% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_associate_service_role_to_account}
\alias{greengrass_associate_service_role_to_account}
\title{Associates a role with your account}
\usage{
greengrass_associate_service_role_to_account(RoleArn)
}
\arguments{
\item{RoleArn}{[required] The ARN of the service role you wish to associate with your account.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  AssociatedAt = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Associates a role with your account. AWS IoT Greengrass will use the
role to access your Lambda functions and AWS IoT resources. This is
necessary for deployments to succeed. The role must have at least
minimum permissions in the policy
''AWSGreengrassResourceAccessRolePolicy''.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$associate_service_role_to_account(
  RoleArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
