% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_create_device_definition}
\alias{greengrass_create_device_definition}
\title{Creates a device definition}
\usage{
greengrass_create_device_definition(AmznClientToken, InitialVersion,
  Name, tags)
}
\arguments{
\item{AmznClientToken}{A client token used to correlate requests and responses.}

\item{InitialVersion}{Information about the initial version of the device definition.}

\item{Name}{The name of the device definition.}

\item{tags}{Tag(s) to add to the new resource.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Arn = "string",
  CreationTimestamp = "string",
  Id = "string",
  LastUpdatedTimestamp = "string",
  LatestVersion = "string",
  LatestVersionArn = "string",
  Name = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Creates a device definition. You may provide the initial version of the
device definition now or use ''CreateDeviceDefinitionVersion'' at a
later time.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_device_definition(
  AmznClientToken = "string",
  InitialVersion = list(
    Devices = list(
      list(
        CertificateArn = "string",
        Id = "string",
        SyncShadow = TRUE|FALSE,
        ThingArn = "string"
      )
    )
  ),
  Name = "string",
  tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
