% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_create_device_definition_version}
\alias{greengrass_create_device_definition_version}
\title{Creates a version of a device definition that has already been defined}
\usage{
greengrass_create_device_definition_version(AmznClientToken,
  DeviceDefinitionId, Devices)
}
\arguments{
\item{AmznClientToken}{A client token used to correlate requests and responses.}

\item{DeviceDefinitionId}{[required] The ID of the device definition.}

\item{Devices}{A list of devices in the definition version.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Arn = "string",
  CreationTimestamp = "string",
  Id = "string",
  Version = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Creates a version of a device definition that has already been defined.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_device_definition_version(
  AmznClientToken = "string",
  DeviceDefinitionId = "string",
  Devices = list(
    list(
      CertificateArn = "string",
      Id = "string",
      SyncShadow = TRUE|FALSE,
      ThingArn = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
