% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_create_function_definition_version}
\alias{greengrass_create_function_definition_version}
\title{Creates a version of a Lambda function definition that has already been
defined}
\usage{
greengrass_create_function_definition_version(AmznClientToken,
  DefaultConfig, FunctionDefinitionId, Functions)
}
\arguments{
\item{AmznClientToken}{A client token used to correlate requests and responses.}

\item{DefaultConfig}{The default configuration that applies to all Lambda functions in this
function definition version. Individual Lambda functions can override
these settings.}

\item{FunctionDefinitionId}{[required] The ID of the Lambda function definition.}

\item{Functions}{A list of Lambda functions in this function definition version.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Arn = "string",
  CreationTimestamp = "string",
  Id = "string",
  Version = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Creates a version of a Lambda function definition that has already been
defined.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_function_definition_version(
  AmznClientToken = "string",
  DefaultConfig = list(
    Execution = list(
      IsolationMode = "GreengrassContainer"|"NoContainer",
      RunAs = list(
        Gid = 123,
        Uid = 123
      )
    )
  ),
  FunctionDefinitionId = "string",
  Functions = list(
    list(
      FunctionArn = "string",
      FunctionConfiguration = list(
        EncodingType = "binary"|"json",
        Environment = list(
          AccessSysfs = TRUE|FALSE,
          Execution = list(
            IsolationMode = "GreengrassContainer"|"NoContainer",
            RunAs = list(
              Gid = 123,
              Uid = 123
            )
          ),
          ResourceAccessPolicies = list(
            list(
              Permission = "ro"|"rw",
              ResourceId = "string"
            )
          ),
          Variables = list(
            "string"
          )
        ),
        ExecArgs = "string",
        Executable = "string",
        MemorySize = 123,
        Pinned = TRUE|FALSE,
        Timeout = 123
      ),
      Id = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
