% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_create_logger_definition}
\alias{greengrass_create_logger_definition}
\title{Creates a logger definition}
\usage{
greengrass_create_logger_definition(AmznClientToken, InitialVersion,
  Name, tags)
}
\arguments{
\item{AmznClientToken}{A client token used to correlate requests and responses.}

\item{InitialVersion}{Information about the initial version of the logger definition.}

\item{Name}{The name of the logger definition.}

\item{tags}{Tag(s) to add to the new resource.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Arn = "string",
  CreationTimestamp = "string",
  Id = "string",
  LastUpdatedTimestamp = "string",
  LatestVersion = "string",
  LatestVersionArn = "string",
  Name = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Creates a logger definition. You may provide the initial version of the
logger definition now or use ''CreateLoggerDefinitionVersion'' at a
later time.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_logger_definition(
  AmznClientToken = "string",
  InitialVersion = list(
    Loggers = list(
      list(
        Component = "GreengrassSystem"|"Lambda",
        Id = "string",
        Level = "DEBUG"|"INFO"|"WARN"|"ERROR"|"FATAL",
        Space = 123,
        Type = "FileSystem"|"AWSCloudWatch"
      )
    )
  ),
  Name = "string",
  tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
