% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_create_logger_definition_version}
\alias{greengrass_create_logger_definition_version}
\title{Creates a version of a logger definition that has already been defined}
\usage{
greengrass_create_logger_definition_version(AmznClientToken,
  LoggerDefinitionId, Loggers)
}
\arguments{
\item{AmznClientToken}{A client token used to correlate requests and responses.}

\item{LoggerDefinitionId}{[required] The ID of the logger definition.}

\item{Loggers}{A list of loggers.}
}
\value{
A list with the following syntax:\preformatted{list(
  Arn = "string",
  CreationTimestamp = "string",
  Id = "string",
  Version = "string"
)
}
}
\description{
Creates a version of a logger definition that has already been defined.
}
\section{Request syntax}{
\preformatted{svc$create_logger_definition_version(
  AmznClientToken = "string",
  LoggerDefinitionId = "string",
  Loggers = list(
    list(
      Component = "GreengrassSystem"|"Lambda",
      Id = "string",
      Level = "DEBUG"|"INFO"|"WARN"|"ERROR"|"FATAL",
      Space = 123,
      Type = "FileSystem"|"AWSCloudWatch"
    )
  )
)
}
}

\keyword{internal}
