% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_create_subscription_definition}
\alias{greengrass_create_subscription_definition}
\title{Creates a subscription definition}
\usage{
greengrass_create_subscription_definition(AmznClientToken,
  InitialVersion, Name, tags)
}
\arguments{
\item{AmznClientToken}{A client token used to correlate requests and responses.}

\item{InitialVersion}{Information about the initial version of the subscription definition.}

\item{Name}{The name of the subscription definition.}

\item{tags}{Tag(s) to add to the new resource.}
}
\value{
A list with the following syntax:\preformatted{list(
  Arn = "string",
  CreationTimestamp = "string",
  Id = "string",
  LastUpdatedTimestamp = "string",
  LatestVersion = "string",
  LatestVersionArn = "string",
  Name = "string"
)
}
}
\description{
Creates a subscription definition. You may provide the initial version
of the subscription definition now or use
''CreateSubscriptionDefinitionVersion'' at a later time.
}
\section{Request syntax}{
\preformatted{svc$create_subscription_definition(
  AmznClientToken = "string",
  InitialVersion = list(
    Subscriptions = list(
      list(
        Id = "string",
        Source = "string",
        Subject = "string",
        Target = "string"
      )
    )
  ),
  Name = "string",
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
