% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_delete_connector_definition}
\alias{greengrass_delete_connector_definition}
\title{Deletes a connector definition}
\usage{
greengrass_delete_connector_definition(ConnectorDefinitionId)
}
\arguments{
\item{ConnectorDefinitionId}{[required] The ID of the connector definition.}
}
\value{
An empty list.
}
\description{
Deletes a connector definition.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_connector_definition(
  ConnectorDefinitionId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
