% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_delete_function_definition}
\alias{greengrass_delete_function_definition}
\title{Deletes a Lambda function definition}
\usage{
greengrass_delete_function_definition(FunctionDefinitionId)
}
\arguments{
\item{FunctionDefinitionId}{[required] The ID of the Lambda function definition.}
}
\value{
An empty list.
}
\description{
Deletes a Lambda function definition.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_function_definition(
  FunctionDefinitionId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
