% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_delete_group}
\alias{greengrass_delete_group}
\title{Deletes a group}
\usage{
greengrass_delete_group(GroupId)
}
\arguments{
\item{GroupId}{[required] The ID of the Greengrass group.}
}
\value{
An empty list.
}
\description{
Deletes a group.
}
\section{Request syntax}{
\preformatted{svc$delete_group(
  GroupId = "string"
)
}
}

\keyword{internal}
