% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_delete_logger_definition}
\alias{greengrass_delete_logger_definition}
\title{Deletes a logger definition}
\usage{
greengrass_delete_logger_definition(LoggerDefinitionId)
}
\arguments{
\item{LoggerDefinitionId}{[required] The ID of the logger definition.}
}
\value{
An empty list.
}
\description{
Deletes a logger definition.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_logger_definition(
  LoggerDefinitionId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
