% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_delete_subscription_definition}
\alias{greengrass_delete_subscription_definition}
\title{Deletes a subscription definition}
\usage{
greengrass_delete_subscription_definition(SubscriptionDefinitionId)
}
\arguments{
\item{SubscriptionDefinitionId}{[required] The ID of the subscription definition.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Deletes a subscription definition.
}
\section{Request syntax}{
\preformatted{svc$delete_subscription_definition(
  SubscriptionDefinitionId = "string"
)
}
}

\keyword{internal}
