% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_get_bulk_deployment_status}
\alias{greengrass_get_bulk_deployment_status}
\title{Returns the status of a bulk deployment}
\usage{
greengrass_get_bulk_deployment_status(BulkDeploymentId)
}
\arguments{
\item{BulkDeploymentId}{[required] The ID of the bulk deployment.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  BulkDeploymentMetrics = list(
    InvalidInputRecords = 123,
    RecordsProcessed = 123,
    RetryAttempts = 123
  ),
  BulkDeploymentStatus = "Initializing"|"Running"|"Completed"|"Stopping"|"Stopped"|"Failed",
  CreatedAt = "string",
  ErrorDetails = list(
    list(
      DetailedErrorCode = "string",
      DetailedErrorMessage = "string"
    )
  ),
  ErrorMessage = "string",
  tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns the status of a bulk deployment.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_bulk_deployment_status(
  BulkDeploymentId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
