% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_get_connector_definition}
\alias{greengrass_get_connector_definition}
\title{Retrieves information about a connector definition}
\usage{
greengrass_get_connector_definition(ConnectorDefinitionId)
}
\arguments{
\item{ConnectorDefinitionId}{[required] The ID of the connector definition.}
}
\value{
A list with the following syntax:\preformatted{list(
  Arn = "string",
  CreationTimestamp = "string",
  Id = "string",
  LastUpdatedTimestamp = "string",
  LatestVersion = "string",
  LatestVersionArn = "string",
  Name = "string",
  tags = list(
    "string"
  )
)
}
}
\description{
Retrieves information about a connector definition.
}
\section{Request syntax}{
\preformatted{svc$get_connector_definition(
  ConnectorDefinitionId = "string"
)
}
}

\keyword{internal}
