% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_get_core_definition}
\alias{greengrass_get_core_definition}
\title{Retrieves information about a core definition version}
\usage{
greengrass_get_core_definition(CoreDefinitionId)
}
\arguments{
\item{CoreDefinitionId}{[required] The ID of the core definition.}
}
\value{
A list with the following syntax:\preformatted{list(
  Arn = "string",
  CreationTimestamp = "string",
  Id = "string",
  LastUpdatedTimestamp = "string",
  LatestVersion = "string",
  LatestVersionArn = "string",
  Name = "string",
  tags = list(
    "string"
  )
)
}
}
\description{
Retrieves information about a core definition version.
}
\section{Request syntax}{
\preformatted{svc$get_core_definition(
  CoreDefinitionId = "string"
)
}
}

\keyword{internal}
