% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_get_core_definition_version}
\alias{greengrass_get_core_definition_version}
\title{Retrieves information about a core definition version}
\usage{
greengrass_get_core_definition_version(CoreDefinitionId,
  CoreDefinitionVersionId)
}
\arguments{
\item{CoreDefinitionId}{[required] The ID of the core definition.}

\item{CoreDefinitionVersionId}{[required] The ID of the core definition version. This value maps to the ''Version'' property of the corresponding ''VersionInformation'' object, which is returned by ''ListCoreDefinitionVersions'' requests. If the version is the last one that was associated with a core definition, the value also maps to the ''LatestVersion'' property of the corresponding ''DefinitionInformation'' object.}
}
\description{
Retrieves information about a core definition version.
}
\section{Request syntax}{
\preformatted{svc$get_core_definition_version(
  CoreDefinitionId = "string",
  CoreDefinitionVersionId = "string"
)
}
}

\keyword{internal}
