% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_get_deployment_status}
\alias{greengrass_get_deployment_status}
\title{Returns the status of a deployment}
\usage{
greengrass_get_deployment_status(DeploymentId, GroupId)
}
\arguments{
\item{DeploymentId}{[required] The ID of the deployment.}

\item{GroupId}{[required] The ID of the Greengrass group.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  DeploymentStatus = "string",
  DeploymentType = "NewDeployment"|"Redeployment"|"ResetDeployment"|"ForceResetDeployment",
  ErrorDetails = list(
    list(
      DetailedErrorCode = "string",
      DetailedErrorMessage = "string"
    )
  ),
  ErrorMessage = "string",
  UpdatedAt = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns the status of a deployment.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_deployment_status(
  DeploymentId = "string",
  GroupId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
