% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_get_function_definition}
\alias{greengrass_get_function_definition}
\title{Retrieves information about a Lambda function definition, including its
creation time and latest version}
\usage{
greengrass_get_function_definition(FunctionDefinitionId)
}
\arguments{
\item{FunctionDefinitionId}{[required] The ID of the Lambda function definition.}
}
\value{
A list with the following syntax:\preformatted{list(
  Arn = "string",
  CreationTimestamp = "string",
  Id = "string",
  LastUpdatedTimestamp = "string",
  LatestVersion = "string",
  LatestVersionArn = "string",
  Name = "string",
  tags = list(
    "string"
  )
)
}
}
\description{
Retrieves information about a Lambda function definition, including its
creation time and latest version.
}
\section{Request syntax}{
\preformatted{svc$get_function_definition(
  FunctionDefinitionId = "string"
)
}
}

\keyword{internal}
