% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_get_service_role_for_account}
\alias{greengrass_get_service_role_for_account}
\title{Retrieves the service role that is attached to your account}
\usage{
greengrass_get_service_role_for_account()
}
\value{
A list with the following syntax:\preformatted{list(
  AssociatedAt = "string",
  RoleArn = "string"
)
}
}
\description{
Retrieves the service role that is attached to your account.
}
\section{Request syntax}{
\preformatted{svc$get_service_role_for_account()
}
}

\keyword{internal}
