% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_get_subscription_definition_version}
\alias{greengrass_get_subscription_definition_version}
\title{Retrieves information about a subscription definition version}
\usage{
greengrass_get_subscription_definition_version(NextToken,
  SubscriptionDefinitionId, SubscriptionDefinitionVersionId)
}
\arguments{
\item{NextToken}{The token for the next set of results, or ''null'' if there are no
additional results.}

\item{SubscriptionDefinitionId}{[required] The ID of the subscription definition.}

\item{SubscriptionDefinitionVersionId}{[required] The ID of the subscription definition version. This value maps to the
''Version'' property of the corresponding ''VersionInformation'' object,
which is returned by ''ListSubscriptionDefinitionVersions'' requests. If
the version is the last one that was associated with a subscription
definition, the value also maps to the ''LatestVersion'' property of the
corresponding ''DefinitionInformation'' object.}
}
\value{
A list with the following syntax:\preformatted{list(
  Arn = "string",
  CreationTimestamp = "string",
  Definition = list(
    Subscriptions = list(
      list(
        Id = "string",
        Source = "string",
        Subject = "string",
        Target = "string"
      )
    )
  ),
  Id = "string",
  NextToken = "string",
  Version = "string"
)
}
}
\description{
Retrieves information about a subscription definition version.
}
\section{Request syntax}{
\preformatted{svc$get_subscription_definition_version(
  NextToken = "string",
  SubscriptionDefinitionId = "string",
  SubscriptionDefinitionVersionId = "string"
)
}
}

\keyword{internal}
