% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_get_thing_runtime_configuration}
\alias{greengrass_get_thing_runtime_configuration}
\title{Get the runtime configuration of a thing}
\usage{
greengrass_get_thing_runtime_configuration(ThingName)
}
\arguments{
\item{ThingName}{[required] The thing name.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  RuntimeConfiguration = list(
    TelemetryConfiguration = list(
      ConfigurationSyncStatus = "InSync"|"OutOfSync",
      Telemetry = "On"|"Off"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Get the runtime configuration of a thing.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_thing_runtime_configuration(
  ThingName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
