% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_list_bulk_deployments}
\alias{greengrass_list_bulk_deployments}
\title{Returns a list of bulk deployments}
\usage{
greengrass_list_bulk_deployments(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{The maximum number of results to be returned per request.}

\item{NextToken}{The token for the next set of results, or ''null'' if there are no
additional results.}
}
\value{
A list with the following syntax:\preformatted{list(
  BulkDeployments = list(
    list(
      BulkDeploymentArn = "string",
      BulkDeploymentId = "string",
      CreatedAt = "string"
    )
  ),
  NextToken = "string"
)
}
}
\description{
Returns a list of bulk deployments.
}
\section{Request syntax}{
\preformatted{svc$list_bulk_deployments(
  MaxResults = "string",
  NextToken = "string"
)
}
}

\keyword{internal}
