% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_list_connector_definitions}
\alias{greengrass_list_connector_definitions}
\title{Retrieves a list of connector definitions}
\usage{
greengrass_list_connector_definitions(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{The maximum number of results to be returned per request.}

\item{NextToken}{The token for the next set of results, or ''null'' if there are no
additional results.}
}
\value{
A list with the following syntax:\preformatted{list(
  Definitions = list(
    list(
      Arn = "string",
      CreationTimestamp = "string",
      Id = "string",
      LastUpdatedTimestamp = "string",
      LatestVersion = "string",
      LatestVersionArn = "string",
      Name = "string",
      Tags = list(
        "string"
      )
    )
  ),
  NextToken = "string"
)
}
}
\description{
Retrieves a list of connector definitions.
}
\section{Request syntax}{
\preformatted{svc$list_connector_definitions(
  MaxResults = "string",
  NextToken = "string"
)
}
}

\keyword{internal}
