% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_list_deployments}
\alias{greengrass_list_deployments}
\title{Returns a history of deployments for the group}
\usage{
greengrass_list_deployments(GroupId, MaxResults, NextToken)
}
\arguments{
\item{GroupId}{[required] The ID of the Greengrass group.}

\item{MaxResults}{The maximum number of results to be returned per request.}

\item{NextToken}{The token for the next set of results, or ''null'' if there are no
additional results.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Deployments = list(
    list(
      CreatedAt = "string",
      DeploymentArn = "string",
      DeploymentId = "string",
      DeploymentType = "NewDeployment"|"Redeployment"|"ResetDeployment"|"ForceResetDeployment",
      GroupArn = "string"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns a history of deployments for the group.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_deployments(
  GroupId = "string",
  MaxResults = "string",
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
