% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_list_device_definition_versions}
\alias{greengrass_list_device_definition_versions}
\title{Lists the versions of a device definition}
\usage{
greengrass_list_device_definition_versions(DeviceDefinitionId,
  MaxResults, NextToken)
}
\arguments{
\item{DeviceDefinitionId}{[required] The ID of the device definition.}

\item{MaxResults}{The maximum number of results to be returned per request.}

\item{NextToken}{The token for the next set of results, or ''null'' if there are no
additional results.}
}
\value{
A list with the following syntax:\preformatted{list(
  NextToken = "string",
  Versions = list(
    list(
      Arn = "string",
      CreationTimestamp = "string",
      Id = "string",
      Version = "string"
    )
  )
)
}
}
\description{
Lists the versions of a device definition.
}
\section{Request syntax}{
\preformatted{svc$list_device_definition_versions(
  DeviceDefinitionId = "string",
  MaxResults = "string",
  NextToken = "string"
)
}
}

\keyword{internal}
