% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_list_group_certificate_authorities}
\alias{greengrass_list_group_certificate_authorities}
\title{Retrieves the current CAs for a group}
\usage{
greengrass_list_group_certificate_authorities(GroupId)
}
\arguments{
\item{GroupId}{[required] The ID of the Greengrass group.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  GroupCertificateAuthorities = list(
    list(
      GroupCertificateAuthorityArn = "string",
      GroupCertificateAuthorityId = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves the current CAs for a group.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_group_certificate_authorities(
  GroupId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
