% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_list_groups}
\alias{greengrass_list_groups}
\title{Retrieves a list of groups}
\usage{
greengrass_list_groups(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{The maximum number of results to be returned per request.}

\item{NextToken}{The token for the next set of results, or ''null'' if there are no additional results.}
}
\description{
Retrieves a list of groups.
}
\section{Request syntax}{
\preformatted{svc$list_groups(
  MaxResults = "string",
  NextToken = "string"
)
}
}

\keyword{internal}
