% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_list_logger_definition_versions}
\alias{greengrass_list_logger_definition_versions}
\title{Lists the versions of a logger definition}
\usage{
greengrass_list_logger_definition_versions(LoggerDefinitionId,
  MaxResults, NextToken)
}
\arguments{
\item{LoggerDefinitionId}{[required] The ID of the logger definition.}

\item{MaxResults}{The maximum number of results to be returned per request.}

\item{NextToken}{The token for the next set of results, or ''null'' if there are no additional results.}
}
\description{
Lists the versions of a logger definition.
}
\section{Request syntax}{
\preformatted{svc$list_logger_definition_versions(
  LoggerDefinitionId = "string",
  MaxResults = "string",
  NextToken = "string"
)
}
}

\keyword{internal}
