% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_list_logger_definitions}
\alias{greengrass_list_logger_definitions}
\title{Retrieves a list of logger definitions}
\usage{
greengrass_list_logger_definitions(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{The maximum number of results to be returned per request.}

\item{NextToken}{The token for the next set of results, or ''null'' if there are no
additional results.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Definitions = list(
    list(
      Arn = "string",
      CreationTimestamp = "string",
      Id = "string",
      LastUpdatedTimestamp = "string",
      LatestVersion = "string",
      LatestVersionArn = "string",
      Name = "string",
      Tags = list(
        "string"
      )
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves a list of logger definitions.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_logger_definitions(
  MaxResults = "string",
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
