% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_start_bulk_deployment}
\alias{greengrass_start_bulk_deployment}
\title{Deploys multiple groups in one operation}
\usage{
greengrass_start_bulk_deployment(AmznClientToken, ExecutionRoleArn,
  InputFileUri, tags)
}
\arguments{
\item{AmznClientToken}{A client token used to correlate requests and responses.}

\item{ExecutionRoleArn}{[required] The ARN of the execution role to associate with the bulk deployment
operation. This IAM role must allow the ''greengrass:CreateDeployment''
action for all group versions that are listed in the input file. This
IAM role must have access to the S3 bucket containing the input file.}

\item{InputFileUri}{[required] The URI of the input file contained in the S3 bucket. The execution role
must have ''getObject'' permissions on this bucket to access the input
file. The input file is a JSON-serialized, line delimited file with
UTF-8 encoding that provides a list of group and version IDs and the
deployment type. This file must be less than 100 MB. Currently, AWS IoT
Greengrass supports only ''NewDeployment'' deployment types.}

\item{tags}{Tag(s) to add to the new resource.}
}
\value{
A list with the following syntax:\preformatted{list(
  BulkDeploymentArn = "string",
  BulkDeploymentId = "string"
)
}
}
\description{
Deploys multiple groups in one operation. This action starts the bulk
deployment of a specified set of group versions. Each group version
deployment will be triggered with an adaptive rate that has a fixed
upper limit. We recommend that you include an ''X-Amzn-Client-Token''
token in every ''StartBulkDeployment'' request. These requests are
idempotent with respect to the token and the request parameters.
}
\section{Request syntax}{
\preformatted{svc$start_bulk_deployment(
  AmznClientToken = "string",
  ExecutionRoleArn = "string",
  InputFileUri = "string",
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
