% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_tag_resource}
\alias{greengrass_tag_resource}
\title{Adds tags to a Greengrass resource}
\usage{
greengrass_tag_resource(ResourceArn, tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{tags}{}
}
\value{
An empty list.
}
\description{
Adds tags to a Greengrass resource. Valid resources are 'Group',
'ConnectorDefinition', 'CoreDefinition', 'DeviceDefinition',
'FunctionDefinition', 'LoggerDefinition', 'SubscriptionDefinition',
'ResourceDefinition', and 'BulkDeployment'.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  ResourceArn = "string",
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
