% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_update_core_definition}
\alias{greengrass_update_core_definition}
\title{Updates a core definition}
\usage{
greengrass_update_core_definition(CoreDefinitionId, Name)
}
\arguments{
\item{CoreDefinitionId}{[required] The ID of the core definition.}

\item{Name}{The name of the definition.}
}
\value{
An empty list.
}
\description{
Updates a core definition.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_core_definition(
  CoreDefinitionId = "string",
  Name = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
