% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_update_device_definition}
\alias{greengrass_update_device_definition}
\title{Updates a device definition}
\usage{
greengrass_update_device_definition(DeviceDefinitionId, Name)
}
\arguments{
\item{DeviceDefinitionId}{[required] The ID of the device definition.}

\item{Name}{The name of the definition.}
}
\value{
An empty list.
}
\description{
Updates a device definition.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_device_definition(
  DeviceDefinitionId = "string",
  Name = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
