% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_update_function_definition}
\alias{greengrass_update_function_definition}
\title{Updates a Lambda function definition}
\usage{
greengrass_update_function_definition(FunctionDefinitionId, Name)
}
\arguments{
\item{FunctionDefinitionId}{[required] The ID of the Lambda function definition.}

\item{Name}{The name of the definition.}
}
\value{
An empty list.
}
\description{
Updates a Lambda function definition.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_function_definition(
  FunctionDefinitionId = "string",
  Name = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
