% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_update_group_certificate_configuration}
\alias{greengrass_update_group_certificate_configuration}
\title{Updates the Certificate expiry time for a group}
\usage{
greengrass_update_group_certificate_configuration(
  CertificateExpiryInMilliseconds, GroupId)
}
\arguments{
\item{CertificateExpiryInMilliseconds}{The amount of time remaining before the certificate expires, in
milliseconds.}

\item{GroupId}{[required] The ID of the Greengrass group.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  CertificateAuthorityExpiryInMilliseconds = "string",
  CertificateExpiryInMilliseconds = "string",
  GroupId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Updates the Certificate expiry time for a group.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_group_certificate_configuration(
  CertificateExpiryInMilliseconds = "string",
  GroupId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
