% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_update_logger_definition}
\alias{greengrass_update_logger_definition}
\title{Updates a logger definition}
\usage{
greengrass_update_logger_definition(LoggerDefinitionId, Name)
}
\arguments{
\item{LoggerDefinitionId}{[required] The ID of the logger definition.}

\item{Name}{The name of the definition.}
}
\value{
An empty list.
}
\description{
Updates a logger definition.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_logger_definition(
  LoggerDefinitionId = "string",
  Name = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
