% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot1clickdevicesservice_operations.R
\name{iot1clickdevicesservice_describe_device}
\alias{iot1clickdevicesservice_describe_device}
\title{Given a device ID, returns a DescribeDeviceResponse object describing
the details of the device}
\usage{
iot1clickdevicesservice_describe_device(DeviceId)
}
\arguments{
\item{DeviceId}{[required] The unique identifier of the device.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  DeviceDescription = list(
    Arn = "string",
    Attributes = list(
      "string"
    ),
    DeviceId = "string",
    Enabled = TRUE|FALSE,
    RemainingLife = 123.0,
    Type = "string",
    Tags = list(
      "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Given a device ID, returns a DescribeDeviceResponse object describing
the details of the device.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_device(
  DeviceId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
