% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot1clickdevicesservice_operations.R
\name{iot1clickdevicesservice_get_device_methods}
\alias{iot1clickdevicesservice_get_device_methods}
\title{Given a device ID, returns the invokable methods associated with the
device}
\usage{
iot1clickdevicesservice_get_device_methods(DeviceId)
}
\arguments{
\item{DeviceId}{[required] The unique identifier of the device.}
}
\description{
Given a device ID, returns the invokable methods associated with the
device.
}
\section{Request syntax}{
\preformatted{svc$get_device_methods(
  DeviceId = "string"
)
}
}

\keyword{internal}
