% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot1clickdevicesservice_operations.R
\name{iot1clickdevicesservice_initiate_device_claim}
\alias{iot1clickdevicesservice_initiate_device_claim}
\title{Given a device ID, initiates a claim request for the associated device}
\usage{
iot1clickdevicesservice_initiate_device_claim(DeviceId)
}
\arguments{
\item{DeviceId}{[required] The unique identifier of the device.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  State = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Given a device ID, initiates a claim request for the associated device.

Claiming a device consists of initiating a claim, then publishing a
device event, and finalizing the claim. For a device of type button, a
device event can be published by simply clicking the device.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$initiate_device_claim(
  DeviceId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
