% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot1clickdevicesservice_operations.R
\name{iot1clickdevicesservice_list_device_events}
\alias{iot1clickdevicesservice_list_device_events}
\title{Using a device ID, returns a DeviceEventsResponse object containing an
array of events for the device}
\usage{
iot1clickdevicesservice_list_device_events(DeviceId, FromTimeStamp,
  MaxResults, NextToken, ToTimeStamp)
}
\arguments{
\item{DeviceId}{[required] The unique identifier of the device.}

\item{FromTimeStamp}{[required] The start date for the device event query, in ISO8061 format. For
example, 2018-03-28T15:45:12.880Z}

\item{MaxResults}{The maximum number of results to return per request. If not set, a
default value of 100 is used.}

\item{NextToken}{The token to retrieve the next set of results.}

\item{ToTimeStamp}{[required] The end date for the device event query, in ISO8061 format. For example,
2018-03-28T15:45:12.880Z}
}
\value{
A list with the following syntax:\preformatted{list(
  Events = list(
    list(
      Device = list(
        Attributes = list(),
        DeviceId = "string",
        Type = "string"
      ),
      StdEvent = "string"
    )
  ),
  NextToken = "string"
)
}
}
\description{
Using a device ID, returns a DeviceEventsResponse object containing an
array of events for the device.
}
\section{Request syntax}{
\preformatted{svc$list_device_events(
  DeviceId = "string",
  FromTimeStamp = as.POSIXct(
    "2015-01-01"
  ),
  MaxResults = 123,
  NextToken = "string",
  ToTimeStamp = as.POSIXct(
    "2015-01-01"
  )
)
}
}

\keyword{internal}
