% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot1clickdevicesservice_operations.R
\name{iot1clickdevicesservice_list_devices}
\alias{iot1clickdevicesservice_list_devices}
\title{Lists the 1-Click compatible devices associated with your AWS account}
\usage{
iot1clickdevicesservice_list_devices(DeviceType, MaxResults, NextToken)
}
\arguments{
\item{DeviceType}{The type of the device, such as "button".}

\item{MaxResults}{The maximum number of results to return per request. If not set, a
default value of 100 is used.}

\item{NextToken}{The token to retrieve the next set of results.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Devices = list(
    list(
      Arn = "string",
      Attributes = list(
        "string"
      ),
      DeviceId = "string",
      Enabled = TRUE|FALSE,
      RemainingLife = 123.0,
      Type = "string",
      Tags = list(
        "string"
      )
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists the 1-Click compatible devices associated with your AWS account.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_devices(
  DeviceType = "string",
  MaxResults = 123,
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
