% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot1clickdevicesservice_operations.R
\name{iot1clickdevicesservice_list_tags_for_resource}
\alias{iot1clickdevicesservice_list_tags_for_resource}
\title{Lists the tags associated with the specified resource ARN}
\usage{
iot1clickdevicesservice_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the resource.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Lists the tags associated with the specified resource ARN.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_tags_for_resource(
  ResourceArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
