% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot1clickdevicesservice_operations.R
\name{iot1clickdevicesservice_tag_resource}
\alias{iot1clickdevicesservice_tag_resource}
\title{Adds or updates the tags associated with the resource ARN}
\usage{
iot1clickdevicesservice_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the resource.}

\item{Tags}{[required] A collection of key/value pairs defining the resource tags. For example,
\{ "tags": \{"key1": "value1", "key2": "value2"\} \}. For more
information, see \href{https://d1.awsstatic.com/whitepapers/aws-tagging-best-practices.pdf}{AWS Tagging Strategies}.}
}
\value{
An empty list.
}
\description{
Adds or updates the tags associated with the resource ARN. See \href{https://docs.aws.amazon.com/iot-1-click/latest/developerguide/1click-appendix.html#1click-limits}{AWS IoT 1-Click Service Limits}
for the maximum number of tags allowed per resource.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  ResourceArn = "string",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
