% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot1clickdevicesservice_operations.R
\name{iot1clickdevicesservice_unclaim_device}
\alias{iot1clickdevicesservice_unclaim_device}
\title{Disassociates a device from your AWS account using its device ID}
\usage{
iot1clickdevicesservice_unclaim_device(DeviceId)
}
\arguments{
\item{DeviceId}{[required] The unique identifier of the device.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  State = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Disassociates a device from your AWS account using its device ID.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$unclaim_device(
  DeviceId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
