% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot1clickprojects_operations.R
\name{iot1clickprojects_describe_placement}
\alias{iot1clickprojects_describe_placement}
\title{Describes a placement in a project}
\usage{
iot1clickprojects_describe_placement(placementName, projectName)
}
\arguments{
\item{placementName}{[required] The name of the placement within a project.}

\item{projectName}{[required] The project containing the placement to be described.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  placement = list(
    projectName = "string",
    placementName = "string",
    attributes = list(
      "string"
    ),
    createdDate = as.POSIXct(
      "2015-01-01"
    ),
    updatedDate = as.POSIXct(
      "2015-01-01"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes a placement in a project.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_placement(
  placementName = "string",
  projectName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
