% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot1clickprojects_operations.R
\name{iot1clickprojects_describe_project}
\alias{iot1clickprojects_describe_project}
\title{Returns an object describing a project}
\usage{
iot1clickprojects_describe_project(projectName)
}
\arguments{
\item{projectName}{[required] The name of the project to be described.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  project = list(
    arn = "string",
    projectName = "string",
    description = "string",
    createdDate = as.POSIXct(
      "2015-01-01"
    ),
    updatedDate = as.POSIXct(
      "2015-01-01"
    ),
    placementTemplate = list(
      defaultAttributes = list(
        "string"
      ),
      deviceTemplates = list(
        list(
          deviceType = "string",
          callbackOverrides = list(
            "string"
          )
        )
      )
    ),
    tags = list(
      "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns an object describing a project.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_project(
  projectName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
