% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot1clickprojects_operations.R
\name{iot1clickprojects_get_devices_in_placement}
\alias{iot1clickprojects_get_devices_in_placement}
\title{Returns an object enumerating the devices in a placement}
\usage{
iot1clickprojects_get_devices_in_placement(projectName, placementName)
}
\arguments{
\item{projectName}{[required] The name of the project containing the placement.}

\item{placementName}{[required] The name of the placement to get the devices from.}
}
\value{
A list with the following syntax:\preformatted{list(
  devices = list(
    "string"
  )
)
}
}
\description{
Returns an object enumerating the devices in a placement.
}
\section{Request syntax}{
\preformatted{svc$get_devices_in_placement(
  projectName = "string",
  placementName = "string"
)
}
}

\keyword{internal}
