% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot1clickprojects_operations.R
\name{iot1clickprojects_list_placements}
\alias{iot1clickprojects_list_placements}
\title{Lists the placement(s) of a project}
\usage{
iot1clickprojects_list_placements(projectName, nextToken, maxResults)
}
\arguments{
\item{projectName}{[required] The project containing the placements to be listed.}

\item{nextToken}{The token to retrieve the next set of results.}

\item{maxResults}{The maximum number of results to return per request. If not set, a
default value of 100 is used.}
}
\value{
A list with the following syntax:\preformatted{list(
  placements = list(
    list(
      projectName = "string",
      placementName = "string",
      createdDate = as.POSIXct(
        "2015-01-01"
      ),
      updatedDate = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  nextToken = "string"
)
}
}
\description{
Lists the placement(s) of a project.
}
\section{Request syntax}{
\preformatted{svc$list_placements(
  projectName = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
