% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot1clickprojects_operations.R
\name{iot1clickprojects_tag_resource}
\alias{iot1clickprojects_tag_resource}
\title{Creates or modifies tags for a resource}
\usage{
iot1clickprojects_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resouce for which tag(s) should be added or modified.}

\item{tags}{[required] The new or modifying tag(s) for the resource. See \href{https://docs.aws.amazon.com/iot-1-click/latest/developerguide/1click-appendix.html#1click-limits}{AWS IoT 1-Click Service Limits}
for the maximum number of tags allowed per resource.}
}
\value{
An empty list.
}
\description{
Creates or modifies tags for a resource. Tags are key/value pairs
(metadata) that can be used to manage a resource. For more information,
see \href{https://d1.awsstatic.com/whitepapers/aws-tagging-best-practices.pdf}{AWS Tagging Strategies}.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  resourceArn = "string",
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
