% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot1clickprojects_operations.R
\name{iot1clickprojects_untag_resource}
\alias{iot1clickprojects_untag_resource}
\title{Removes one or more tags (metadata key/value pairs) from a resource}
\usage{
iot1clickprojects_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource whose tag you want to remove.}

\item{tagKeys}{[required] The keys of those tags which you want to remove.}
}
\value{
An empty list.
}
\description{
Removes one or more tags (metadata key/value pairs) from a resource.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$untag_resource(
  resourceArn = "string",
  tagKeys = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
