% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_associate_targets_with_job}
\alias{iot_associate_targets_with_job}
\title{Associates a group with a continuous job}
\usage{
iot_associate_targets_with_job(targets, jobId, comment)
}
\arguments{
\item{targets}{[required] A list of thing group ARNs that define the targets of the job.}

\item{jobId}{[required] The unique identifier you assigned to this job when it was created.}

\item{comment}{An optional comment string describing why the job was associated with
the targets.}
}
\description{
Associates a group with a continuous job. The following criteria must be
met:
}
\details{
\itemize{
\item The job must have been created with the \code{targetSelection} field set
to "CONTINUOUS".
\item The job status must currently be "IN\\_PROGRESS".
\item The total number of targets associated with a job must not
exceed 100.
}
}
\section{Request syntax}{
\preformatted{svc$associate_targets_with_job(
  targets = list(
    "string"
  ),
  jobId = "string",
  comment = "string"
)
}
}

\keyword{internal}
