% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_attach_policy}
\alias{iot_attach_policy}
\title{Attaches a policy to the specified target}
\usage{
iot_attach_policy(policyName, target)
}
\arguments{
\item{policyName}{[required] The name of the policy to attach.}

\item{target}{[required] The
\href{https://docs.aws.amazon.com/iot/latest/developerguide/security-iam.html}{identity}
to which the policy is attached.}
}
\value{
An empty list.
}
\description{
Attaches a policy to the specified target.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$attach_policy(
  policyName = "string",
  target = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
