% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_attach_thing_principal}
\alias{iot_attach_thing_principal}
\title{Attaches the specified principal to the specified thing}
\usage{
iot_attach_thing_principal(thingName, principal)
}
\arguments{
\item{thingName}{[required] The name of the thing.}

\item{principal}{[required] The principal, which can be a certificate ARN (as returned from the
CreateCertificate operation) or an Amazon Cognito ID.}
}
\value{
An empty list.
}
\description{
Attaches the specified principal to the specified thing. A principal can
be X.509 certificates, IAM users, groups, and roles, Amazon Cognito
identities or federated identities.
}
\section{Request syntax}{
\preformatted{svc$attach_thing_principal(
  thingName = "string",
  principal = "string"
)
}
}

\keyword{internal}
