% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_cancel_audit_mitigation_actions_task}
\alias{iot_cancel_audit_mitigation_actions_task}
\title{Cancels a mitigation action task that is in progress}
\usage{
iot_cancel_audit_mitigation_actions_task(taskId)
}
\arguments{
\item{taskId}{[required] The unique identifier for the task that you want to cancel.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Cancels a mitigation action task that is in progress. If the task is not
in progress, an InvalidRequestException occurs.
}
\section{Request syntax}{
\preformatted{svc$cancel_audit_mitigation_actions_task(
  taskId = "string"
)
}
}

\keyword{internal}
