% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_cancel_detect_mitigation_actions_task}
\alias{iot_cancel_detect_mitigation_actions_task}
\title{Cancels a Device Defender ML Detect mitigation action}
\usage{
iot_cancel_detect_mitigation_actions_task(taskId)
}
\arguments{
\item{taskId}{[required] The unique identifier of the task.}
}
\value{
An empty list.
}
\description{
Cancels a Device Defender ML Detect mitigation action.
}
\section{Request syntax}{
\preformatted{svc$cancel_detect_mitigation_actions_task(
  taskId = "string"
)
}
}

\keyword{internal}
