% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_cancel_job}
\alias{iot_cancel_job}
\title{Cancels a job}
\usage{
iot_cancel_job(jobId, reasonCode, comment, force)
}
\arguments{
\item{jobId}{[required] The unique identifier you assigned to this job when it was created.}

\item{reasonCode}{(Optional)A reason code string that explains why the job was canceled.}

\item{comment}{An optional comment string describing why the job was canceled.}

\item{force}{(Optional) If \code{true} job executions with status "IN_PROGRESS" and
"QUEUED" are canceled, otherwise only job executions with status
"QUEUED" are canceled. The default is \code{false}.

Canceling a job which is "IN_PROGRESS", will cause a device which is
executing the job to be unable to update the job execution status. Use
caution and ensure that each device executing a job which is canceled is
able to recover to a valid state.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  jobArn = "string",
  jobId = "string",
  description = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Cancels a job.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$cancel_job(
  jobId = "string",
  reasonCode = "string",
  comment = "string",
  force = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
