% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_confirm_topic_rule_destination}
\alias{iot_confirm_topic_rule_destination}
\title{Confirms a topic rule destination}
\usage{
iot_confirm_topic_rule_destination(confirmationToken)
}
\arguments{
\item{confirmationToken}{[required] The token used to confirm ownership or access to the topic rule
confirmation URL.}
}
\value{
An empty list.
}
\description{
Confirms a topic rule destination. When you create a rule requiring a
destination, AWS IoT sends a confirmation message to the endpoint or
base address you specify. The message includes a token which you pass
back when calling
\code{\link[=iot_confirm_topic_rule_destination]{confirm_topic_rule_destination}}
to confirm that you own or have access to the endpoint.
}
\section{Request syntax}{
\preformatted{svc$confirm_topic_rule_destination(
  confirmationToken = "string"
)
}
}

\keyword{internal}
