% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_create_audit_suppression}
\alias{iot_create_audit_suppression}
\title{Creates a Device Defender audit suppression}
\usage{
iot_create_audit_suppression(checkName, resourceIdentifier,
  expirationDate, suppressIndefinitely, description, clientRequestToken)
}
\arguments{
\item{checkName}{[required]}

\item{resourceIdentifier}{[required]}

\item{expirationDate}{The epoch timestamp in seconds at which this suppression expires.}

\item{suppressIndefinitely}{Indicates whether a suppression should exist indefinitely or not.}

\item{description}{The description of the audit suppression.}

\item{clientRequestToken}{[required] The epoch timestamp in seconds at which this suppression expires.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Creates a Device Defender audit suppression.
}
\section{Request syntax}{
\preformatted{svc$create_audit_suppression(
  checkName = "string",
  resourceIdentifier = list(
    deviceCertificateId = "string",
    caCertificateId = "string",
    cognitoIdentityPoolId = "string",
    clientId = "string",
    policyVersionIdentifier = list(
      policyName = "string",
      policyVersionId = "string"
    ),
    account = "string",
    iamRoleArn = "string",
    roleAliasArn = "string"
  ),
  expirationDate = as.POSIXct(
    "2015-01-01"
  ),
  suppressIndefinitely = TRUE|FALSE,
  description = "string",
  clientRequestToken = "string"
)
}
}

\keyword{internal}
