% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_create_custom_metric}
\alias{iot_create_custom_metric}
\title{Use this API to define a Custom Metric published by your devices to
Device Defender}
\usage{
iot_create_custom_metric(metricName, displayName, metricType, tags,
  clientRequestToken)
}
\arguments{
\item{metricName}{[required] The name of the custom metric. This will be used in the metric report
submitted from the device/thing. Shouldn't begin with \verb{aws:}. Cannot be
updated once defined.}

\item{displayName}{Field represents a friendly name in the console for the custom metric;
it doesn't have to be unique. Don't use this name as the metric
identifier in the device metric report. Can be updated once defined.}

\item{metricType}{[required] The type of the custom metric. Types include \code{string-list},
\code{ip-address-list}, \code{number-list}, and \code{number}.}

\item{tags}{Metadata that can be used to manage the custom metric.}

\item{clientRequestToken}{[required] Each custom metric must have a unique client request token. If you try
to create a new custom metric that already exists with a different
token, an exception occurs. If you omit this value, AWS SDKs will
automatically generate a unique client request.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  metricName = "string",
  metricArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Use this API to define a Custom Metric published by your devices to
Device Defender.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_custom_metric(
  metricName = "string",
  displayName = "string",
  metricType = "string-list"|"ip-address-list"|"number-list"|"number",
  tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  clientRequestToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
