% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_create_dimension}
\alias{iot_create_dimension}
\title{Create a dimension that you can use to limit the scope of a metric used
in a security profile for AWS IoT Device Defender}
\usage{
iot_create_dimension(name, type, stringValues, tags, clientRequestToken)
}
\arguments{
\item{name}{[required] A unique identifier for the dimension. Choose something that describes
the type and value to make it easy to remember what it does.}

\item{type}{[required] Specifies the type of dimension. Supported types: \code{TOPIC_FILTER.}}

\item{stringValues}{[required] Specifies the value or list of values for the dimension. For
\code{TOPIC_FILTER} dimensions, this is a pattern used to match the MQTT
topic (for example, "admin/#").}

\item{tags}{Metadata that can be used to manage the dimension.}

\item{clientRequestToken}{[required] Each dimension must have a unique client request token. If you try to
create a new dimension with the same token as a dimension that already
exists, an exception occurs. If you omit this value, AWS SDKs will
automatically generate a unique client request.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  name = "string",
  arn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Create a dimension that you can use to limit the scope of a metric used
in a security profile for AWS IoT Device Defender. For example, using a
\code{TOPIC_FILTER} dimension, you can narrow down the scope of the metric
only to MQTT topics whose name match the pattern specified in the
dimension.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_dimension(
  name = "string",
  type = "TOPIC_FILTER",
  stringValues = list(
    "string"
  ),
  tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  clientRequestToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
