% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_create_dynamic_thing_group}
\alias{iot_create_dynamic_thing_group}
\title{Creates a dynamic thing group}
\usage{
iot_create_dynamic_thing_group(thingGroupName, thingGroupProperties,
  indexName, queryString, queryVersion, tags)
}
\arguments{
\item{thingGroupName}{[required] The dynamic thing group name to create.}

\item{thingGroupProperties}{The dynamic thing group properties.}

\item{indexName}{The dynamic thing group index name.

Currently one index is supported: \"AWS\\_Things\".}

\item{queryString}{[required] The dynamic thing group search query string.

See \href{https://docs.aws.amazon.com/iot/latest/developerguide/query-syntax.html}{Query Syntax}
for information about query string syntax.}

\item{queryVersion}{The dynamic thing group query version.

Currently one query version is supported: \"2017-09-30\". If not
specified, the query version defaults to this value.}

\item{tags}{Metadata which can be used to manage the dynamic thing group.}
}
\description{
Creates a dynamic thing group.
}
\section{Request syntax}{
\preformatted{svc$create_dynamic_thing_group(
  thingGroupName = "string",
  thingGroupProperties = list(
    thingGroupDescription = "string",
    attributePayload = list(
      attributes = list(
        "string"
      ),
      merge = TRUE|FALSE
    )
  ),
  indexName = "string",
  queryString = "string",
  queryVersion = "string",
  tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
