% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_create_job}
\alias{iot_create_job}
\title{Creates a job}
\usage{
iot_create_job(jobId, targets, documentSource, document, description,
  presignedUrlConfig, targetSelection, jobExecutionsRolloutConfig,
  abortConfig, timeoutConfig, tags)
}
\arguments{
\item{jobId}{[required] A job identifier which must be unique for your AWS account. We recommend
using a UUID. Alpha-numeric characters, "-" and "\\_" are valid for use
here.}

\item{targets}{[required] A list of things and thing groups to which the job should be sent.}

\item{documentSource}{An S3 link to the job document.}

\item{document}{The job document.

If the job document resides in an S3 bucket, you must use a placeholder
link when specifying the document.

The placeholder link is of the following form:

\verb{$\\\{aws:iot:s3-presigned-url:https://s3.amazonaws.com/<i>bucket</i>/<i>key</i>\\\}}

where \emph{bucket} is your bucket name and \emph{key} is the object in the bucket
to which you are linking.}

\item{description}{A short text description of the job.}

\item{presignedUrlConfig}{Configuration information for pre-signed S3 URLs.}

\item{targetSelection}{Specifies whether the job will continue to run (CONTINUOUS), or will be
complete after all those things specified as targets have completed the
job (SNAPSHOT). If continuous, the job may also be run on a thing when a
change is detected in a target. For example, a job will run on a thing
when the thing is added to a target group, even after the job was
completed by all things originally in the group.}

\item{jobExecutionsRolloutConfig}{Allows you to create a staged rollout of the job.}

\item{abortConfig}{Allows you to create criteria to abort a job.}

\item{timeoutConfig}{Specifies the amount of time each device has to finish its execution of
the job. The timer is started when the job execution status is set to
\code{IN_PROGRESS}. If the job execution status is not set to another
terminal state before the time expires, it will be automatically set to
\code{TIMED_OUT}.}

\item{tags}{Metadata which can be used to manage the job.}
}
\description{
Creates a job.
}
\section{Request syntax}{
\preformatted{svc$create_job(
  jobId = "string",
  targets = list(
    "string"
  ),
  documentSource = "string",
  document = "string",
  description = "string",
  presignedUrlConfig = list(
    roleArn = "string",
    expiresInSec = 123
  ),
  targetSelection = "CONTINUOUS"|"SNAPSHOT",
  jobExecutionsRolloutConfig = list(
    maximumPerMinute = 123,
    exponentialRate = list(
      baseRatePerMinute = 123,
      incrementFactor = 123.0,
      rateIncreaseCriteria = list(
        numberOfNotifiedThings = 123,
        numberOfSucceededThings = 123
      )
    )
  ),
  abortConfig = list(
    criteriaList = list(
      list(
        failureType = "FAILED"|"REJECTED"|"TIMED_OUT"|"ALL",
        action = "CANCEL",
        thresholdPercentage = 123.0,
        minNumberOfExecutedThings = 123
      )
    )
  ),
  timeoutConfig = list(
    inProgressTimeoutInMinutes = 123
  ),
  tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
