% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_create_keys_and_certificate}
\alias{iot_create_keys_and_certificate}
\title{Creates a 2048-bit RSA key pair and issues an X}
\usage{
iot_create_keys_and_certificate(setAsActive)
}
\arguments{
\item{setAsActive}{Specifies whether the certificate is active.}
}
\value{
A list with the following syntax:\preformatted{list(
  certificateArn = "string",
  certificateId = "string",
  certificatePem = "string",
  keyPair = list(
    PublicKey = "string",
    PrivateKey = "string"
  )
)
}
}
\description{
Creates a 2048-bit RSA key pair and issues an X.509 certificate using
the issued public key. You can also call
\code{\link[=iot_create_keys_and_certificate]{create_keys_and_certificate}} over
MQTT from a device, for more information, see \href{https://docs.aws.amazon.com/iot/latest/developerguide/provision-wo-cert.html#provision-mqtt-api}{Provisioning MQTT API}.

\strong{Note} This is the only time AWS IoT issues the private key for this
certificate, so it is important to keep it in a secure location.
}
\section{Request syntax}{
\preformatted{svc$create_keys_and_certificate(
  setAsActive = TRUE|FALSE
)
}
}

\keyword{internal}
