% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_create_policy_version}
\alias{iot_create_policy_version}
\title{Creates a new version of the specified AWS IoT policy}
\usage{
iot_create_policy_version(policyName, policyDocument, setAsDefault)
}
\arguments{
\item{policyName}{[required] The policy name.}

\item{policyDocument}{[required] The JSON document that describes the policy. Minimum length of 1.
Maximum length of 2048, excluding whitespace.}

\item{setAsDefault}{Specifies whether the policy version is set as the default. When this
parameter is true, the new policy version becomes the operative version
(that is, the version that is in effect for the certificates to which
the policy is attached).}
}
\value{
A list with the following syntax:\preformatted{list(
  policyArn = "string",
  policyDocument = "string",
  policyVersionId = "string",
  isDefaultVersion = TRUE|FALSE
)
}
}
\description{
Creates a new version of the specified AWS IoT policy. To update a
policy, create a new policy version. A managed policy can have up to
five versions. If the policy has five versions, you must use
\code{\link[=iot_delete_policy_version]{delete_policy_version}} to delete an
existing version before you create a new one.

Optionally, you can set the new version as the policy's default version.
The default version is the operative version (that is, the version that
is in effect for the certificates to which the policy is attached).
}
\section{Request syntax}{
\preformatted{svc$create_policy_version(
  policyName = "string",
  policyDocument = "string",
  setAsDefault = TRUE|FALSE
)
}
}

\keyword{internal}
