% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_create_provisioning_claim}
\alias{iot_create_provisioning_claim}
\title{Creates a provisioning claim}
\usage{
iot_create_provisioning_claim(templateName)
}
\arguments{
\item{templateName}{[required] The name of the provisioning template to use.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  certificateId = "string",
  certificatePem = "string",
  keyPair = list(
    PublicKey = "string",
    PrivateKey = "string"
  ),
  expiration = as.POSIXct(
    "2015-01-01"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Creates a provisioning claim.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_provisioning_claim(
  templateName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
