% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_create_provisioning_template}
\alias{iot_create_provisioning_template}
\title{Creates a fleet provisioning template}
\usage{
iot_create_provisioning_template(templateName, description,
  templateBody, enabled, provisioningRoleArn, preProvisioningHook, tags)
}
\arguments{
\item{templateName}{[required] The name of the fleet provisioning template.}

\item{description}{The description of the fleet provisioning template.}

\item{templateBody}{[required] The JSON formatted contents of the fleet provisioning template.}

\item{enabled}{True to enable the fleet provisioning template, otherwise false.}

\item{provisioningRoleArn}{[required] The role ARN for the role associated with the fleet provisioning
template. This IoT role grants permission to provision a device.}

\item{preProvisioningHook}{Creates a pre-provisioning hook template.}

\item{tags}{Metadata which can be used to manage the fleet provisioning template.

For URI Request parameters use format: ...key1=value1&key2=value2...

For the CLI command-line parameter use format: &&tags
"key1=value1&key2=value2..."

For the cli-input-json file use format: "tags":
"key1=value1&key2=value2..."}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  templateArn = "string",
  templateName = "string",
  defaultVersionId = 123
)
}\if{html}{\out{</div>}}
}
\description{
Creates a fleet provisioning template.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_provisioning_template(
  templateName = "string",
  description = "string",
  templateBody = "string",
  enabled = TRUE|FALSE,
  provisioningRoleArn = "string",
  preProvisioningHook = list(
    payloadVersion = "string",
    targetArn = "string"
  ),
  tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
