% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_create_role_alias}
\alias{iot_create_role_alias}
\title{Creates a role alias}
\usage{
iot_create_role_alias(roleAlias, roleArn, credentialDurationSeconds,
  tags)
}
\arguments{
\item{roleAlias}{[required] The role alias that points to a role ARN. This allows you to change the
role without having to update the device.}

\item{roleArn}{[required] The role ARN.}

\item{credentialDurationSeconds}{How long (in seconds) the credentials will be valid.}

\item{tags}{Metadata which can be used to manage the role alias.

For URI Request parameters use format: ...key1=value1&key2=value2...

For the CLI command-line parameter use format: &&tags
"key1=value1&key2=value2..."

For the cli-input-json file use format: "tags":
"key1=value1&key2=value2..."}
}
\value{
A list with the following syntax:\preformatted{list(
  roleAlias = "string",
  roleAliasArn = "string"
)
}
}
\description{
Creates a role alias.
}
\section{Request syntax}{
\preformatted{svc$create_role_alias(
  roleAlias = "string",
  roleArn = "string",
  credentialDurationSeconds = 123,
  tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
