% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_create_scheduled_audit}
\alias{iot_create_scheduled_audit}
\title{Creates a scheduled audit that is run at a specified time interval}
\usage{
iot_create_scheduled_audit(frequency, dayOfMonth, dayOfWeek,
  targetCheckNames, scheduledAuditName, tags)
}
\arguments{
\item{frequency}{[required] How often the scheduled audit takes place. Can be one of \"DAILY\",
\"WEEKLY\", \"BIWEEKLY\" or \"MONTHLY\". The start time of each audit is
determined by the system.}

\item{dayOfMonth}{The day of the month on which the scheduled audit takes place. Can be
\"1\" through \"31\" or \"LAST\". This field is required if the
\"frequency\" parameter is set to \"MONTHLY\". If days 29-31 are
specified, and the month does not have that many days, the audit takes
place on the \"LAST\" day of the month.}

\item{dayOfWeek}{The day of the week on which the scheduled audit takes place. Can be one
of \"SUN\", \"MON\", \"TUE\", \"WED\", \"THU\", \"FRI\", or \"SAT\".
This field is required if the \"frequency\" parameter is set to
\"WEEKLY\" or \"BIWEEKLY\".}

\item{targetCheckNames}{[required] Which checks are performed during the scheduled audit. Checks must be
enabled for your account. (Use \code{DescribeAccountAuditConfiguration} to
see the list of all checks, including those that are enabled or use
\code{UpdateAccountAuditConfiguration} to select which checks are enabled.)}

\item{scheduledAuditName}{[required] The name you want to give to the scheduled audit. (Max. 128 chars)}

\item{tags}{Metadata that can be used to manage the scheduled audit.}
}
\description{
Creates a scheduled audit that is run at a specified time interval.
}
\section{Request syntax}{
\preformatted{svc$create_scheduled_audit(
  frequency = "DAILY"|"WEEKLY"|"BIWEEKLY"|"MONTHLY",
  dayOfMonth = "string",
  dayOfWeek = "SUN"|"MON"|"TUE"|"WED"|"THU"|"FRI"|"SAT",
  targetCheckNames = list(
    "string"
  ),
  scheduledAuditName = "string",
  tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
