% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_create_security_profile}
\alias{iot_create_security_profile}
\title{Creates a Device Defender security profile}
\usage{
iot_create_security_profile(securityProfileName,
  securityProfileDescription, behaviors, alertTargets,
  additionalMetricsToRetain, additionalMetricsToRetainV2, tags)
}
\arguments{
\item{securityProfileName}{[required] The name you are giving to the security profile.}

\item{securityProfileDescription}{A description of the security profile.}

\item{behaviors}{Specifies the behaviors that, when violated by a device (thing), cause
an alert.}

\item{alertTargets}{Specifies the destinations to which alerts are sent. (Alerts are always
sent to the console.) Alerts are generated when a device (thing)
violates a behavior.}

\item{additionalMetricsToRetain}{\emph{Please use CreateSecurityProfileRequest$additionalMetricsToRetainV2
instead.}

A list of metrics whose data is retained (stored). By default, data is
retained for any metric used in the profile's \code{behaviors}, but it is
also retained for any metric specified here. Can be used with custom
metrics; cannot be used with dimensions.}

\item{additionalMetricsToRetainV2}{A list of metrics whose data is retained (stored). By default, data is
retained for any metric used in the profile's \code{behaviors}, but it is
also retained for any metric specified here. Can be used with custom
metrics; cannot be used with dimensions.}

\item{tags}{Metadata that can be used to manage the security profile.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  securityProfileName = "string",
  securityProfileArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Creates a Device Defender security profile.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_security_profile(
  securityProfileName = "string",
  securityProfileDescription = "string",
  behaviors = list(
    list(
      name = "string",
      metric = "string",
      metricDimension = list(
        dimensionName = "string",
        operator = "IN"|"NOT_IN"
      ),
      criteria = list(
        comparisonOperator = "less-than"|"less-than-equals"|"greater-than"|"greater-than-equals"|"in-cidr-set"|"not-in-cidr-set"|"in-port-set"|"not-in-port-set"|"in-set"|"not-in-set",
        value = list(
          count = 123,
          cidrs = list(
            "string"
          ),
          ports = list(
            123
          ),
          number = 123.0,
          numbers = list(
            123.0
          ),
          strings = list(
            "string"
          )
        ),
        durationSeconds = 123,
        consecutiveDatapointsToAlarm = 123,
        consecutiveDatapointsToClear = 123,
        statisticalThreshold = list(
          statistic = "string"
        ),
        mlDetectionConfig = list(
          confidenceLevel = "LOW"|"MEDIUM"|"HIGH"
        )
      ),
      suppressAlerts = TRUE|FALSE
    )
  ),
  alertTargets = list(
    list(
      alertTargetArn = "string",
      roleArn = "string"
    )
  ),
  additionalMetricsToRetain = list(
    "string"
  ),
  additionalMetricsToRetainV2 = list(
    list(
      metric = "string",
      metricDimension = list(
        dimensionName = "string",
        operator = "IN"|"NOT_IN"
      )
    )
  ),
  tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
