% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_create_thing_group}
\alias{iot_create_thing_group}
\title{Create a thing group}
\usage{
iot_create_thing_group(thingGroupName, parentGroupName,
  thingGroupProperties, tags)
}
\arguments{
\item{thingGroupName}{[required] The thing group name to create.}

\item{parentGroupName}{The name of the parent thing group.}

\item{thingGroupProperties}{The thing group properties.}

\item{tags}{Metadata which can be used to manage the thing group.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  thingGroupName = "string",
  thingGroupArn = "string",
  thingGroupId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Create a thing group.

This is a control plane operation. See
\href{https://docs.aws.amazon.com/iot/latest/developerguide/iot-authorization.html}{Authorization}
for information about authorizing control plane actions.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_thing_group(
  thingGroupName = "string",
  parentGroupName = "string",
  thingGroupProperties = list(
    thingGroupDescription = "string",
    attributePayload = list(
      attributes = list(
        "string"
      ),
      merge = TRUE|FALSE
    )
  ),
  tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
