% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_create_thing_type}
\alias{iot_create_thing_type}
\title{Creates a new thing type}
\usage{
iot_create_thing_type(thingTypeName, thingTypeProperties, tags)
}
\arguments{
\item{thingTypeName}{[required] The name of the thing type.}

\item{thingTypeProperties}{The ThingTypeProperties for the thing type to create. It contains
information about the new thing type including a description, and a list
of searchable thing attribute names.}

\item{tags}{Metadata which can be used to manage the thing type.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  thingTypeName = "string",
  thingTypeArn = "string",
  thingTypeId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Creates a new thing type.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_thing_type(
  thingTypeName = "string",
  thingTypeProperties = list(
    thingTypeDescription = "string",
    searchableAttributes = list(
      "string"
    )
  ),
  tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
