% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_create_topic_rule_destination}
\alias{iot_create_topic_rule_destination}
\title{Creates a topic rule destination}
\usage{
iot_create_topic_rule_destination(destinationConfiguration)
}
\arguments{
\item{destinationConfiguration}{[required] The topic rule destination configuration.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  topicRuleDestination = list(
    arn = "string",
    status = "ENABLED"|"IN_PROGRESS"|"DISABLED"|"ERROR"|"DELETING",
    createdAt = as.POSIXct(
      "2015-01-01"
    ),
    lastUpdatedAt = as.POSIXct(
      "2015-01-01"
    ),
    statusReason = "string",
    httpUrlProperties = list(
      confirmationUrl = "string"
    ),
    vpcProperties = list(
      subnetIds = list(
        "string"
      ),
      securityGroups = list(
        "string"
      ),
      vpcId = "string",
      roleArn = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Creates a topic rule destination. The destination must be confirmed
prior to use.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_topic_rule_destination(
  destinationConfiguration = list(
    httpUrlConfiguration = list(
      confirmationUrl = "string"
    ),
    vpcConfiguration = list(
      subnetIds = list(
        "string"
      ),
      securityGroups = list(
        "string"
      ),
      vpcId = "string",
      roleArn = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
