% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_delete_account_audit_configuration}
\alias{iot_delete_account_audit_configuration}
\title{Restores the default settings for Device Defender audits for this
account}
\usage{
iot_delete_account_audit_configuration(deleteScheduledAudits)
}
\arguments{
\item{deleteScheduledAudits}{If true, all scheduled audits are deleted.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Restores the default settings for Device Defender audits for this
account. Any configuration data you entered is deleted and all audit
checks are reset to disabled.
}
\section{Request syntax}{
\preformatted{svc$delete_account_audit_configuration(
  deleteScheduledAudits = TRUE|FALSE
)
}
}

\keyword{internal}
