% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_delete_audit_suppression}
\alias{iot_delete_audit_suppression}
\title{Deletes a Device Defender audit suppression}
\usage{
iot_delete_audit_suppression(checkName, resourceIdentifier)
}
\arguments{
\item{checkName}{[required]}

\item{resourceIdentifier}{[required]}
}
\value{
An empty list.
}
\description{
Deletes a Device Defender audit suppression.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_audit_suppression(
  checkName = "string",
  resourceIdentifier = list(
    deviceCertificateId = "string",
    caCertificateId = "string",
    cognitoIdentityPoolId = "string",
    clientId = "string",
    policyVersionIdentifier = list(
      policyName = "string",
      policyVersionId = "string"
    ),
    account = "string",
    iamRoleArn = "string",
    roleAliasArn = "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
