% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_delete_billing_group}
\alias{iot_delete_billing_group}
\title{Deletes the billing group}
\usage{
iot_delete_billing_group(billingGroupName, expectedVersion)
}
\arguments{
\item{billingGroupName}{[required] The name of the billing group.}

\item{expectedVersion}{The expected version of the billing group. If the version of the billing
group does not match the expected version specified in the request, the
\code{\link[=iot_delete_billing_group]{delete_billing_group}} request is rejected
with a \code{VersionConflictException}.}
}
\value{
An empty list.
}
\description{
Deletes the billing group.
}
\section{Request syntax}{
\preformatted{svc$delete_billing_group(
  billingGroupName = "string",
  expectedVersion = 123
)
}
}

\keyword{internal}
