% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_delete_ca_certificate}
\alias{iot_delete_ca_certificate}
\title{Deletes a registered CA certificate}
\usage{
iot_delete_ca_certificate(certificateId)
}
\arguments{
\item{certificateId}{[required] The ID of the certificate to delete. (The last part of the certificate
ARN contains the certificate ID.)}
}
\description{
Deletes a registered CA certificate.
}
\section{Request syntax}{
\preformatted{svc$delete_ca_certificate(
  certificateId = "string"
)
}
}

\keyword{internal}
