% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_delete_certificate}
\alias{iot_delete_certificate}
\title{Deletes the specified certificate}
\usage{
iot_delete_certificate(certificateId, forceDelete)
}
\arguments{
\item{certificateId}{[required] The ID of the certificate. (The last part of the certificate ARN
contains the certificate ID.)}

\item{forceDelete}{Forces the deletion of a certificate if it is inactive and is not
attached to an IoT thing.}
}
\description{
Deletes the specified certificate.
}
\details{
A certificate cannot be deleted if it has a policy or IoT thing attached
to it or if its status is set to ACTIVE. To delete a certificate, first
use the DetachPrincipalPolicy API to detach all policies. Next, use the
UpdateCertificate API to set the certificate to the INACTIVE status.
}
\section{Request syntax}{
\preformatted{svc$delete_certificate(
  certificateId = "string",
  forceDelete = TRUE|FALSE
)
}
}

\keyword{internal}
