% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_delete_dimension}
\alias{iot_delete_dimension}
\title{Removes the specified dimension from your AWS account}
\usage{
iot_delete_dimension(name)
}
\arguments{
\item{name}{[required] The unique identifier for the dimension that you want to delete.}
}
\description{
Removes the specified dimension from your AWS account.
}
\section{Request syntax}{
\preformatted{svc$delete_dimension(
  name = "string"
)
}
}

\keyword{internal}
