% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_delete_dynamic_thing_group}
\alias{iot_delete_dynamic_thing_group}
\title{Deletes a dynamic thing group}
\usage{
iot_delete_dynamic_thing_group(thingGroupName, expectedVersion)
}
\arguments{
\item{thingGroupName}{[required] The name of the dynamic thing group to delete.}

\item{expectedVersion}{The expected version of the dynamic thing group to delete.}
}
\value{
An empty list.
}
\description{
Deletes a dynamic thing group.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_dynamic_thing_group(
  thingGroupName = "string",
  expectedVersion = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
