% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_delete_mitigation_action}
\alias{iot_delete_mitigation_action}
\title{Deletes a defined mitigation action from your AWS account}
\usage{
iot_delete_mitigation_action(actionName)
}
\arguments{
\item{actionName}{[required] The name of the mitigation action that you want to delete.}
}
\value{
An empty list.
}
\description{
Deletes a defined mitigation action from your AWS account.
}
\section{Request syntax}{
\preformatted{svc$delete_mitigation_action(
  actionName = "string"
)
}
}

\keyword{internal}
