% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_delete_ota_update}
\alias{iot_delete_ota_update}
\title{Delete an OTA update}
\usage{
iot_delete_ota_update(otaUpdateId, deleteStream, forceDeleteAWSJob)
}
\arguments{
\item{otaUpdateId}{[required] The ID of the OTA update to delete.}

\item{deleteStream}{When true, the stream created by the OTAUpdate process is deleted when
the OTA update is deleted. Ignored if the stream specified in the
OTAUpdate is supplied by the user.}

\item{forceDeleteAWSJob}{When true, deletes the AWS job created by the OTAUpdate process even if
it is "IN_PROGRESS". Otherwise, if the job is not in a terminal state
("COMPLETED" or "CANCELED") an exception will occur. The default is
false.}
}
\value{
An empty list.
}
\description{
Delete an OTA update.
}
\section{Request syntax}{
\preformatted{svc$delete_ota_update(
  otaUpdateId = "string",
  deleteStream = TRUE|FALSE,
  forceDeleteAWSJob = TRUE|FALSE
)
}
}

\keyword{internal}
