% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_delete_policy}
\alias{iot_delete_policy}
\title{Deletes the specified policy}
\usage{
iot_delete_policy(policyName)
}
\arguments{
\item{policyName}{[required] The name of the policy to delete.}
}
\description{
Deletes the specified policy.
}
\details{
A policy cannot be deleted if it has non-default versions or it is
attached to any certificate.

To delete a policy, use the DeletePolicyVersion API to delete all
non-default versions of the policy; use the DetachPrincipalPolicy API to
detach the policy from any certificate; and then use the DeletePolicy
API to delete the policy.

When a policy is deleted using DeletePolicy, its default version is
deleted with it.
}
\section{Request syntax}{
\preformatted{svc$delete_policy(
  policyName = "string"
)
}
}

\keyword{internal}
