% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_delete_provisioning_template}
\alias{iot_delete_provisioning_template}
\title{Deletes a fleet provisioning template}
\usage{
iot_delete_provisioning_template(templateName)
}
\arguments{
\item{templateName}{[required] The name of the fleet provision template to delete.}
}
\value{
An empty list.
}
\description{
Deletes a fleet provisioning template.
}
\section{Request syntax}{
\preformatted{svc$delete_provisioning_template(
  templateName = "string"
)
}
}

\keyword{internal}
