% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_delete_provisioning_template_version}
\alias{iot_delete_provisioning_template_version}
\title{Deletes a fleet provisioning template version}
\usage{
iot_delete_provisioning_template_version(templateName, versionId)
}
\arguments{
\item{templateName}{[required] The name of the fleet provisioning template version to delete.}

\item{versionId}{[required] The fleet provisioning template version ID to delete.}
}
\value{
An empty list.
}
\description{
Deletes a fleet provisioning template version.
}
\section{Request syntax}{
\preformatted{svc$delete_provisioning_template_version(
  templateName = "string",
  versionId = 123
)
}
}

\keyword{internal}
