% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_delete_registration_code}
\alias{iot_delete_registration_code}
\title{Deletes a CA certificate registration code}
\usage{
iot_delete_registration_code()
}
\value{
An empty list.
}
\description{
Deletes a CA certificate registration code.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_registration_code()
}\if{html}{\out{</div>}}
}

\keyword{internal}
