% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_delete_scheduled_audit}
\alias{iot_delete_scheduled_audit}
\title{Deletes a scheduled audit}
\usage{
iot_delete_scheduled_audit(scheduledAuditName)
}
\arguments{
\item{scheduledAuditName}{[required] The name of the scheduled audit you want to delete.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Deletes a scheduled audit.
}
\section{Request syntax}{
\preformatted{svc$delete_scheduled_audit(
  scheduledAuditName = "string"
)
}
}

\keyword{internal}
