% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_delete_stream}
\alias{iot_delete_stream}
\title{Deletes a stream}
\usage{
iot_delete_stream(streamId)
}
\arguments{
\item{streamId}{[required] The stream ID.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Deletes a stream.
}
\section{Request syntax}{
\preformatted{svc$delete_stream(
  streamId = "string"
)
}
}

\keyword{internal}
