% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_delete_thing_type}
\alias{iot_delete_thing_type}
\title{Deletes the specified thing type}
\usage{
iot_delete_thing_type(thingTypeName)
}
\arguments{
\item{thingTypeName}{[required] The name of the thing type.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Deletes the specified thing type. You cannot delete a thing type if it
has things associated with it. To delete a thing type, first mark it as
deprecated by calling
\code{\link[=iot_deprecate_thing_type]{deprecate_thing_type}}, then remove any
associated things by calling \code{\link[=iot_update_thing]{update_thing}} to
change the thing type on any associated thing, and finally use
\code{\link[=iot_delete_thing_type]{delete_thing_type}} to delete the thing type.
}
\section{Request syntax}{
\preformatted{svc$delete_thing_type(
  thingTypeName = "string"
)
}
}

\keyword{internal}
