% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_delete_topic_rule}
\alias{iot_delete_topic_rule}
\title{Deletes the rule}
\usage{
iot_delete_topic_rule(ruleName)
}
\arguments{
\item{ruleName}{[required] The name of the rule.}
}
\value{
An empty list.
}
\description{
Deletes the rule.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_topic_rule(
  ruleName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
