% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_deprecate_thing_type}
\alias{iot_deprecate_thing_type}
\title{Deprecates a thing type}
\usage{
iot_deprecate_thing_type(thingTypeName, undoDeprecate)
}
\arguments{
\item{thingTypeName}{[required] The name of the thing type to deprecate.}

\item{undoDeprecate}{Whether to undeprecate a deprecated thing type. If \strong{true}, the thing
type will not be deprecated anymore and you can associate it with
things.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Deprecates a thing type. You can not associate new things with
deprecated thing type.
}
\section{Request syntax}{
\preformatted{svc$deprecate_thing_type(
  thingTypeName = "string",
  undoDeprecate = TRUE|FALSE
)
}
}

\keyword{internal}
