% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_account_audit_configuration}
\alias{iot_describe_account_audit_configuration}
\title{Gets information about the Device Defender audit settings for this
account}
\usage{
iot_describe_account_audit_configuration()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  roleArn = "string",
  auditNotificationTargetConfigurations = list(
    list(
      targetArn = "string",
      roleArn = "string",
      enabled = TRUE|FALSE
    )
  ),
  auditCheckConfigurations = list(
    list(
      enabled = TRUE|FALSE
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets information about the Device Defender audit settings for this
account. Settings include how audit notifications are sent and which
audit checks are enabled or disabled.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_account_audit_configuration()
}\if{html}{\out{</div>}}
}

\keyword{internal}
