% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_audit_task}
\alias{iot_describe_audit_task}
\title{Gets information about a Device Defender audit}
\usage{
iot_describe_audit_task(taskId)
}
\arguments{
\item{taskId}{[required] The ID of the audit whose information you want to get.}
}
\description{
Gets information about a Device Defender audit.
}
\section{Request syntax}{
\preformatted{svc$describe_audit_task(
  taskId = "string"
)
}
}

\keyword{internal}
