% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_billing_group}
\alias{iot_describe_billing_group}
\title{Returns information about a billing group}
\usage{
iot_describe_billing_group(billingGroupName)
}
\arguments{
\item{billingGroupName}{[required] The name of the billing group.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  billingGroupName = "string",
  billingGroupId = "string",
  billingGroupArn = "string",
  version = 123,
  billingGroupProperties = list(
    billingGroupDescription = "string"
  ),
  billingGroupMetadata = list(
    creationDate = as.POSIXct(
      "2015-01-01"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns information about a billing group.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_billing_group(
  billingGroupName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
