% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_ca_certificate}
\alias{iot_describe_ca_certificate}
\title{Describes a registered CA certificate}
\usage{
iot_describe_ca_certificate(certificateId)
}
\arguments{
\item{certificateId}{[required] The CA certificate identifier.}
}
\value{
A list with the following syntax:\preformatted{list(
  certificateDescription = list(
    certificateArn = "string",
    certificateId = "string",
    status = "ACTIVE"|"INACTIVE",
    certificatePem = "string",
    ownedBy = "string",
    creationDate = as.POSIXct(
      "2015-01-01"
    ),
    autoRegistrationStatus = "ENABLE"|"DISABLE",
    lastModifiedDate = as.POSIXct(
      "2015-01-01"
    ),
    customerVersion = 123,
    generationId = "string",
    validity = list(
      notBefore = as.POSIXct(
        "2015-01-01"
      ),
      notAfter = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  registrationConfig = list(
    templateBody = "string",
    roleArn = "string"
  )
)
}
}
\description{
Describes a registered CA certificate.
}
\section{Request syntax}{
\preformatted{svc$describe_ca_certificate(
  certificateId = "string"
)
}
}

\keyword{internal}
