% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_detect_mitigation_actions_task}
\alias{iot_describe_detect_mitigation_actions_task}
\title{Gets information about a Device Defender ML Detect mitigation action}
\usage{
iot_describe_detect_mitigation_actions_task(taskId)
}
\arguments{
\item{taskId}{[required] The unique identifier of the task.}
}
\description{
Gets information about a Device Defender ML Detect mitigation action.
}
\section{Request syntax}{
\preformatted{svc$describe_detect_mitigation_actions_task(
  taskId = "string"
)
}
}

\keyword{internal}
