% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_domain_configuration}
\alias{iot_describe_domain_configuration}
\title{Gets summary information about a domain configuration}
\usage{
iot_describe_domain_configuration(domainConfigurationName)
}
\arguments{
\item{domainConfigurationName}{[required] The name of the domain configuration.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  domainConfigurationName = "string",
  domainConfigurationArn = "string",
  domainName = "string",
  serverCertificates = list(
    list(
      serverCertificateArn = "string",
      serverCertificateStatus = "INVALID"|"VALID",
      serverCertificateStatusDetail = "string"
    )
  ),
  authorizerConfig = list(
    defaultAuthorizerName = "string",
    allowAuthorizerOverride = TRUE|FALSE
  ),
  domainConfigurationStatus = "ENABLED"|"DISABLED",
  serviceType = "DATA"|"CREDENTIAL_PROVIDER"|"JOBS",
  domainType = "ENDPOINT"|"AWS_MANAGED"|"CUSTOMER_MANAGED",
  lastStatusChangeDate = as.POSIXct(
    "2015-01-01"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets summary information about a domain configuration.

The domain configuration feature is in public preview and is subject to
change.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_domain_configuration(
  domainConfigurationName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
