% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_mitigation_action}
\alias{iot_describe_mitigation_action}
\title{Gets information about a mitigation action}
\usage{
iot_describe_mitigation_action(actionName)
}
\arguments{
\item{actionName}{[required] The friendly name that uniquely identifies the mitigation action.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  actionName = "string",
  actionType = "UPDATE_DEVICE_CERTIFICATE"|"UPDATE_CA_CERTIFICATE"|"ADD_THINGS_TO_THING_GROUP"|"REPLACE_DEFAULT_POLICY_VERSION"|"ENABLE_IOT_LOGGING"|"PUBLISH_FINDING_TO_SNS",
  actionArn = "string",
  actionId = "string",
  roleArn = "string",
  actionParams = list(
    updateDeviceCertificateParams = list(
      action = "DEACTIVATE"
    ),
    updateCACertificateParams = list(
      action = "DEACTIVATE"
    ),
    addThingsToThingGroupParams = list(
      thingGroupNames = list(
        "string"
      ),
      overrideDynamicGroups = TRUE|FALSE
    ),
    replaceDefaultPolicyVersionParams = list(
      templateName = "BLANK_POLICY"
    ),
    enableIoTLoggingParams = list(
      roleArnForLogging = "string",
      logLevel = "DEBUG"|"INFO"|"ERROR"|"WARN"|"DISABLED"
    ),
    publishFindingToSnsParams = list(
      topicArn = "string"
    )
  ),
  creationDate = as.POSIXct(
    "2015-01-01"
  ),
  lastModifiedDate = as.POSIXct(
    "2015-01-01"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets information about a mitigation action.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_mitigation_action(
  actionName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
