% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_provisioning_template}
\alias{iot_describe_provisioning_template}
\title{Returns information about a fleet provisioning template}
\usage{
iot_describe_provisioning_template(templateName)
}
\arguments{
\item{templateName}{[required] The name of the fleet provisioning template.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  templateArn = "string",
  templateName = "string",
  description = "string",
  creationDate = as.POSIXct(
    "2015-01-01"
  ),
  lastModifiedDate = as.POSIXct(
    "2015-01-01"
  ),
  defaultVersionId = 123,
  templateBody = "string",
  enabled = TRUE|FALSE,
  provisioningRoleArn = "string",
  preProvisioningHook = list(
    payloadVersion = "string",
    targetArn = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns information about a fleet provisioning template.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_provisioning_template(
  templateName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
