% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_provisioning_template_version}
\alias{iot_describe_provisioning_template_version}
\title{Returns information about a fleet provisioning template version}
\usage{
iot_describe_provisioning_template_version(templateName, versionId)
}
\arguments{
\item{templateName}{[required] The template name.}

\item{versionId}{[required] The fleet provisioning template version ID.}
}
\value{
A list with the following syntax:\preformatted{list(
  versionId = 123,
  creationDate = as.POSIXct(
    "2015-01-01"
  ),
  templateBody = "string",
  isDefaultVersion = TRUE|FALSE
)
}
}
\description{
Returns information about a fleet provisioning template version.
}
\section{Request syntax}{
\preformatted{svc$describe_provisioning_template_version(
  templateName = "string",
  versionId = 123
)
}
}

\keyword{internal}
