% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_role_alias}
\alias{iot_describe_role_alias}
\title{Describes a role alias}
\usage{
iot_describe_role_alias(roleAlias)
}
\arguments{
\item{roleAlias}{[required] The role alias to describe.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  roleAliasDescription = list(
    roleAlias = "string",
    roleAliasArn = "string",
    roleArn = "string",
    owner = "string",
    credentialDurationSeconds = 123,
    creationDate = as.POSIXct(
      "2015-01-01"
    ),
    lastModifiedDate = as.POSIXct(
      "2015-01-01"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes a role alias.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_role_alias(
  roleAlias = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
