% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_security_profile}
\alias{iot_describe_security_profile}
\title{Gets information about a Device Defender security profile}
\usage{
iot_describe_security_profile(securityProfileName)
}
\arguments{
\item{securityProfileName}{[required] The name of the security profile whose information you want to get.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  securityProfileName = "string",
  securityProfileArn = "string",
  securityProfileDescription = "string",
  behaviors = list(
    list(
      name = "string",
      metric = "string",
      metricDimension = list(
        dimensionName = "string",
        operator = "IN"|"NOT_IN"
      ),
      criteria = list(
        comparisonOperator = "less-than"|"less-than-equals"|"greater-than"|"greater-than-equals"|"in-cidr-set"|"not-in-cidr-set"|"in-port-set"|"not-in-port-set"|"in-set"|"not-in-set",
        value = list(
          count = 123,
          cidrs = list(
            "string"
          ),
          ports = list(
            123
          ),
          number = 123.0,
          numbers = list(
            123.0
          ),
          strings = list(
            "string"
          )
        ),
        durationSeconds = 123,
        consecutiveDatapointsToAlarm = 123,
        consecutiveDatapointsToClear = 123,
        statisticalThreshold = list(
          statistic = "string"
        ),
        mlDetectionConfig = list(
          confidenceLevel = "LOW"|"MEDIUM"|"HIGH"
        )
      ),
      suppressAlerts = TRUE|FALSE
    )
  ),
  alertTargets = list(
    list(
      alertTargetArn = "string",
      roleArn = "string"
    )
  ),
  additionalMetricsToRetain = list(
    "string"
  ),
  additionalMetricsToRetainV2 = list(
    list(
      metric = "string",
      metricDimension = list(
        dimensionName = "string",
        operator = "IN"|"NOT_IN"
      )
    )
  ),
  version = 123,
  creationDate = as.POSIXct(
    "2015-01-01"
  ),
  lastModifiedDate = as.POSIXct(
    "2015-01-01"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets information about a Device Defender security profile.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_security_profile(
  securityProfileName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
