% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_thing}
\alias{iot_describe_thing}
\title{Gets information about the specified thing}
\usage{
iot_describe_thing(thingName)
}
\arguments{
\item{thingName}{[required] The name of the thing.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  defaultClientId = "string",
  thingName = "string",
  thingId = "string",
  thingArn = "string",
  thingTypeName = "string",
  attributes = list(
    "string"
  ),
  version = 123,
  billingGroupName = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Gets information about the specified thing.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_thing(
  thingName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
