% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_thing_group}
\alias{iot_describe_thing_group}
\title{Describe a thing group}
\usage{
iot_describe_thing_group(thingGroupName)
}
\arguments{
\item{thingGroupName}{[required] The name of the thing group.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  thingGroupName = "string",
  thingGroupId = "string",
  thingGroupArn = "string",
  version = 123,
  thingGroupProperties = list(
    thingGroupDescription = "string",
    attributePayload = list(
      attributes = list(
        "string"
      ),
      merge = TRUE|FALSE
    )
  ),
  thingGroupMetadata = list(
    parentGroupName = "string",
    rootToParentThingGroups = list(
      list(
        groupName = "string",
        groupArn = "string"
      )
    ),
    creationDate = as.POSIXct(
      "2015-01-01"
    )
  ),
  indexName = "string",
  queryString = "string",
  queryVersion = "string",
  status = "ACTIVE"|"BUILDING"|"REBUILDING"
)
}\if{html}{\out{</div>}}
}
\description{
Describe a thing group.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_thing_group(
  thingGroupName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
