% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_thing_registration_task}
\alias{iot_describe_thing_registration_task}
\title{Describes a bulk thing provisioning task}
\usage{
iot_describe_thing_registration_task(taskId)
}
\arguments{
\item{taskId}{[required] The task ID.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  taskId = "string",
  creationDate = as.POSIXct(
    "2015-01-01"
  ),
  lastModifiedDate = as.POSIXct(
    "2015-01-01"
  ),
  templateBody = "string",
  inputFileBucket = "string",
  inputFileKey = "string",
  roleArn = "string",
  status = "InProgress"|"Completed"|"Failed"|"Cancelled"|"Cancelling",
  message = "string",
  successCount = 123,
  failureCount = 123,
  percentageProgress = 123
)
}\if{html}{\out{</div>}}
}
\description{
Describes a bulk thing provisioning task.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_thing_registration_task(
  taskId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
