% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_detach_principal_policy}
\alias{iot_detach_principal_policy}
\title{Removes the specified policy from the specified certificate}
\usage{
iot_detach_principal_policy(policyName, principal)
}
\arguments{
\item{policyName}{[required] The name of the policy to detach.}

\item{principal}{[required] The principal.

If the principal is a certificate, specify the certificate ARN. If the
principal is an Amazon Cognito identity, specify the identity ID.}
}
\description{
Removes the specified policy from the specified certificate.
}
\details{
\strong{Note:} This API is deprecated. Please use DetachPolicy instead.
}
\section{Request syntax}{
\preformatted{svc$detach_principal_policy(
  policyName = "string",
  principal = "string"
)
}
}

\keyword{internal}
