% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_detach_security_profile}
\alias{iot_detach_security_profile}
\title{Disassociates a Device Defender security profile from a thing group or
from this account}
\usage{
iot_detach_security_profile(securityProfileName,
  securityProfileTargetArn)
}
\arguments{
\item{securityProfileName}{[required] The security profile that is detached.}

\item{securityProfileTargetArn}{[required] The ARN of the thing group from which the security profile is detached.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Disassociates a Device Defender security profile from a thing group or
from this account.
}
\section{Request syntax}{
\preformatted{svc$detach_security_profile(
  securityProfileName = "string",
  securityProfileTargetArn = "string"
)
}
}

\keyword{internal}
