% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_detach_thing_principal}
\alias{iot_detach_thing_principal}
\title{Detaches the specified principal from the specified thing}
\usage{
iot_detach_thing_principal(thingName, principal)
}
\arguments{
\item{thingName}{[required] The name of the thing.}

\item{principal}{[required] If the principal is a certificate, this value must be ARN of the
certificate. If the principal is an Amazon Cognito identity, this value
must be the ID of the Amazon Cognito identity.}
}
\description{
Detaches the specified principal from the specified thing. A principal
can be X.509 certificates, IAM users, groups, and roles, Amazon Cognito
identities or federated identities.
}
\details{
This call is asynchronous. It might take several seconds for the
detachment to propagate.
}
\section{Request syntax}{
\preformatted{svc$detach_thing_principal(
  thingName = "string",
  principal = "string"
)
}
}

\keyword{internal}
