% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_get_behavior_model_training_summaries}
\alias{iot_get_behavior_model_training_summaries}
\title{Returns a Device Defender's ML Detect Security Profile training model's
status}
\usage{
iot_get_behavior_model_training_summaries(securityProfileName,
  maxResults, nextToken)
}
\arguments{
\item{securityProfileName}{The name of the security profile.}

\item{maxResults}{The maximum number of results to return at one time. The default is 25.}

\item{nextToken}{The token for the next set of results.}
}
\description{
Returns a Device Defender's ML Detect Security Profile training model's
status.
}
\section{Request syntax}{
\preformatted{svc$get_behavior_model_training_summaries(
  securityProfileName = "string",
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
