% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_get_job_document}
\alias{iot_get_job_document}
\title{Gets a job document}
\usage{
iot_get_job_document(jobId)
}
\arguments{
\item{jobId}{[required] The unique identifier you assigned to this job when it was created.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  document = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Gets a job document.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_job_document(
  jobId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
