% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_get_logging_options}
\alias{iot_get_logging_options}
\title{Gets the logging options}
\usage{
iot_get_logging_options()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  roleArn = "string",
  logLevel = "DEBUG"|"INFO"|"ERROR"|"WARN"|"DISABLED"
)
}\if{html}{\out{</div>}}
}
\description{
Gets the logging options.

NOTE: use of this command is not recommended. Use
\code{\link[=iot_get_v2_logging_options]{get_v2_logging_options}} instead.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_logging_options()
}\if{html}{\out{</div>}}
}

\keyword{internal}
