% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_get_percentiles}
\alias{iot_get_percentiles}
\title{Groups the aggregated values that match the query into percentile
groupings}
\usage{
iot_get_percentiles(indexName, queryString, aggregationField,
  queryVersion, percents)
}
\arguments{
\item{indexName}{The name of the index to search.}

\item{queryString}{[required] The query string.}

\item{aggregationField}{The field to aggregate.}

\item{queryVersion}{The query version.}

\item{percents}{The percentile groups returned.}
}
\description{
Groups the aggregated values that match the query into percentile
groupings. The default percentile groupings are: 1,5,25,50,75,95,99,
although you can specify your own when you call \code{GetPercentiles}. This
function returns a value for each percentile group specified (or the
default percentile groupings). The percentile group \"1\" contains the
aggregated field value that occurs in approximately one percent of the
values that match the query. The percentile group \"5\" contains the
aggregated field value that occurs in approximately five percent of the
values that match the query, and so on. The result is an approximation,
the more values that match the query, the more accurate the percentile
values.
}
\section{Request syntax}{
\preformatted{svc$get_percentiles(
  indexName = "string",
  queryString = "string",
  aggregationField = "string",
  queryVersion = "string",
  percents = list(
    123.0
  )
)
}
}

\keyword{internal}
