% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_get_policy}
\alias{iot_get_policy}
\title{Gets information about the specified policy with the policy document of
the default version}
\usage{
iot_get_policy(policyName)
}
\arguments{
\item{policyName}{[required] The name of the policy.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  policyName = "string",
  policyArn = "string",
  policyDocument = "string",
  defaultVersionId = "string",
  creationDate = as.POSIXct(
    "2015-01-01"
  ),
  lastModifiedDate = as.POSIXct(
    "2015-01-01"
  ),
  generationId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Gets information about the specified policy with the policy document of
the default version.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_policy(
  policyName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
