% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_get_registration_code}
\alias{iot_get_registration_code}
\title{Gets a registration code used to register a CA certificate with AWS IoT}
\usage{
iot_get_registration_code()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  registrationCode = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Gets a registration code used to register a CA certificate with AWS IoT.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_registration_code()
}\if{html}{\out{</div>}}
}

\keyword{internal}
