% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_get_topic_rule_destination}
\alias{iot_get_topic_rule_destination}
\title{Gets information about a topic rule destination}
\usage{
iot_get_topic_rule_destination(arn)
}
\arguments{
\item{arn}{[required] The ARN of the topic rule destination.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  topicRuleDestination = list(
    arn = "string",
    status = "ENABLED"|"IN_PROGRESS"|"DISABLED"|"ERROR"|"DELETING",
    createdAt = as.POSIXct(
      "2015-01-01"
    ),
    lastUpdatedAt = as.POSIXct(
      "2015-01-01"
    ),
    statusReason = "string",
    httpUrlProperties = list(
      confirmationUrl = "string"
    ),
    vpcProperties = list(
      subnetIds = list(
        "string"
      ),
      securityGroups = list(
        "string"
      ),
      vpcId = "string",
      roleArn = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets information about a topic rule destination.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_topic_rule_destination(
  arn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
