% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_get_v2_logging_options}
\alias{iot_get_v2_logging_options}
\title{Gets the fine grained logging options}
\usage{
iot_get_v2_logging_options()
}
\value{
A list with the following syntax:\preformatted{list(
  roleArn = "string",
  defaultLogLevel = "DEBUG"|"INFO"|"ERROR"|"WARN"|"DISABLED",
  disableAllLogs = TRUE|FALSE
)
}
}
\description{
Gets the fine grained logging options.
}
\section{Request syntax}{
\preformatted{svc$get_v2_logging_options()
}
}

\keyword{internal}
