% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_attached_policies}
\alias{iot_list_attached_policies}
\title{Lists the policies attached to the specified thing group}
\usage{
iot_list_attached_policies(target, recursive, marker, pageSize)
}
\arguments{
\item{target}{[required] The group or principal for which the policies will be listed. Valid
principals are CertificateArn
(arn:aws:iot:\emph{region}:\emph{accountId}:cert/\emph{certificateId}), thingGroupArn
(arn:aws:iot:\emph{region}:\emph{accountId}:thinggroup/\emph{groupName}) and CognitoId
(\emph{region}:\emph{id}).}

\item{recursive}{When true, recursively list attached policies.}

\item{marker}{The token to retrieve the next set of results.}

\item{pageSize}{The maximum number of results to be returned per request.}
}
\description{
Lists the policies attached to the specified thing group.
}
\section{Request syntax}{
\preformatted{svc$list_attached_policies(
  target = "string",
  recursive = TRUE|FALSE,
  marker = "string",
  pageSize = 123
)
}
}

\keyword{internal}
