% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_audit_mitigation_actions_executions}
\alias{iot_list_audit_mitigation_actions_executions}
\title{Gets the status of audit mitigation action tasks that were executed}
\usage{
iot_list_audit_mitigation_actions_executions(taskId, actionStatus,
  findingId, maxResults, nextToken)
}
\arguments{
\item{taskId}{[required] Specify this filter to limit results to actions for a specific audit
mitigation actions task.}

\item{actionStatus}{Specify this filter to limit results to those with a specific status.}

\item{findingId}{[required] Specify this filter to limit results to those that were applied to a
specific audit finding.}

\item{maxResults}{The maximum number of results to return at one time. The default is 25.}

\item{nextToken}{The token for the next set of results.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  actionsExecutions = list(
    list(
      taskId = "string",
      findingId = "string",
      actionName = "string",
      actionId = "string",
      status = "IN_PROGRESS"|"COMPLETED"|"FAILED"|"CANCELED"|"SKIPPED"|"PENDING",
      startTime = as.POSIXct(
        "2015-01-01"
      ),
      endTime = as.POSIXct(
        "2015-01-01"
      ),
      errorCode = "string",
      message = "string"
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Gets the status of audit mitigation action tasks that were executed.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_audit_mitigation_actions_executions(
  taskId = "string",
  actionStatus = "IN_PROGRESS"|"COMPLETED"|"FAILED"|"CANCELED"|"SKIPPED"|"PENDING",
  findingId = "string",
  maxResults = 123,
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
