% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_audit_mitigation_actions_tasks}
\alias{iot_list_audit_mitigation_actions_tasks}
\title{Gets a list of audit mitigation action tasks that match the specified
filters}
\usage{
iot_list_audit_mitigation_actions_tasks(auditTaskId, findingId,
  taskStatus, maxResults, nextToken, startTime, endTime)
}
\arguments{
\item{auditTaskId}{Specify this filter to limit results to tasks that were applied to
results for a specific audit.}

\item{findingId}{Specify this filter to limit results to tasks that were applied to a
specific audit finding.}

\item{taskStatus}{Specify this filter to limit results to tasks that are in a specific
state.}

\item{maxResults}{The maximum number of results to return at one time. The default is 25.}

\item{nextToken}{The token for the next set of results.}

\item{startTime}{[required] Specify this filter to limit results to tasks that began on or after a
specific date and time.}

\item{endTime}{[required] Specify this filter to limit results to tasks that were completed or
canceled on or before a specific date and time.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  tasks = list(
    list(
      taskId = "string",
      startTime = as.POSIXct(
        "2015-01-01"
      ),
      taskStatus = "IN_PROGRESS"|"COMPLETED"|"FAILED"|"CANCELED"
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Gets a list of audit mitigation action tasks that match the specified
filters.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_audit_mitigation_actions_tasks(
  auditTaskId = "string",
  findingId = "string",
  taskStatus = "IN_PROGRESS"|"COMPLETED"|"FAILED"|"CANCELED",
  maxResults = 123,
  nextToken = "string",
  startTime = as.POSIXct(
    "2015-01-01"
  ),
  endTime = as.POSIXct(
    "2015-01-01"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
