% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_audit_suppressions}
\alias{iot_list_audit_suppressions}
\title{Lists your Device Defender audit listings}
\usage{
iot_list_audit_suppressions(checkName, resourceIdentifier,
  ascendingOrder, nextToken, maxResults)
}
\arguments{
\item{checkName}{}

\item{resourceIdentifier}{}

\item{ascendingOrder}{Determines whether suppressions are listed in ascending order by
expiration date or not. If parameter isn't provided,
\code{ascendingOrder=true}.}

\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to return at one time. The default is 25.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  suppressions = list(
    list(
      checkName = "string",
      resourceIdentifier = list(
        deviceCertificateId = "string",
        caCertificateId = "string",
        cognitoIdentityPoolId = "string",
        clientId = "string",
        policyVersionIdentifier = list(
          policyName = "string",
          policyVersionId = "string"
        ),
        account = "string",
        iamRoleArn = "string",
        roleAliasArn = "string"
      ),
      expirationDate = as.POSIXct(
        "2015-01-01"
      ),
      suppressIndefinitely = TRUE|FALSE,
      description = "string"
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists your Device Defender audit listings.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_audit_suppressions(
  checkName = "string",
  resourceIdentifier = list(
    deviceCertificateId = "string",
    caCertificateId = "string",
    cognitoIdentityPoolId = "string",
    clientId = "string",
    policyVersionIdentifier = list(
      policyName = "string",
      policyVersionId = "string"
    ),
    account = "string",
    iamRoleArn = "string",
    roleAliasArn = "string"
  ),
  ascendingOrder = TRUE|FALSE,
  nextToken = "string",
  maxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
