% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_authorizers}
\alias{iot_list_authorizers}
\title{Lists the authorizers registered in your account}
\usage{
iot_list_authorizers(pageSize, marker, ascendingOrder, status)
}
\arguments{
\item{pageSize}{The maximum number of results to return at one time.}

\item{marker}{A marker used to get the next set of results.}

\item{ascendingOrder}{Return the list of authorizers in ascending alphabetical order.}

\item{status}{The status of the list authorizers request.}
}
\value{
A list with the following syntax:\preformatted{list(
  authorizers = list(
    list(
      authorizerName = "string",
      authorizerArn = "string"
    )
  ),
  nextMarker = "string"
)
}
}
\description{
Lists the authorizers registered in your account.
}
\section{Request syntax}{
\preformatted{svc$list_authorizers(
  pageSize = 123,
  marker = "string",
  ascendingOrder = TRUE|FALSE,
  status = "ACTIVE"|"INACTIVE"
)
}
}

\keyword{internal}
