% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_ca_certificates}
\alias{iot_list_ca_certificates}
\title{Lists the CA certificates registered for your AWS account}
\usage{
iot_list_ca_certificates(pageSize, marker, ascendingOrder)
}
\arguments{
\item{pageSize}{The result page size.}

\item{marker}{The marker for the next set of results.}

\item{ascendingOrder}{Determines the order of the results.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  certificates = list(
    list(
      certificateArn = "string",
      certificateId = "string",
      status = "ACTIVE"|"INACTIVE",
      creationDate = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  nextMarker = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists the CA certificates registered for your AWS account.

The results are paginated with a default page size of 25. You can use
the returned marker to retrieve additional results.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_ca_certificates(
  pageSize = 123,
  marker = "string",
  ascendingOrder = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
