% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_certificates}
\alias{iot_list_certificates}
\title{Lists the certificates registered in your AWS account}
\usage{
iot_list_certificates(pageSize, marker, ascendingOrder)
}
\arguments{
\item{pageSize}{The result page size.}

\item{marker}{The marker for the next set of results.}

\item{ascendingOrder}{Specifies the order for results. If True, the results are returned in
ascending order, based on the creation date.}
}
\value{
A list with the following syntax:\preformatted{list(
  certificates = list(
    list(
      certificateArn = "string",
      certificateId = "string",
      status = "ACTIVE"|"INACTIVE"|"REVOKED"|"PENDING_TRANSFER"|"REGISTER_INACTIVE"|"PENDING_ACTIVATION",
      certificateMode = "DEFAULT"|"SNI_ONLY",
      creationDate = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  nextMarker = "string"
)
}
}
\description{
Lists the certificates registered in your AWS account.

The results are paginated with a default page size of 25. You can use
the returned marker to retrieve additional results.
}
\section{Request syntax}{
\preformatted{svc$list_certificates(
  pageSize = 123,
  marker = "string",
  ascendingOrder = TRUE|FALSE
)
}
}

\keyword{internal}
