% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_certificates_by_ca}
\alias{iot_list_certificates_by_ca}
\title{List the device certificates signed by the specified CA certificate}
\usage{
iot_list_certificates_by_ca(caCertificateId, pageSize, marker,
  ascendingOrder)
}
\arguments{
\item{caCertificateId}{[required] The ID of the CA certificate. This operation will list all registered
device certificate that were signed by this CA certificate.}

\item{pageSize}{The result page size.}

\item{marker}{The marker for the next set of results.}

\item{ascendingOrder}{Specifies the order for results. If True, the results are returned in
ascending order, based on the creation date.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  certificates = list(
    list(
      certificateArn = "string",
      certificateId = "string",
      status = "ACTIVE"|"INACTIVE"|"REVOKED"|"PENDING_TRANSFER"|"REGISTER_INACTIVE"|"PENDING_ACTIVATION",
      certificateMode = "DEFAULT"|"SNI_ONLY",
      creationDate = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  nextMarker = "string"
)
}\if{html}{\out{</div>}}
}
\description{
List the device certificates signed by the specified CA certificate.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_certificates_by_ca(
  caCertificateId = "string",
  pageSize = 123,
  marker = "string",
  ascendingOrder = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
