% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_custom_metrics}
\alias{iot_list_custom_metrics}
\title{Lists your Device Defender detect custom metrics}
\usage{
iot_list_custom_metrics(nextToken, maxResults)
}
\arguments{
\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to return at one time. The default is 25.}
}
\description{
Lists your Device Defender detect custom metrics.
}
\section{Request syntax}{
\preformatted{svc$list_custom_metrics(
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
