% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_detect_mitigation_actions_executions}
\alias{iot_list_detect_mitigation_actions_executions}
\title{Lists mitigation actions executions for a Device Defender ML Detect
Security Profile}
\usage{
iot_list_detect_mitigation_actions_executions(taskId, violationId,
  thingName, startTime, endTime, maxResults, nextToken)
}
\arguments{
\item{taskId}{The unique identifier of the task.}

\item{violationId}{The unique identifier of the violation.}

\item{thingName}{The name of the thing whose mitigation actions are listed.}

\item{startTime}{A filter to limit results to those found after the specified time. You
must specify either the startTime and endTime or the taskId, but not
both.}

\item{endTime}{The end of the time period for which ML Detect mitigation actions
executions are returned.}

\item{maxResults}{The maximum number of results to return at one time. The default is 25.}

\item{nextToken}{The token for the next set of results.}
}
\value{
A list with the following syntax:\preformatted{list(
  actionsExecutions = list(
    list(
      taskId = "string",
      violationId = "string",
      actionName = "string",
      thingName = "string",
      executionStartDate = as.POSIXct(
        "2015-01-01"
      ),
      executionEndDate = as.POSIXct(
        "2015-01-01"
      ),
      status = "IN_PROGRESS"|"SUCCESSFUL"|"FAILED"|"SKIPPED",
      errorCode = "string",
      message = "string"
    )
  ),
  nextToken = "string"
)
}
}
\description{
Lists mitigation actions executions for a Device Defender ML Detect
Security Profile.
}
\section{Request syntax}{
\preformatted{svc$list_detect_mitigation_actions_executions(
  taskId = "string",
  violationId = "string",
  thingName = "string",
  startTime = as.POSIXct(
    "2015-01-01"
  ),
  endTime = as.POSIXct(
    "2015-01-01"
  ),
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
