% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_detect_mitigation_actions_tasks}
\alias{iot_list_detect_mitigation_actions_tasks}
\title{List of Device Defender ML Detect mitigation actions tasks}
\usage{
iot_list_detect_mitigation_actions_tasks(maxResults, nextToken,
  startTime, endTime)
}
\arguments{
\item{maxResults}{The maximum number of results to return at one time. The default is 25.}

\item{nextToken}{The token for the next set of results.}

\item{startTime}{[required] A filter to limit results to those found after the specified time. You
must specify either the startTime and endTime or the taskId, but not
both.}

\item{endTime}{[required] The end of the time period for which ML Detect mitigation actions tasks
are returned.}
}
\description{
List of Device Defender ML Detect mitigation actions tasks.
}
\section{Request syntax}{
\preformatted{svc$list_detect_mitigation_actions_tasks(
  maxResults = 123,
  nextToken = "string",
  startTime = as.POSIXct(
    "2015-01-01"
  ),
  endTime = as.POSIXct(
    "2015-01-01"
  )
)
}
}

\keyword{internal}
