% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_domain_configurations}
\alias{iot_list_domain_configurations}
\title{Gets a list of domain configurations for the user}
\usage{
iot_list_domain_configurations(marker, pageSize, serviceType)
}
\arguments{
\item{marker}{The marker for the next set of results.}

\item{pageSize}{The result page size.}

\item{serviceType}{The type of service delivered by the endpoint.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  domainConfigurations = list(
    list(
      domainConfigurationName = "string",
      domainConfigurationArn = "string",
      serviceType = "DATA"|"CREDENTIAL_PROVIDER"|"JOBS"
    )
  ),
  nextMarker = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Gets a list of domain configurations for the user. This list is sorted
alphabetically by domain configuration name.

The domain configuration feature is in public preview and is subject to
change.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_domain_configurations(
  marker = "string",
  pageSize = 123,
  serviceType = "DATA"|"CREDENTIAL_PROVIDER"|"JOBS"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
