% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_job_executions_for_job}
\alias{iot_list_job_executions_for_job}
\title{Lists the job executions for a job}
\usage{
iot_list_job_executions_for_job(jobId, status, maxResults, nextToken)
}
\arguments{
\item{jobId}{[required] The unique identifier you assigned to this job when it was created.}

\item{status}{The status of the job.}

\item{maxResults}{The maximum number of results to be returned per request.}

\item{nextToken}{The token to retrieve the next set of results.}
}
\description{
Lists the job executions for a job.
}
\section{Request syntax}{
\preformatted{svc$list_job_executions_for_job(
  jobId = "string",
  status = "QUEUED"|"IN_PROGRESS"|"SUCCEEDED"|"FAILED"|"TIMED_OUT"|"REJECTED"|"REMOVED"|"CANCELED",
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
