% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_job_executions_for_thing}
\alias{iot_list_job_executions_for_thing}
\title{Lists the job executions for the specified thing}
\usage{
iot_list_job_executions_for_thing(thingName, status, namespaceId,
  maxResults, nextToken)
}
\arguments{
\item{thingName}{[required] The thing name.}

\item{status}{An optional filter that lets you search for jobs that have the specified
status.}

\item{namespaceId}{The namespace used to indicate that a job is a customer-managed job.

When you specify a value for this parameter, AWS IoT Core sends jobs
notifications to MQTT topics that contain the value in the following
format.

\verb{$aws/things/THING_NAME/jobs/JOB_ID/notify-namespace-NAMESPACE_ID/}

The \code{namespaceId} feature is in public preview.}

\item{maxResults}{The maximum number of results to be returned per request.}

\item{nextToken}{The token to retrieve the next set of results.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  executionSummaries = list(
    list(
      jobId = "string",
      jobExecutionSummary = list(
        status = "QUEUED"|"IN_PROGRESS"|"SUCCEEDED"|"FAILED"|"TIMED_OUT"|"REJECTED"|"REMOVED"|"CANCELED",
        queuedAt = as.POSIXct(
          "2015-01-01"
        ),
        startedAt = as.POSIXct(
          "2015-01-01"
        ),
        lastUpdatedAt = as.POSIXct(
          "2015-01-01"
        ),
        executionNumber = 123
      )
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists the job executions for the specified thing.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_job_executions_for_thing(
  thingName = "string",
  status = "QUEUED"|"IN_PROGRESS"|"SUCCEEDED"|"FAILED"|"TIMED_OUT"|"REJECTED"|"REMOVED"|"CANCELED",
  namespaceId = "string",
  maxResults = 123,
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
