% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_jobs}
\alias{iot_list_jobs}
\title{Lists jobs}
\usage{
iot_list_jobs(status, targetSelection, maxResults, nextToken,
  thingGroupName, thingGroupId, namespaceId)
}
\arguments{
\item{status}{An optional filter that lets you search for jobs that have the specified
status.}

\item{targetSelection}{Specifies whether the job will continue to run (CONTINUOUS), or will be
complete after all those things specified as targets have completed the
job (SNAPSHOT). If continuous, the job may also be run on a thing when a
change is detected in a target. For example, a job will run on a thing
when the thing is added to a target group, even after the job was
completed by all things originally in the group.}

\item{maxResults}{The maximum number of results to return per request.}

\item{nextToken}{The token to retrieve the next set of results.}

\item{thingGroupName}{A filter that limits the returned jobs to those for the specified group.}

\item{thingGroupId}{A filter that limits the returned jobs to those for the specified group.}

\item{namespaceId}{The namespace used to indicate that a job is a customer-managed job.

When you specify a value for this parameter, AWS IoT Core sends jobs
notifications to MQTT topics that contain the value in the following
format.

\verb{$aws/things/THING_NAME/jobs/JOB_ID/notify-namespace-NAMESPACE_ID/}

The \code{namespaceId} feature is in public preview.}
}
\value{
A list with the following syntax:\preformatted{list(
  jobs = list(
    list(
      jobArn = "string",
      jobId = "string",
      thingGroupId = "string",
      targetSelection = "CONTINUOUS"|"SNAPSHOT",
      status = "IN_PROGRESS"|"CANCELED"|"COMPLETED"|"DELETION_IN_PROGRESS",
      createdAt = as.POSIXct(
        "2015-01-01"
      ),
      lastUpdatedAt = as.POSIXct(
        "2015-01-01"
      ),
      completedAt = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  nextToken = "string"
)
}
}
\description{
Lists jobs.
}
\section{Request syntax}{
\preformatted{svc$list_jobs(
  status = "IN_PROGRESS"|"CANCELED"|"COMPLETED"|"DELETION_IN_PROGRESS",
  targetSelection = "CONTINUOUS"|"SNAPSHOT",
  maxResults = 123,
  nextToken = "string",
  thingGroupName = "string",
  thingGroupId = "string",
  namespaceId = "string"
)
}
}

\keyword{internal}
