% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_mitigation_actions}
\alias{iot_list_mitigation_actions}
\title{Gets a list of all mitigation actions that match the specified filter
criteria}
\usage{
iot_list_mitigation_actions(actionType, maxResults, nextToken)
}
\arguments{
\item{actionType}{Specify a value to limit the result to mitigation actions with a
specific action type.}

\item{maxResults}{The maximum number of results to return at one time. The default is 25.}

\item{nextToken}{The token for the next set of results.}
}
\value{
A list with the following syntax:\preformatted{list(
  actionIdentifiers = list(
    list(
      actionName = "string",
      actionArn = "string",
      creationDate = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  nextToken = "string"
)
}
}
\description{
Gets a list of all mitigation actions that match the specified filter
criteria.
}
\section{Request syntax}{
\preformatted{svc$list_mitigation_actions(
  actionType = "UPDATE_DEVICE_CERTIFICATE"|"UPDATE_CA_CERTIFICATE"|"ADD_THINGS_TO_THING_GROUP"|"REPLACE_DEFAULT_POLICY_VERSION"|"ENABLE_IOT_LOGGING"|"PUBLISH_FINDING_TO_SNS",
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
