% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_ota_updates}
\alias{iot_list_ota_updates}
\title{Lists OTA updates}
\usage{
iot_list_ota_updates(maxResults, nextToken, otaUpdateStatus)
}
\arguments{
\item{maxResults}{The maximum number of results to return at one time.}

\item{nextToken}{A token used to retrieve the next set of results.}

\item{otaUpdateStatus}{The OTA update job status.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  otaUpdates = list(
    list(
      otaUpdateId = "string",
      otaUpdateArn = "string",
      creationDate = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists OTA updates.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_ota_updates(
  maxResults = 123,
  nextToken = "string",
  otaUpdateStatus = "CREATE_PENDING"|"CREATE_IN_PROGRESS"|"CREATE_COMPLETE"|"CREATE_FAILED"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
