% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_principal_policies}
\alias{iot_list_principal_policies}
\title{Lists the policies attached to the specified principal}
\usage{
iot_list_principal_policies(principal, marker, pageSize, ascendingOrder)
}
\arguments{
\item{principal}{[required] The principal. Valid principals are CertificateArn
(arn:aws:iot:\emph{region}:\emph{accountId}:cert/\emph{certificateId}), thingGroupArn
(arn:aws:iot:\emph{region}:\emph{accountId}:thinggroup/\emph{groupName}) and CognitoId
(\emph{region}:\emph{id}).}

\item{marker}{The marker for the next set of results.}

\item{pageSize}{The result page size.}

\item{ascendingOrder}{Specifies the order for results. If true, results are returned in
ascending creation order.}
}
\description{
Lists the policies attached to the specified principal. If you use an
Cognito identity, the ID must be in \href{https://docs.aws.amazon.com/cognitoidentity/latest/APIReference/API_GetCredentialsForIdentity.html#API_GetCredentialsForIdentity_RequestSyntax}{AmazonCognito Identity format}.

\strong{Note:} This API is deprecated. Please use
\code{\link[=iot_list_attached_policies]{list_attached_policies}} instead.
}
\section{Request syntax}{
\preformatted{svc$list_principal_policies(
  principal = "string",
  marker = "string",
  pageSize = 123,
  ascendingOrder = TRUE|FALSE
)
}
}

\keyword{internal}
