% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_provisioning_template_versions}
\alias{iot_list_provisioning_template_versions}
\title{A list of fleet provisioning template versions}
\usage{
iot_list_provisioning_template_versions(templateName, maxResults,
  nextToken)
}
\arguments{
\item{templateName}{[required] The name of the fleet provisioning template.}

\item{maxResults}{The maximum number of results to return at one time.}

\item{nextToken}{A token to retrieve the next set of results.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  versions = list(
    list(
      versionId = 123,
      creationDate = as.POSIXct(
        "2015-01-01"
      ),
      isDefaultVersion = TRUE|FALSE
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
A list of fleet provisioning template versions.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_provisioning_template_versions(
  templateName = "string",
  maxResults = 123,
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
