% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_provisioning_templates}
\alias{iot_list_provisioning_templates}
\title{Lists the fleet provisioning templates in your AWS account}
\usage{
iot_list_provisioning_templates(maxResults, nextToken)
}
\arguments{
\item{maxResults}{The maximum number of results to return at one time.}

\item{nextToken}{A token to retrieve the next set of results.}
}
\description{
Lists the fleet provisioning templates in your AWS account.
}
\section{Request syntax}{
\preformatted{svc$list_provisioning_templates(
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
