% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_scheduled_audits}
\alias{iot_list_scheduled_audits}
\title{Lists all of your scheduled audits}
\usage{
iot_list_scheduled_audits(nextToken, maxResults)
}
\arguments{
\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to return at one time. The default is 25.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  scheduledAudits = list(
    list(
      scheduledAuditName = "string",
      scheduledAuditArn = "string",
      frequency = "DAILY"|"WEEKLY"|"BIWEEKLY"|"MONTHLY",
      dayOfMonth = "string",
      dayOfWeek = "SUN"|"MON"|"TUE"|"WED"|"THU"|"FRI"|"SAT"
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists all of your scheduled audits.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_scheduled_audits(
  nextToken = "string",
  maxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
