% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_security_profiles}
\alias{iot_list_security_profiles}
\title{Lists the Device Defender security profiles you've created}
\usage{
iot_list_security_profiles(nextToken, maxResults, dimensionName,
  metricName)
}
\arguments{
\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to return at one time.}

\item{dimensionName}{A filter to limit results to the security profiles that use the defined
dimension. Cannot be used with \code{metricName}}

\item{metricName}{The name of the custom metric. Cannot be used with \code{dimensionName}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  securityProfileIdentifiers = list(
    list(
      name = "string",
      arn = "string"
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists the Device Defender security profiles you've created. You can
filter security profiles by dimension or custom metric.

\code{dimensionName} and \code{metricName} cannot be used in the same request.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_security_profiles(
  nextToken = "string",
  maxResults = 123,
  dimensionName = "string",
  metricName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
