% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_security_profiles_for_target}
\alias{iot_list_security_profiles_for_target}
\title{Lists the Device Defender security profiles attached to a target (thing
group)}
\usage{
iot_list_security_profiles_for_target(nextToken, maxResults, recursive,
  securityProfileTargetArn)
}
\arguments{
\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to return at one time.}

\item{recursive}{If true, return child groups as well.}

\item{securityProfileTargetArn}{[required] The ARN of the target (thing group) whose attached security profiles you
want to get.}
}
\description{
Lists the Device Defender security profiles attached to a target (thing
group).
}
\section{Request syntax}{
\preformatted{svc$list_security_profiles_for_target(
  nextToken = "string",
  maxResults = 123,
  recursive = TRUE|FALSE,
  securityProfileTargetArn = "string"
)
}
}

\keyword{internal}
