% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_streams}
\alias{iot_list_streams}
\title{Lists all of the streams in your AWS account}
\usage{
iot_list_streams(maxResults, nextToken, ascendingOrder)
}
\arguments{
\item{maxResults}{The maximum number of results to return at a time.}

\item{nextToken}{A token used to get the next set of results.}

\item{ascendingOrder}{Set to true to return the list of streams in ascending order.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  streams = list(
    list(
      streamId = "string",
      streamArn = "string",
      streamVersion = 123,
      description = "string"
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists all of the streams in your AWS account.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_streams(
  maxResults = 123,
  nextToken = "string",
  ascendingOrder = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
