% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_targets_for_security_profile}
\alias{iot_list_targets_for_security_profile}
\title{Lists the targets (thing groups) associated with a given Device Defender
security profile}
\usage{
iot_list_targets_for_security_profile(securityProfileName, nextToken,
  maxResults)
}
\arguments{
\item{securityProfileName}{[required] The security profile.}

\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to return at one time.}
}
\value{
A list with the following syntax:\preformatted{list(
  securityProfileTargets = list(
    list(
      arn = "string"
    )
  ),
  nextToken = "string"
)
}
}
\description{
Lists the targets (thing groups) associated with a given Device Defender
security profile.
}
\section{Request syntax}{
\preformatted{svc$list_targets_for_security_profile(
  securityProfileName = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
