% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_thing_groups}
\alias{iot_list_thing_groups}
\title{List the thing groups in your account}
\usage{
iot_list_thing_groups(nextToken, maxResults, parentGroup,
  namePrefixFilter, recursive)
}
\arguments{
\item{nextToken}{To retrieve the next set of results, the \code{nextToken} value from a
previous response; otherwise \strong{null} to receive the first set of
results.}

\item{maxResults}{The maximum number of results to return at one time.}

\item{parentGroup}{A filter that limits the results to those with the specified parent
group.}

\item{namePrefixFilter}{A filter that limits the results to those with the specified name
prefix.}

\item{recursive}{If true, return child groups as well.}
}
\description{
List the thing groups in your account.
}
\section{Request syntax}{
\preformatted{svc$list_thing_groups(
  nextToken = "string",
  maxResults = 123,
  parentGroup = "string",
  namePrefixFilter = "string",
  recursive = TRUE|FALSE
)
}
}

\keyword{internal}
