% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_thing_groups_for_thing}
\alias{iot_list_thing_groups_for_thing}
\title{List the thing groups to which the specified thing belongs}
\usage{
iot_list_thing_groups_for_thing(thingName, nextToken, maxResults)
}
\arguments{
\item{thingName}{[required] The thing name.}

\item{nextToken}{To retrieve the next set of results, the \code{nextToken} value from a
previous response; otherwise \strong{null} to receive the first set of
results.}

\item{maxResults}{The maximum number of results to return at one time.}
}
\value{
A list with the following syntax:\preformatted{list(
  thingGroups = list(
    list(
      groupName = "string",
      groupArn = "string"
    )
  ),
  nextToken = "string"
)
}
}
\description{
List the thing groups to which the specified thing belongs.
}
\section{Request syntax}{
\preformatted{svc$list_thing_groups_for_thing(
  thingName = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
