% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_thing_principals}
\alias{iot_list_thing_principals}
\title{Lists the principals associated with the specified thing}
\usage{
iot_list_thing_principals(nextToken, maxResults, thingName)
}
\arguments{
\item{nextToken}{To retrieve the next set of results, the \code{nextToken} value from a
previous response; otherwise \strong{null} to receive the first set of
results.}

\item{maxResults}{The maximum number of results to return in this operation.}

\item{thingName}{[required] The name of the thing.}
}
\value{
A list with the following syntax:\preformatted{list(
  principals = list(
    "string"
  ),
  nextToken = "string"
)
}
}
\description{
Lists the principals associated with the specified thing. A principal
can be X.509 certificates, IAM users, groups, and roles, Amazon Cognito
identities or federated identities.
}
\section{Request syntax}{
\preformatted{svc$list_thing_principals(
  nextToken = "string",
  maxResults = 123,
  thingName = "string"
)
}
}

\keyword{internal}
