% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_thing_registration_tasks}
\alias{iot_list_thing_registration_tasks}
\title{List bulk thing provisioning tasks}
\usage{
iot_list_thing_registration_tasks(nextToken, maxResults, status)
}
\arguments{
\item{nextToken}{To retrieve the next set of results, the \code{nextToken} value from a
previous response; otherwise \strong{null} to receive the first set of
results.}

\item{maxResults}{The maximum number of results to return at one time.}

\item{status}{The status of the bulk thing provisioning task.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  taskIds = list(
    "string"
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
List bulk thing provisioning tasks.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_thing_registration_tasks(
  nextToken = "string",
  maxResults = 123,
  status = "InProgress"|"Completed"|"Failed"|"Cancelled"|"Cancelling"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
