% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_things}
\alias{iot_list_things}
\title{Lists your things}
\usage{
iot_list_things(nextToken, maxResults, attributeName, attributeValue,
  thingTypeName, usePrefixAttributeValue)
}
\arguments{
\item{nextToken}{To retrieve the next set of results, the \code{nextToken} value from a
previous response; otherwise \strong{null} to receive the first set of
results.}

\item{maxResults}{The maximum number of results to return in this operation.}

\item{attributeName}{The attribute name used to search for things.}

\item{attributeValue}{The attribute value used to search for things.}

\item{thingTypeName}{The name of the thing type used to search for things.}

\item{usePrefixAttributeValue}{When \code{true}, the action returns the thing resources with attribute
values that start with the \code{attributeValue} provided.

When \code{false}, or not present, the action returns only the thing
resources with attribute values that match the entire \code{attributeValue}
provided.}
}
\value{
A list with the following syntax:\preformatted{list(
  things = list(
    list(
      thingName = "string",
      thingTypeName = "string",
      thingArn = "string",
      attributes = list(
        "string"
      ),
      version = 123
    )
  ),
  nextToken = "string"
)
}
}
\description{
Lists your things. Use the \strong{attributeName} and \strong{attributeValue}
parameters to filter your things. For example, calling
\code{\link[=iot_list_things]{list_things}} with attributeName=Color and
attributeValue=Red retrieves all things in the registry that contain an
attribute \strong{Color} with the value \strong{Red}.

You will not be charged for calling this API if an \verb{Access denied} error
is returned. You will also not be charged if no attributes or pagination
token was provided in request and no pagination token and no results
were returned.
}
\section{Request syntax}{
\preformatted{svc$list_things(
  nextToken = "string",
  maxResults = 123,
  attributeName = "string",
  attributeValue = "string",
  thingTypeName = "string",
  usePrefixAttributeValue = TRUE|FALSE
)
}
}

\keyword{internal}
