% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_things_in_billing_group}
\alias{iot_list_things_in_billing_group}
\title{Lists the things you have added to the given billing group}
\usage{
iot_list_things_in_billing_group(billingGroupName, nextToken,
  maxResults)
}
\arguments{
\item{billingGroupName}{[required] The name of the billing group.}

\item{nextToken}{To retrieve the next set of results, the \code{nextToken} value from a
previous response; otherwise \strong{null} to receive the first set of
results.}

\item{maxResults}{The maximum number of results to return per request.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  things = list(
    "string"
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists the things you have added to the given billing group.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_things_in_billing_group(
  billingGroupName = "string",
  nextToken = "string",
  maxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
