% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_things_in_thing_group}
\alias{iot_list_things_in_thing_group}
\title{Lists the things in the specified group}
\usage{
iot_list_things_in_thing_group(thingGroupName, recursive, nextToken,
  maxResults)
}
\arguments{
\item{thingGroupName}{[required] The thing group name.}

\item{recursive}{When true, list things in this thing group and in all child groups as
well.}

\item{nextToken}{To retrieve the next set of results, the \code{nextToken} value from a
previous response; otherwise \strong{null} to receive the first set of
results.}

\item{maxResults}{The maximum number of results to return at one time.}
}
\value{
A list with the following syntax:\preformatted{list(
  things = list(
    "string"
  ),
  nextToken = "string"
)
}
}
\description{
Lists the things in the specified group.
}
\section{Request syntax}{
\preformatted{svc$list_things_in_thing_group(
  thingGroupName = "string",
  recursive = TRUE|FALSE,
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
