% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_v2_logging_levels}
\alias{iot_list_v2_logging_levels}
\title{Lists logging levels}
\usage{
iot_list_v2_logging_levels(targetType, nextToken, maxResults)
}
\arguments{
\item{targetType}{The type of resource for which you are configuring logging. Must be
\code{THING_Group}.}

\item{nextToken}{To retrieve the next set of results, the \code{nextToken} value from a
previous response; otherwise \strong{null} to receive the first set of
results.}

\item{maxResults}{The maximum number of results to return at one time.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  logTargetConfigurations = list(
    list(
      logTarget = list(
        targetType = "DEFAULT"|"THING_GROUP",
        targetName = "string"
      ),
      logLevel = "DEBUG"|"INFO"|"ERROR"|"WARN"|"DISABLED"
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists logging levels.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_v2_logging_levels(
  targetType = "DEFAULT"|"THING_GROUP",
  nextToken = "string",
  maxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
