% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_violation_events}
\alias{iot_list_violation_events}
\title{Lists the Device Defender security profile violations discovered during
the given time period}
\usage{
iot_list_violation_events(startTime, endTime, thingName,
  securityProfileName, behaviorCriteriaType, listSuppressedAlerts,
  nextToken, maxResults)
}
\arguments{
\item{startTime}{[required] The start time for the alerts to be listed.}

\item{endTime}{[required] The end time for the alerts to be listed.}

\item{thingName}{A filter to limit results to those alerts caused by the specified thing.}

\item{securityProfileName}{A filter to limit results to those alerts generated by the specified
security profile.}

\item{behaviorCriteriaType}{The criteria for a behavior.}

\item{listSuppressedAlerts}{A list of all suppressed alerts.}

\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to return at one time.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  violationEvents = list(
    list(
      violationId = "string",
      thingName = "string",
      securityProfileName = "string",
      behavior = list(
        name = "string",
        metric = "string",
        metricDimension = list(
          dimensionName = "string",
          operator = "IN"|"NOT_IN"
        ),
        criteria = list(
          comparisonOperator = "less-than"|"less-than-equals"|"greater-than"|"greater-than-equals"|"in-cidr-set"|"not-in-cidr-set"|"in-port-set"|"not-in-port-set"|"in-set"|"not-in-set",
          value = list(
            count = 123,
            cidrs = list(
              "string"
            ),
            ports = list(
              123
            ),
            number = 123.0,
            numbers = list(
              123.0
            ),
            strings = list(
              "string"
            )
          ),
          durationSeconds = 123,
          consecutiveDatapointsToAlarm = 123,
          consecutiveDatapointsToClear = 123,
          statisticalThreshold = list(
            statistic = "string"
          ),
          mlDetectionConfig = list(
            confidenceLevel = "LOW"|"MEDIUM"|"HIGH"
          )
        ),
        suppressAlerts = TRUE|FALSE
      ),
      metricValue = list(
        count = 123,
        cidrs = list(
          "string"
        ),
        ports = list(
          123
        ),
        number = 123.0,
        numbers = list(
          123.0
        ),
        strings = list(
          "string"
        )
      ),
      violationEventAdditionalInfo = list(
        confidenceLevel = "LOW"|"MEDIUM"|"HIGH"
      ),
      violationEventType = "in-alarm"|"alarm-cleared"|"alarm-invalidated",
      violationEventTime = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists the Device Defender security profile violations discovered during
the given time period. You can use filters to limit the results to those
alerts issued for a particular security profile, behavior, or thing
(device).
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_violation_events(
  startTime = as.POSIXct(
    "2015-01-01"
  ),
  endTime = as.POSIXct(
    "2015-01-01"
  ),
  thingName = "string",
  securityProfileName = "string",
  behaviorCriteriaType = "STATIC"|"STATISTICAL"|"MACHINE_LEARNING",
  listSuppressedAlerts = TRUE|FALSE,
  nextToken = "string",
  maxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
