% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_register_certificate}
\alias{iot_register_certificate}
\title{Registers a device certificate with AWS IoT}
\usage{
iot_register_certificate(certificatePem, caCertificatePem, setAsActive,
  status)
}
\arguments{
\item{certificatePem}{[required] The certificate data, in PEM format.}

\item{caCertificatePem}{The CA certificate used to sign the device certificate being registered.}

\item{setAsActive}{A boolean value that specifies if the certificate is set to active.}

\item{status}{The status of the register certificate request.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  certificateArn = "string",
  certificateId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Registers a device certificate with AWS IoT. If you have more than one
CA certificate that has the same subject field, you must specify the CA
certificate that was used to sign the device certificate being
registered.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$register_certificate(
  certificatePem = "string",
  caCertificatePem = "string",
  setAsActive = TRUE|FALSE,
  status = "ACTIVE"|"INACTIVE"|"REVOKED"|"PENDING_TRANSFER"|"REGISTER_INACTIVE"|"PENDING_ACTIVATION"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
