% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_register_certificate_without_ca}
\alias{iot_register_certificate_without_ca}
\title{Register a certificate that does not have a certificate authority (CA)}
\usage{
iot_register_certificate_without_ca(certificatePem, status)
}
\arguments{
\item{certificatePem}{[required] The certificate data, in PEM format.}

\item{status}{The status of the register certificate request.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  certificateArn = "string",
  certificateId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Register a certificate that does not have a certificate authority (CA).
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$register_certificate_without_ca(
  certificatePem = "string",
  status = "ACTIVE"|"INACTIVE"|"REVOKED"|"PENDING_TRANSFER"|"REGISTER_INACTIVE"|"PENDING_ACTIVATION"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
