% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_register_thing}
\alias{iot_register_thing}
\title{Provisions a thing in the device registry}
\usage{
iot_register_thing(templateBody, parameters)
}
\arguments{
\item{templateBody}{[required] The provisioning template. See \href{https://docs.aws.amazon.com/iot/latest/developerguide/provision-w-cert.html}{Provisioning Devices That Have Device Certificates}
for more information.}

\item{parameters}{The parameters for provisioning a thing. See \href{https://docs.aws.amazon.com/iot/latest/developerguide/provision-template.html}{Provisioning Templates}
for more information.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  certificatePem = "string",
  resourceArns = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Provisions a thing in the device registry. RegisterThing calls other AWS
IoT control plane APIs. These calls might exceed your account level \href{https://docs.aws.amazon.com/general/latest/gr/aws_service_limits.html#limits_iot}{AWS IoT Throttling Limits}
and cause throttle errors. Please contact AWS Customer Support to raise
your throttling limits if necessary.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$register_thing(
  templateBody = "string",
  parameters = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
