% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_set_default_policy_version}
\alias{iot_set_default_policy_version}
\title{Sets the specified version of the specified policy as the policy's
default (operative) version}
\usage{
iot_set_default_policy_version(policyName, policyVersionId)
}
\arguments{
\item{policyName}{[required] The policy name.}

\item{policyVersionId}{[required] The policy version ID.}
}
\value{
An empty list.
}
\description{
Sets the specified version of the specified policy as the policy's
default (operative) version. This action affects all certificates to
which the policy is attached. To list the principals the policy is
attached to, use the ListPrincipalPolicy API.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$set_default_policy_version(
  policyName = "string",
  policyVersionId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
