% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_set_logging_options}
\alias{iot_set_logging_options}
\title{Sets the logging options}
\usage{
iot_set_logging_options(loggingOptionsPayload)
}
\arguments{
\item{loggingOptionsPayload}{[required] The logging options payload.}
}
\value{
An empty list.
}
\description{
Sets the logging options.

NOTE: use of this command is not recommended. Use
\code{\link[=iot_set_v2_logging_options]{set_v2_logging_options}} instead.
}
\section{Request syntax}{
\preformatted{svc$set_logging_options(
  loggingOptionsPayload = list(
    roleArn = "string",
    logLevel = "DEBUG"|"INFO"|"ERROR"|"WARN"|"DISABLED"
  )
)
}
}

\keyword{internal}
