% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_set_v2_logging_level}
\alias{iot_set_v2_logging_level}
\title{Sets the logging level}
\usage{
iot_set_v2_logging_level(logTarget, logLevel)
}
\arguments{
\item{logTarget}{[required] The log target.}

\item{logLevel}{[required] The log level.}
}
\value{
An empty list.
}
\description{
Sets the logging level.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$set_v2_logging_level(
  logTarget = list(
    targetType = "DEFAULT"|"THING_GROUP",
    targetName = "string"
  ),
  logLevel = "DEBUG"|"INFO"|"ERROR"|"WARN"|"DISABLED"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
