% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_set_v2_logging_options}
\alias{iot_set_v2_logging_options}
\title{Sets the logging options for the V2 logging service}
\usage{
iot_set_v2_logging_options(roleArn, defaultLogLevel, disableAllLogs)
}
\arguments{
\item{roleArn}{The ARN of the role that allows IoT to write to Cloudwatch logs.}

\item{defaultLogLevel}{The default logging level.}

\item{disableAllLogs}{If true all logs are disabled. The default is false.}
}
\value{
An empty list.
}
\description{
Sets the logging options for the V2 logging service.
}
\section{Request syntax}{
\preformatted{svc$set_v2_logging_options(
  roleArn = "string",
  defaultLogLevel = "DEBUG"|"INFO"|"ERROR"|"WARN"|"DISABLED",
  disableAllLogs = TRUE|FALSE
)
}
}

\keyword{internal}
