% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_start_audit_mitigation_actions_task}
\alias{iot_start_audit_mitigation_actions_task}
\title{Starts a task that applies a set of mitigation actions to the specified
target}
\usage{
iot_start_audit_mitigation_actions_task(taskId, target,
  auditCheckToActionsMapping, clientRequestToken)
}
\arguments{
\item{taskId}{[required] A unique identifier for the task. You can use this identifier to check
the status of the task or to cancel it.}

\item{target}{[required] Specifies the audit findings to which the mitigation actions are
applied. You can apply them to a type of audit check, to all findings
from an audit, or to a specific set of findings.}

\item{auditCheckToActionsMapping}{[required] For an audit check, specifies which mitigation actions to apply. Those
actions must be defined in your AWS account.}

\item{clientRequestToken}{[required] Each audit mitigation task must have a unique client request token. If
you try to start a new task with the same token as a task that already
exists, an exception occurs. If you omit this value, a unique client
request token is generated automatically.}
}
\value{
A list with the following syntax:\preformatted{list(
  taskId = "string"
)
}
}
\description{
Starts a task that applies a set of mitigation actions to the specified
target.
}
\section{Request syntax}{
\preformatted{svc$start_audit_mitigation_actions_task(
  taskId = "string",
  target = list(
    auditTaskId = "string",
    findingIds = list(
      "string"
    ),
    auditCheckToReasonCodeFilter = list(
      list(
        "string"
      )
    )
  ),
  auditCheckToActionsMapping = list(
    list(
      "string"
    )
  ),
  clientRequestToken = "string"
)
}
}

\keyword{internal}
