% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_start_detect_mitigation_actions_task}
\alias{iot_start_detect_mitigation_actions_task}
\title{Starts a Device Defender ML Detect mitigation actions task}
\usage{
iot_start_detect_mitigation_actions_task(taskId, target, actions,
  violationEventOccurrenceRange, includeOnlyActiveViolations,
  includeSuppressedAlerts, clientRequestToken)
}
\arguments{
\item{taskId}{[required] The unique identifier of the task.}

\item{target}{[required] Specifies the ML Detect findings to which the mitigation actions are
applied.}

\item{actions}{[required] The actions to be performed when a device has unexpected behavior.}

\item{violationEventOccurrenceRange}{Specifies the time period of which violation events occurred between.}

\item{includeOnlyActiveViolations}{Specifies to list only active violations.}

\item{includeSuppressedAlerts}{Specifies to include suppressed alerts.}

\item{clientRequestToken}{[required] Each mitigation action task must have a unique client request token. If
you try to create a new task with the same token as a task that already
exists, an exception occurs. If you omit this value, AWS SDKs will
automatically generate a unique client request.}
}
\value{
A list with the following syntax:\preformatted{list(
  taskId = "string"
)
}
}
\description{
Starts a Device Defender ML Detect mitigation actions task.
}
\section{Request syntax}{
\preformatted{svc$start_detect_mitigation_actions_task(
  taskId = "string",
  target = list(
    violationIds = list(
      "string"
    ),
    securityProfileName = "string",
    behaviorName = "string"
  ),
  actions = list(
    "string"
  ),
  violationEventOccurrenceRange = list(
    startTime = as.POSIXct(
      "2015-01-01"
    ),
    endTime = as.POSIXct(
      "2015-01-01"
    )
  ),
  includeOnlyActiveViolations = TRUE|FALSE,
  includeSuppressedAlerts = TRUE|FALSE,
  clientRequestToken = "string"
)
}
}

\keyword{internal}
