% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_start_on_demand_audit_task}
\alias{iot_start_on_demand_audit_task}
\title{Starts an on-demand Device Defender audit}
\usage{
iot_start_on_demand_audit_task(targetCheckNames)
}
\arguments{
\item{targetCheckNames}{[required] Which checks are performed during the audit. The checks you specify must
be enabled for your account or an exception occurs. Use
\code{\link[=iot_describe_account_audit_configuration]{describe_account_audit_configuration}}
to see the list of all checks, including those that are enabled or
\code{\link[=iot_update_account_audit_configuration]{update_account_audit_configuration}}
to select which checks are enabled.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  taskId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Starts an on-demand Device Defender audit.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$start_on_demand_audit_task(
  targetCheckNames = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
