% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_tag_resource}
\alias{iot_tag_resource}
\title{Adds to or modifies the tags of the given resource}
\usage{
iot_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource.}

\item{tags}{[required] The new or modified tags for the resource.}
}
\value{
An empty list.
}
\description{
Adds to or modifies the tags of the given resource. Tags are metadata
which can be used to manage a resource.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  resourceArn = "string",
  tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
