% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_test_authorization}
\alias{iot_test_authorization}
\title{Tests if a specified principal is authorized to perform an AWS IoT
action on a specified resource}
\usage{
iot_test_authorization(principal, cognitoIdentityPoolId, authInfos,
  clientId, policyNamesToAdd, policyNamesToSkip)
}
\arguments{
\item{principal}{The principal. Valid principals are CertificateArn
(arn:aws:iot:\emph{region}:\emph{accountId}:cert/\emph{certificateId}), thingGroupArn
(arn:aws:iot:\emph{region}:\emph{accountId}:thinggroup/\emph{groupName}) and CognitoId
(\emph{region}:\emph{id}).}

\item{cognitoIdentityPoolId}{The Cognito identity pool ID.}

\item{authInfos}{[required] A list of authorization info objects. Simulating authorization will
create a response for each \code{authInfo} object in the list.}

\item{clientId}{The MQTT client ID.}

\item{policyNamesToAdd}{When testing custom authorization, the policies specified here are
treated as if they are attached to the principal being authorized.}

\item{policyNamesToSkip}{When testing custom authorization, the policies specified here are
treated as if they are not attached to the principal being authorized.}
}
\value{
A list with the following syntax:\preformatted{list(
  authResults = list(
    list(
      authInfo = list(
        actionType = "PUBLISH"|"SUBSCRIBE"|"RECEIVE"|"CONNECT",
        resources = list(
          "string"
        )
      ),
      allowed = list(
        policies = list(
          list(
            policyName = "string",
            policyArn = "string"
          )
        )
      ),
      denied = list(
        implicitDeny = list(
          policies = list(
            list(
              policyName = "string",
              policyArn = "string"
            )
          )
        ),
        explicitDeny = list(
          policies = list(
            list(
              policyName = "string",
              policyArn = "string"
            )
          )
        )
      ),
      authDecision = "ALLOWED"|"EXPLICIT_DENY"|"IMPLICIT_DENY",
      missingContextValues = list(
        "string"
      )
    )
  )
)
}
}
\description{
Tests if a specified principal is authorized to perform an AWS IoT
action on a specified resource. Use this to test and debug the
authorization behavior of devices that connect to the AWS IoT device
gateway.
}
\section{Request syntax}{
\preformatted{svc$test_authorization(
  principal = "string",
  cognitoIdentityPoolId = "string",
  authInfos = list(
    list(
      actionType = "PUBLISH"|"SUBSCRIBE"|"RECEIVE"|"CONNECT",
      resources = list(
        "string"
      )
    )
  ),
  clientId = "string",
  policyNamesToAdd = list(
    "string"
  ),
  policyNamesToSkip = list(
    "string"
  )
)
}
}

\keyword{internal}
