% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_update_ca_certificate}
\alias{iot_update_ca_certificate}
\title{Updates a registered CA certificate}
\usage{
iot_update_ca_certificate(certificateId, newStatus,
  newAutoRegistrationStatus, registrationConfig, removeAutoRegistration)
}
\arguments{
\item{certificateId}{[required] The CA certificate identifier.}

\item{newStatus}{The updated status of the CA certificate.

\strong{Note:} The status value REGISTER_INACTIVE is deprecated and should
not be used.}

\item{newAutoRegistrationStatus}{The new value for the auto registration status. Valid values are:
"ENABLE" or "DISABLE".}

\item{registrationConfig}{Information about the registration configuration.}

\item{removeAutoRegistration}{If true, removes auto registration.}
}
\value{
An empty list.
}
\description{
Updates a registered CA certificate.
}
\section{Request syntax}{
\preformatted{svc$update_ca_certificate(
  certificateId = "string",
  newStatus = "ACTIVE"|"INACTIVE",
  newAutoRegistrationStatus = "ENABLE"|"DISABLE",
  registrationConfig = list(
    templateBody = "string",
    roleArn = "string"
  ),
  removeAutoRegistration = TRUE|FALSE
)
}
}

\keyword{internal}
