% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_update_custom_metric}
\alias{iot_update_custom_metric}
\title{Updates a Device Defender detect custom metric}
\usage{
iot_update_custom_metric(metricName, displayName)
}
\arguments{
\item{metricName}{[required] The name of the custom metric. Cannot be updated.}

\item{displayName}{[required] Field represents a friendly name in the console for the custom metric,
it doesn't have to be unique. Don't use this name as the metric
identifier in the device metric report. Can be updated.}
}
\value{
A list with the following syntax:\preformatted{list(
  metricName = "string",
  metricArn = "string",
  metricType = "string-list"|"ip-address-list"|"number-list"|"number",
  displayName = "string",
  creationDate = as.POSIXct(
    "2015-01-01"
  ),
  lastModifiedDate = as.POSIXct(
    "2015-01-01"
  )
)
}
}
\description{
Updates a Device Defender detect custom metric.
}
\section{Request syntax}{
\preformatted{svc$update_custom_metric(
  metricName = "string",
  displayName = "string"
)
}
}

\keyword{internal}
