% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_update_domain_configuration}
\alias{iot_update_domain_configuration}
\title{Updates values stored in the domain configuration}
\usage{
iot_update_domain_configuration(domainConfigurationName,
  authorizerConfig, domainConfigurationStatus, removeAuthorizerConfig)
}
\arguments{
\item{domainConfigurationName}{[required] The name of the domain configuration to be updated.}

\item{authorizerConfig}{An object that specifies the authorization service for a domain.}

\item{domainConfigurationStatus}{The status to which the domain configuration should be updated.}

\item{removeAuthorizerConfig}{Removes the authorization configuration from a domain.}
}
\value{
A list with the following syntax:\preformatted{list(
  domainConfigurationName = "string",
  domainConfigurationArn = "string"
)
}
}
\description{
Updates values stored in the domain configuration. Domain configurations
for default endpoints can't be updated.

The domain configuration feature is in public preview and is subject to
change.
}
\section{Request syntax}{
\preformatted{svc$update_domain_configuration(
  domainConfigurationName = "string",
  authorizerConfig = list(
    defaultAuthorizerName = "string",
    allowAuthorizerOverride = TRUE|FALSE
  ),
  domainConfigurationStatus = "ENABLED"|"DISABLED",
  removeAuthorizerConfig = TRUE|FALSE
)
}
}

\keyword{internal}
