% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_update_event_configurations}
\alias{iot_update_event_configurations}
\title{Updates the event configurations}
\usage{
iot_update_event_configurations(eventConfigurations)
}
\arguments{
\item{eventConfigurations}{The new event configuration values.}
}
\value{
An empty list.
}
\description{
Updates the event configurations.
}
\section{Request syntax}{
\preformatted{svc$update_event_configurations(
  eventConfigurations = list(
    list(
      Enabled = TRUE|FALSE
    )
  )
)
}
}

\keyword{internal}
