% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_update_indexing_configuration}
\alias{iot_update_indexing_configuration}
\title{Updates the search configuration}
\usage{
iot_update_indexing_configuration(thingIndexingConfiguration,
  thingGroupIndexingConfiguration)
}
\arguments{
\item{thingIndexingConfiguration}{Thing indexing configuration.}

\item{thingGroupIndexingConfiguration}{Thing group indexing configuration.}
}
\value{
An empty list.
}
\description{
Updates the search configuration.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_indexing_configuration(
  thingIndexingConfiguration = list(
    thingIndexingMode = "OFF"|"REGISTRY"|"REGISTRY_AND_SHADOW",
    thingConnectivityIndexingMode = "OFF"|"STATUS",
    managedFields = list(
      list(
        name = "string",
        type = "Number"|"String"|"Boolean"
      )
    ),
    customFields = list(
      list(
        name = "string",
        type = "Number"|"String"|"Boolean"
      )
    )
  ),
  thingGroupIndexingConfiguration = list(
    thingGroupIndexingMode = "OFF"|"ON",
    managedFields = list(
      list(
        name = "string",
        type = "Number"|"String"|"Boolean"
      )
    ),
    customFields = list(
      list(
        name = "string",
        type = "Number"|"String"|"Boolean"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
