% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_update_job}
\alias{iot_update_job}
\title{Updates supported fields of the specified job}
\usage{
iot_update_job(jobId, description, presignedUrlConfig,
  jobExecutionsRolloutConfig, abortConfig, timeoutConfig, namespaceId)
}
\arguments{
\item{jobId}{[required] The ID of the job to be updated.}

\item{description}{A short text description of the job.}

\item{presignedUrlConfig}{Configuration information for pre-signed S3 URLs.}

\item{jobExecutionsRolloutConfig}{Allows you to create a staged rollout of the job.}

\item{abortConfig}{Allows you to create criteria to abort a job.}

\item{timeoutConfig}{Specifies the amount of time each device has to finish its execution of
the job. The timer is started when the job execution status is set to
\code{IN_PROGRESS}. If the job execution status is not set to another
terminal state before the time expires, it will be automatically set to
\code{TIMED_OUT}.}

\item{namespaceId}{The namespace used to indicate that a job is a customer-managed job.

When you specify a value for this parameter, AWS IoT Core sends jobs
notifications to MQTT topics that contain the value in the following
format.

\verb{$aws/things/<i>THING_NAME</i>/jobs/<i>JOB_ID</i>/notify-namespace-<i>NAMESPACE_ID</i>/}

The \code{namespaceId} feature is in public preview.}
}
\description{
Updates supported fields of the specified job.
}
\section{Request syntax}{
\preformatted{svc$update_job(
  jobId = "string",
  description = "string",
  presignedUrlConfig = list(
    roleArn = "string",
    expiresInSec = 123
  ),
  jobExecutionsRolloutConfig = list(
    maximumPerMinute = 123,
    exponentialRate = list(
      baseRatePerMinute = 123,
      incrementFactor = 123.0,
      rateIncreaseCriteria = list(
        numberOfNotifiedThings = 123,
        numberOfSucceededThings = 123
      )
    )
  ),
  abortConfig = list(
    criteriaList = list(
      list(
        failureType = "FAILED"|"REJECTED"|"TIMED_OUT"|"ALL",
        action = "CANCEL",
        thresholdPercentage = 123.0,
        minNumberOfExecutedThings = 123
      )
    )
  ),
  timeoutConfig = list(
    inProgressTimeoutInMinutes = 123
  ),
  namespaceId = "string"
)
}
}

\keyword{internal}
