% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_update_mitigation_action}
\alias{iot_update_mitigation_action}
\title{Updates the definition for the specified mitigation action}
\usage{
iot_update_mitigation_action(actionName, roleArn, actionParams)
}
\arguments{
\item{actionName}{[required] The friendly name for the mitigation action. You can't change the name
by using \code{UpdateMitigationAction}. Instead, you must delete and
re-create the mitigation action with the new name.}

\item{roleArn}{The ARN of the IAM role that is used to apply the mitigation action.}

\item{actionParams}{Defines the type of action and the parameters for that action.}
}
\description{
Updates the definition for the specified mitigation action.
}
\section{Request syntax}{
\preformatted{svc$update_mitigation_action(
  actionName = "string",
  roleArn = "string",
  actionParams = list(
    updateDeviceCertificateParams = list(
      action = "DEACTIVATE"
    ),
    updateCACertificateParams = list(
      action = "DEACTIVATE"
    ),
    addThingsToThingGroupParams = list(
      thingGroupNames = list(
        "string"
      ),
      overrideDynamicGroups = TRUE|FALSE
    ),
    replaceDefaultPolicyVersionParams = list(
      templateName = "BLANK_POLICY"
    ),
    enableIoTLoggingParams = list(
      roleArnForLogging = "string",
      logLevel = "DEBUG"|"INFO"|"ERROR"|"WARN"|"DISABLED"
    ),
    publishFindingToSnsParams = list(
      topicArn = "string"
    )
  )
)
}
}

\keyword{internal}
