% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_update_provisioning_template}
\alias{iot_update_provisioning_template}
\title{Updates a fleet provisioning template}
\usage{
iot_update_provisioning_template(templateName, description, enabled,
  defaultVersionId, provisioningRoleArn, preProvisioningHook,
  removePreProvisioningHook)
}
\arguments{
\item{templateName}{[required] The name of the fleet provisioning template.}

\item{description}{The description of the fleet provisioning template.}

\item{enabled}{True to enable the fleet provisioning template, otherwise false.}

\item{defaultVersionId}{The ID of the default provisioning template version.}

\item{provisioningRoleArn}{The ARN of the role associated with the provisioning template. This IoT
role grants permission to provision a device.}

\item{preProvisioningHook}{Updates the pre-provisioning hook template.}

\item{removePreProvisioningHook}{Removes pre-provisioning hook template.}
}
\value{
An empty list.
}
\description{
Updates a fleet provisioning template.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_provisioning_template(
  templateName = "string",
  description = "string",
  enabled = TRUE|FALSE,
  defaultVersionId = 123,
  provisioningRoleArn = "string",
  preProvisioningHook = list(
    payloadVersion = "string",
    targetArn = "string"
  ),
  removePreProvisioningHook = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
