% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_update_role_alias}
\alias{iot_update_role_alias}
\title{Updates a role alias}
\usage{
iot_update_role_alias(roleAlias, roleArn, credentialDurationSeconds)
}
\arguments{
\item{roleAlias}{[required] The role alias to update.}

\item{roleArn}{The role ARN.}

\item{credentialDurationSeconds}{The number of seconds the credential will be valid.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  roleAlias = "string",
  roleAliasArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Updates a role alias.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_role_alias(
  roleAlias = "string",
  roleArn = "string",
  credentialDurationSeconds = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
