% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_update_scheduled_audit}
\alias{iot_update_scheduled_audit}
\title{Updates a scheduled audit, including which checks are performed and how
often the audit takes place}
\usage{
iot_update_scheduled_audit(frequency, dayOfMonth, dayOfWeek,
  targetCheckNames, scheduledAuditName)
}
\arguments{
\item{frequency}{How often the scheduled audit takes place, either \code{DAILY}, \code{WEEKLY},
\code{BIWEEKLY}, or \code{MONTHLY}. The start time of each audit is determined by
the system.}

\item{dayOfMonth}{The day of the month on which the scheduled audit takes place. This can
be \code{1} through \code{31} or \code{LAST}. This field is required if the \code{frequency}
parameter is set to \code{MONTHLY}. If days 29-31 are specified, and the
month does not have that many days, the audit takes place on the "LAST"
day of the month.}

\item{dayOfWeek}{The day of the week on which the scheduled audit takes place. This can
be one of \code{SUN}, \code{MON}, \code{TUE}, \code{WED}, \code{THU}, \code{FRI}, or \code{SAT}. This field
is required if the "frequency" parameter is set to \code{WEEKLY} or
\code{BIWEEKLY}.}

\item{targetCheckNames}{Which checks are performed during the scheduled audit. Checks must be
enabled for your account. (Use
\code{\link[=iot_describe_account_audit_configuration]{describe_account_audit_configuration}}
to see the list of all checks, including those that are enabled or use
\code{\link[=iot_update_account_audit_configuration]{update_account_audit_configuration}}
to select which checks are enabled.)}

\item{scheduledAuditName}{[required] The name of the scheduled audit. (Max. 128 chars)}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  scheduledAuditArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Updates a scheduled audit, including which checks are performed and how
often the audit takes place.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_scheduled_audit(
  frequency = "DAILY"|"WEEKLY"|"BIWEEKLY"|"MONTHLY",
  dayOfMonth = "string",
  dayOfWeek = "SUN"|"MON"|"TUE"|"WED"|"THU"|"FRI"|"SAT",
  targetCheckNames = list(
    "string"
  ),
  scheduledAuditName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
