% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_update_security_profile}
\alias{iot_update_security_profile}
\title{Updates a Device Defender security profile}
\usage{
iot_update_security_profile(securityProfileName,
  securityProfileDescription, behaviors, alertTargets,
  additionalMetricsToRetain, additionalMetricsToRetainV2, deleteBehaviors,
  deleteAlertTargets, deleteAdditionalMetricsToRetain, expectedVersion)
}
\arguments{
\item{securityProfileName}{[required] The name of the security profile you want to update.}

\item{securityProfileDescription}{A description of the security profile.}

\item{behaviors}{Specifies the behaviors that, when violated by a device (thing), cause
an alert.}

\item{alertTargets}{Where the alerts are sent. (Alerts are always sent to the console.)}

\item{additionalMetricsToRetain}{\emph{Please use UpdateSecurityProfileRequest$additionalMetricsToRetainV2
instead.}

A list of metrics whose data is retained (stored). By default, data is
retained for any metric used in the profile's \code{behaviors}, but it is
also retained for any metric specified here. Can be used with custom
metrics; cannot be used with dimensions.}

\item{additionalMetricsToRetainV2}{A list of metrics whose data is retained (stored). By default, data is
retained for any metric used in the profile's behaviors, but it is also
retained for any metric specified here. Can be used with custom metrics;
cannot be used with dimensions.}

\item{deleteBehaviors}{If true, delete all \code{behaviors} defined for this security profile. If
any \code{behaviors} are defined in the current invocation, an exception
occurs.}

\item{deleteAlertTargets}{If true, delete all \code{alertTargets} defined for this security profile. If
any \code{alertTargets} are defined in the current invocation, an exception
occurs.}

\item{deleteAdditionalMetricsToRetain}{If true, delete all \code{additionalMetricsToRetain} defined for this
security profile. If any \code{additionalMetricsToRetain} are defined in the
current invocation, an exception occurs.}

\item{expectedVersion}{The expected version of the security profile. A new version is generated
whenever the security profile is updated. If you specify a value that is
different from the actual version, a \code{VersionConflictException} is
thrown.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  securityProfileName = "string",
  securityProfileArn = "string",
  securityProfileDescription = "string",
  behaviors = list(
    list(
      name = "string",
      metric = "string",
      metricDimension = list(
        dimensionName = "string",
        operator = "IN"|"NOT_IN"
      ),
      criteria = list(
        comparisonOperator = "less-than"|"less-than-equals"|"greater-than"|"greater-than-equals"|"in-cidr-set"|"not-in-cidr-set"|"in-port-set"|"not-in-port-set"|"in-set"|"not-in-set",
        value = list(
          count = 123,
          cidrs = list(
            "string"
          ),
          ports = list(
            123
          ),
          number = 123.0,
          numbers = list(
            123.0
          ),
          strings = list(
            "string"
          )
        ),
        durationSeconds = 123,
        consecutiveDatapointsToAlarm = 123,
        consecutiveDatapointsToClear = 123,
        statisticalThreshold = list(
          statistic = "string"
        ),
        mlDetectionConfig = list(
          confidenceLevel = "LOW"|"MEDIUM"|"HIGH"
        )
      ),
      suppressAlerts = TRUE|FALSE
    )
  ),
  alertTargets = list(
    list(
      alertTargetArn = "string",
      roleArn = "string"
    )
  ),
  additionalMetricsToRetain = list(
    "string"
  ),
  additionalMetricsToRetainV2 = list(
    list(
      metric = "string",
      metricDimension = list(
        dimensionName = "string",
        operator = "IN"|"NOT_IN"
      )
    )
  ),
  version = 123,
  creationDate = as.POSIXct(
    "2015-01-01"
  ),
  lastModifiedDate = as.POSIXct(
    "2015-01-01"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Updates a Device Defender security profile.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_security_profile(
  securityProfileName = "string",
  securityProfileDescription = "string",
  behaviors = list(
    list(
      name = "string",
      metric = "string",
      metricDimension = list(
        dimensionName = "string",
        operator = "IN"|"NOT_IN"
      ),
      criteria = list(
        comparisonOperator = "less-than"|"less-than-equals"|"greater-than"|"greater-than-equals"|"in-cidr-set"|"not-in-cidr-set"|"in-port-set"|"not-in-port-set"|"in-set"|"not-in-set",
        value = list(
          count = 123,
          cidrs = list(
            "string"
          ),
          ports = list(
            123
          ),
          number = 123.0,
          numbers = list(
            123.0
          ),
          strings = list(
            "string"
          )
        ),
        durationSeconds = 123,
        consecutiveDatapointsToAlarm = 123,
        consecutiveDatapointsToClear = 123,
        statisticalThreshold = list(
          statistic = "string"
        ),
        mlDetectionConfig = list(
          confidenceLevel = "LOW"|"MEDIUM"|"HIGH"
        )
      ),
      suppressAlerts = TRUE|FALSE
    )
  ),
  alertTargets = list(
    list(
      alertTargetArn = "string",
      roleArn = "string"
    )
  ),
  additionalMetricsToRetain = list(
    "string"
  ),
  additionalMetricsToRetainV2 = list(
    list(
      metric = "string",
      metricDimension = list(
        dimensionName = "string",
        operator = "IN"|"NOT_IN"
      )
    )
  ),
  deleteBehaviors = TRUE|FALSE,
  deleteAlertTargets = TRUE|FALSE,
  deleteAdditionalMetricsToRetain = TRUE|FALSE,
  expectedVersion = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
