% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_update_stream}
\alias{iot_update_stream}
\title{Updates an existing stream}
\usage{
iot_update_stream(streamId, description, files, roleArn)
}
\arguments{
\item{streamId}{[required] The stream ID.}

\item{description}{The description of the stream.}

\item{files}{The files associated with the stream.}

\item{roleArn}{An IAM role that allows the IoT service principal assumes to access your
S3 files.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  streamId = "string",
  streamArn = "string",
  description = "string",
  streamVersion = 123
)
}\if{html}{\out{</div>}}
}
\description{
Updates an existing stream. The stream version will be incremented by
one.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_stream(
  streamId = "string",
  description = "string",
  files = list(
    list(
      fileId = 123,
      s3Location = list(
        bucket = "string",
        key = "string",
        version = "string"
      )
    )
  ),
  roleArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
