% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_update_thing}
\alias{iot_update_thing}
\title{Updates the data for a thing}
\usage{
iot_update_thing(thingName, thingTypeName, attributePayload,
  expectedVersion, removeThingType)
}
\arguments{
\item{thingName}{[required] The name of the thing to update.

You can't change a thing's name. To change a thing's name, you must
create a new thing, give it the new name, and then delete the old thing.}

\item{thingTypeName}{The name of the thing type.}

\item{attributePayload}{A list of thing attributes, a JSON string containing name-value pairs.
For example:

\verb{\{\\"attributes\\":\{\\"name1\\":\\"value2\\"\}\}}

This data is used to add new attributes or update existing attributes.}

\item{expectedVersion}{The expected version of the thing record in the registry. If the version
of the record in the registry does not match the expected version
specified in the request, the \code{\link[=iot_update_thing]{update_thing}} request
is rejected with a \code{VersionConflictException}.}

\item{removeThingType}{Remove a thing type association. If \strong{true}, the association is
removed.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Updates the data for a thing.
}
\section{Request syntax}{
\preformatted{svc$update_thing(
  thingName = "string",
  thingTypeName = "string",
  attributePayload = list(
    attributes = list(
      "string"
    ),
    merge = TRUE|FALSE
  ),
  expectedVersion = 123,
  removeThingType = TRUE|FALSE
)
}
}

\keyword{internal}
