% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_update_thing_group}
\alias{iot_update_thing_group}
\title{Update a thing group}
\usage{
iot_update_thing_group(thingGroupName, thingGroupProperties,
  expectedVersion)
}
\arguments{
\item{thingGroupName}{[required] The thing group to update.}

\item{thingGroupProperties}{[required] The thing group properties.}

\item{expectedVersion}{The expected version of the thing group. If this does not match the
version of the thing group being updated, the update will fail.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  version = 123
)
}\if{html}{\out{</div>}}
}
\description{
Update a thing group.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_thing_group(
  thingGroupName = "string",
  thingGroupProperties = list(
    thingGroupDescription = "string",
    attributePayload = list(
      attributes = list(
        "string"
      ),
      merge = TRUE|FALSE
    )
  ),
  expectedVersion = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
