% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_cancel_pipeline_reprocessing}
\alias{iotanalytics_cancel_pipeline_reprocessing}
\title{Cancels the reprocessing of data through the pipeline}
\usage{
iotanalytics_cancel_pipeline_reprocessing(pipelineName, reprocessingId)
}
\arguments{
\item{pipelineName}{[required] The name of pipeline for which data reprocessing is canceled.}

\item{reprocessingId}{[required] The ID of the reprocessing task (returned by
\code{\link[=iotanalytics_start_pipeline_reprocessing]{start_pipeline_reprocessing}}).}
}
\value{
An empty list.
}
\description{
Cancels the reprocessing of data through the pipeline.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$cancel_pipeline_reprocessing(
  pipelineName = "string",
  reprocessingId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
