% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_create_datastore}
\alias{iotanalytics_create_datastore}
\title{Creates a data store, which is a repository for messages}
\usage{
iotanalytics_create_datastore(datastoreName, datastoreStorage,
  retentionPeriod, tags)
}
\arguments{
\item{datastoreName}{[required] The name of the data store.}

\item{datastoreStorage}{Where data store data is stored. You may choose one of
"serviceManagedS3" or "customerManagedS3" storage. If not specified, the
default is "serviceManagedS3". This cannot be changed after the data
store is created.}

\item{retentionPeriod}{How long, in days, message data is kept for the data store. When
"customerManagedS3" storage is selected, this parameter is ignored.}

\item{tags}{Metadata which can be used to manage the data store.}
}
\description{
Creates a data store, which is a repository for messages.
}
\section{Request syntax}{
\preformatted{svc$create_datastore(
  datastoreName = "string",
  datastoreStorage = list(
    serviceManagedS3 = list(),
    customerManagedS3 = list(
      bucket = "string",
      keyPrefix = "string",
      roleArn = "string"
    )
  ),
  retentionPeriod = list(
    unlimited = TRUE|FALSE,
    numberOfDays = 123
  ),
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  )
)
}
}

\keyword{internal}
