% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_create_pipeline}
\alias{iotanalytics_create_pipeline}
\title{Creates a pipeline}
\usage{
iotanalytics_create_pipeline(pipelineName, pipelineActivities, tags)
}
\arguments{
\item{pipelineName}{[required] The name of the pipeline.}

\item{pipelineActivities}{[required] A list of \code{PipelineActivity} objects. Activities perform transformations
on your messages, such as removing, renaming or adding message
attributes; filtering messages based on attribute values; invoking your
Lambda functions on messages for advanced processing; or performing
mathematical transformations to normalize device data.

The list can be 2-25 \code{PipelineActivity} objects and must contain both a
\code{channel} and a \code{datastore} activity. Each entry in the list must
contain only one activity. For example:

\verb{pipelineActivities = \\\[ \\\{ "channel": \\\{ ... \\\} \\\}, \\\{ "lambda": \\\{ ... \\\} \\\}, ... \\\]}}

\item{tags}{Metadata which can be used to manage the pipeline.}
}
\description{
Creates a pipeline. A pipeline consumes messages from a channel and
allows you to process the messages before storing them in a data store.
You must specify both a \code{channel} and a \code{datastore} activity and,
optionally, as many as 23 additional activities in the
\code{pipelineActivities} array.
}
\section{Request syntax}{
\preformatted{svc$create_pipeline(
  pipelineName = "string",
  pipelineActivities = list(
    list(
      channel = list(
        name = "string",
        channelName = "string",
        next = "string"
      ),
      lambda = list(
        name = "string",
        lambdaName = "string",
        batchSize = 123,
        next = "string"
      ),
      datastore = list(
        name = "string",
        datastoreName = "string"
      ),
      addAttributes = list(
        name = "string",
        attributes = list(
          "string"
        ),
        next = "string"
      ),
      removeAttributes = list(
        name = "string",
        attributes = list(
          "string"
        ),
        next = "string"
      ),
      selectAttributes = list(
        name = "string",
        attributes = list(
          "string"
        ),
        next = "string"
      ),
      filter = list(
        name = "string",
        filter = "string",
        next = "string"
      ),
      math = list(
        name = "string",
        attribute = "string",
        math = "string",
        next = "string"
      ),
      deviceRegistryEnrich = list(
        name = "string",
        attribute = "string",
        thingName = "string",
        roleArn = "string",
        next = "string"
      ),
      deviceShadowEnrich = list(
        name = "string",
        attribute = "string",
        thingName = "string",
        roleArn = "string",
        next = "string"
      )
    )
  ),
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  )
)
}
}

\keyword{internal}
