% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_delete_datastore}
\alias{iotanalytics_delete_datastore}
\title{Deletes the specified data store}
\usage{
iotanalytics_delete_datastore(datastoreName)
}
\arguments{
\item{datastoreName}{[required] The name of the data store to delete.}
}
\value{
An empty list.
}
\description{
Deletes the specified data store.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_datastore(
  datastoreName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
