% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_describe_datastore}
\alias{iotanalytics_describe_datastore}
\title{Retrieves information about a data store}
\usage{
iotanalytics_describe_datastore(datastoreName, includeStatistics)
}
\arguments{
\item{datastoreName}{[required] The name of the data store}

\item{includeStatistics}{If true, additional statistical information about the data store is
included in the response. This feature cannot be used with a data store
whose S3 storage is customer-managed.}
}
\description{
Retrieves information about a data store.
}
\section{Request syntax}{
\preformatted{svc$describe_datastore(
  datastoreName = "string",
  includeStatistics = TRUE|FALSE
)
}
}

\keyword{internal}
