% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_get_dataset_content}
\alias{iotanalytics_get_dataset_content}
\title{Retrieves the contents of a data set as presigned URIs}
\usage{
iotanalytics_get_dataset_content(datasetName, versionId)
}
\arguments{
\item{datasetName}{[required] The name of the data set whose contents are retrieved.}

\item{versionId}{The version of the data set whose contents are retrieved. You can also
use the strings "$LATEST" or "$LATEST_SUCCEEDED" to retrieve the
contents of the latest or latest successfully completed data set. If not
specified, "$LATEST_SUCCEEDED" is the default.}
}
\value{
A list with the following syntax:\preformatted{list(
  entries = list(
    list(
      entryName = "string",
      dataURI = "string"
    )
  ),
  timestamp = as.POSIXct(
    "2015-01-01"
  ),
  status = list(
    state = "CREATING"|"SUCCEEDED"|"FAILED",
    reason = "string"
  )
)
}
}
\description{
Retrieves the contents of a data set as presigned URIs.
}
\section{Request syntax}{
\preformatted{svc$get_dataset_content(
  datasetName = "string",
  versionId = "string"
)
}
}

\keyword{internal}
