% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_list_datasets}
\alias{iotanalytics_list_datasets}
\title{Retrieves information about data sets}
\usage{
iotanalytics_list_datasets(nextToken, maxResults)
}
\arguments{
\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to return in this request.

The default value is 100.}
}
\value{
A list with the following syntax:\preformatted{list(
  datasetSummaries = list(
    list(
      datasetName = "string",
      status = "CREATING"|"ACTIVE"|"DELETING",
      creationTime = as.POSIXct(
        "2015-01-01"
      ),
      lastUpdateTime = as.POSIXct(
        "2015-01-01"
      ),
      triggers = list(
        list(
          schedule = list(
            expression = "string"
          ),
          dataset = list(
            name = "string"
          )
        )
      ),
      actions = list(
        list(
          actionName = "string",
          actionType = "QUERY"|"CONTAINER"
        )
      )
    )
  ),
  nextToken = "string"
)
}
}
\description{
Retrieves information about data sets.
}
\section{Request syntax}{
\preformatted{svc$list_datasets(
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
