% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_put_logging_options}
\alias{iotanalytics_put_logging_options}
\title{Sets or updates the AWS IoT Analytics logging options}
\usage{
iotanalytics_put_logging_options(loggingOptions)
}
\arguments{
\item{loggingOptions}{[required] The new values of the AWS IoT Analytics logging options.}
}
\description{
Sets or updates the AWS IoT Analytics logging options.

If you update the value of any \code{loggingOptions} field, it takes up to
one minute for the change to take effect. Also, if you change the policy
attached to the role you specified in the \code{roleArn} field (for example,
to correct an invalid policy), it takes up to five minutes for that
change to take effect.
}
\section{Request syntax}{
\preformatted{svc$put_logging_options(
  loggingOptions = list(
    roleArn = "string",
    level = "ERROR",
    enabled = TRUE|FALSE
  )
)
}
}

\keyword{internal}
