% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_start_pipeline_reprocessing}
\alias{iotanalytics_start_pipeline_reprocessing}
\title{Starts the reprocessing of raw message data through the pipeline}
\usage{
iotanalytics_start_pipeline_reprocessing(pipelineName, startTime,
  endTime, channelMessages)
}
\arguments{
\item{pipelineName}{[required] The name of the pipeline on which to start reprocessing.}

\item{startTime}{The start time (inclusive) of raw message data that is reprocessed.

If you specify a value for the \code{startTime} parameter, you must not use
the \code{channelMessages} object.}

\item{endTime}{The end time (exclusive) of raw message data that is reprocessed.

If you specify a value for the \code{endTime} parameter, you must not use the
\code{channelMessages} object.}

\item{channelMessages}{Specifies one or more sets of channel messages that you want to
reprocess.

If you use the \code{channelMessages} object, you must not specify a value
for \code{startTime} and \code{endTime}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  reprocessingId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Starts the reprocessing of raw message data through the pipeline.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$start_pipeline_reprocessing(
  pipelineName = "string",
  startTime = as.POSIXct(
    "2015-01-01"
  ),
  endTime = as.POSIXct(
    "2015-01-01"
  ),
  channelMessages = list(
    s3Paths = list(
      "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
