% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_tag_resource}
\alias{iotanalytics_tag_resource}
\title{Adds to or modifies the tags of the given resource}
\usage{
iotanalytics_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource whose tags you want to modify.}

\item{tags}{[required] The new or modified tags for the resource.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Adds to or modifies the tags of the given resource. Tags are metadata
that can be used to manage a resource.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  resourceArn = "string",
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  )
)
}
}

\keyword{internal}
