% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_update_channel}
\alias{iotanalytics_update_channel}
\title{Updates the settings of a channel}
\usage{
iotanalytics_update_channel(channelName, channelStorage,
  retentionPeriod)
}
\arguments{
\item{channelName}{[required] The name of the channel to be updated.}

\item{channelStorage}{Where channel data is stored. You can choose one of \code{serviceManagedS3}
or \code{customerManagedS3} storage. If not specified, the default is
\code{serviceManagedS3}. You cannot change this storage option after the
channel is created.}

\item{retentionPeriod}{How long, in days, message data is kept for the channel. The retention
period cannot be updated if the channel's S3 storage is
customer-managed.}
}
\description{
Updates the settings of a channel.
}
\section{Request syntax}{
\preformatted{svc$update_channel(
  channelName = "string",
  channelStorage = list(
    serviceManagedS3 = list(),
    customerManagedS3 = list(
      bucket = "string",
      keyPrefix = "string",
      roleArn = "string"
    )
  ),
  retentionPeriod = list(
    unlimited = TRUE|FALSE,
    numberOfDays = 123
  )
)
}
}

\keyword{internal}
