% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_update_dataset}
\alias{iotanalytics_update_dataset}
\title{Updates the settings of a data set}
\usage{
iotanalytics_update_dataset(datasetName, actions, triggers,
  contentDeliveryRules, retentionPeriod, versioningConfiguration,
  lateDataRules)
}
\arguments{
\item{datasetName}{[required] The name of the data set to update.}

\item{actions}{[required] A list of \code{DatasetAction} objects.}

\item{triggers}{A list of \code{DatasetTrigger} objects. The list can be empty or can contain
up to five \code{DatasetTrigger} objects.}

\item{contentDeliveryRules}{When dataset contents are created, they are delivered to destinations
specified here.}

\item{retentionPeriod}{How long, in days, dataset contents are kept for the dataset.}

\item{versioningConfiguration}{Optional. How many versions of dataset contents are kept. If not
specified or set to null, only the latest version plus the latest
succeeded version (if they are different) are kept for the time period
specified by the \code{retentionPeriod} parameter. For more information, see
\href{https://docs.aws.amazon.com/iotanalytics/latest/userguide/getting-started.html#aws-iot-analytics-dataset-versions}{Keeping Multiple Versions of AWS IoT Analytics Data Sets}
in the \emph{AWS IoT Analytics User Guide}.}

\item{lateDataRules}{A list of data rules that send notifications to Amazon CloudWatch, when
data arrives late. To specify \code{lateDataRules}, the dataset must use a
\href{https://docs.aws.amazon.com/iotanalytics/latest/APIReference/API_DeltaTime.html}{DeltaTimer}
filter.}
}
\value{
An empty list.
}
\description{
Updates the settings of a data set.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_dataset(
  datasetName = "string",
  actions = list(
    list(
      actionName = "string",
      queryAction = list(
        sqlQuery = "string",
        filters = list(
          list(
            deltaTime = list(
              offsetSeconds = 123,
              timeExpression = "string"
            )
          )
        )
      ),
      containerAction = list(
        image = "string",
        executionRoleArn = "string",
        resourceConfiguration = list(
          computeType = "ACU_1"|"ACU_2",
          volumeSizeInGB = 123
        ),
        variables = list(
          list(
            name = "string",
            stringValue = "string",
            doubleValue = 123.0,
            datasetContentVersionValue = list(
              datasetName = "string"
            ),
            outputFileUriValue = list(
              fileName = "string"
            )
          )
        )
      )
    )
  ),
  triggers = list(
    list(
      schedule = list(
        expression = "string"
      ),
      dataset = list(
        name = "string"
      )
    )
  ),
  contentDeliveryRules = list(
    list(
      entryName = "string",
      destination = list(
        iotEventsDestinationConfiguration = list(
          inputName = "string",
          roleArn = "string"
        ),
        s3DestinationConfiguration = list(
          bucket = "string",
          key = "string",
          glueConfiguration = list(
            tableName = "string",
            databaseName = "string"
          ),
          roleArn = "string"
        )
      )
    )
  ),
  retentionPeriod = list(
    unlimited = TRUE|FALSE,
    numberOfDays = 123
  ),
  versioningConfiguration = list(
    unlimited = TRUE|FALSE,
    maxVersions = 123
  ),
  lateDataRules = list(
    list(
      ruleName = "string",
      ruleConfiguration = list(
        deltaTimeSessionWindowConfiguration = list(
          timeoutInMinutes = 123
        )
      )
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
