% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_update_datastore}
\alias{iotanalytics_update_datastore}
\title{Updates the settings of a data store}
\usage{
iotanalytics_update_datastore(datastoreName, retentionPeriod,
  datastoreStorage, fileFormatConfiguration)
}
\arguments{
\item{datastoreName}{[required] The name of the data store to be updated.}

\item{retentionPeriod}{How long, in days, message data is kept for the data store. The
retention period cannot be updated if the data store's S3 storage is
customer-managed.}

\item{datastoreStorage}{Where data store data is stored. You can choose one of
\code{serviceManagedS3} or \code{customerManagedS3} storage. If not specified, the
default is\code{serviceManagedS3}. You cannot change this storage option
after the data store is created.}

\item{fileFormatConfiguration}{Contains the configuration information of file formats. AWS IoT
Analytics data stores support JSON and
\href{https://parquet.apache.org/}{Parquet}.

The default file format is JSON. You can specify only one format.

You can't change the file format after you create the data store.}
}
\value{
An empty list.
}
\description{
Updates the settings of a data store.
}
\section{Request syntax}{
\preformatted{svc$update_datastore(
  datastoreName = "string",
  retentionPeriod = list(
    unlimited = TRUE|FALSE,
    numberOfDays = 123
  ),
  datastoreStorage = list(
    serviceManagedS3 = list(),
    customerManagedS3 = list(
      bucket = "string",
      keyPrefix = "string",
      roleArn = "string"
    )
  ),
  fileFormatConfiguration = list(
    jsonConfiguration = list(),
    parquetConfiguration = list(
      schemaDefinition = list(
        columns = list(
          list(
            name = "string",
            type = "string"
          )
        )
      )
    )
  )
)
}
}

\keyword{internal}
