% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotdataplane_service.R
\name{iotdataplane}
\alias{iotdataplane}
\title{AWS IoT Data Plane}
\usage{
iotdataplane(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\value{
A client for the service. You can call the service's operations using
syntax like \code{svc$operation(...)}, where \code{svc} is the name you've assigned
to the client. The available operations are listed in the
Operations section.
}
\description{
AWS IoT

AWS IoT-Data enables secure, bi-directional communication between
Internet-connected things (such as sensors, actuators, embedded devices,
or smart appliances) and the AWS cloud. It implements a broker for
applications and things to publish messages over HTTP (Publish) and
retrieve, update, and delete shadows. A shadow is a persistent
representation of your things and their state in the AWS cloud.

Find the endpoint address for actions in the AWS IoT data plane by
running this CLI command:

\verb{aws iot describe-endpoint --endpoint-type iot:Data-ATS}

The service name used by \href{https://docs.aws.amazon.com/general/latest/gr/signature-version-4.html}{AWS Signature Version 4}
to sign requests is: \emph{iotdevicegateway}.
}
\section{Service syntax}{
\preformatted{svc <- iotdataplane(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=iotdataplane_delete_thing_shadow]{delete_thing_shadow} \tab Deletes the shadow for the specified thing\cr
\link[=iotdataplane_get_thing_shadow]{get_thing_shadow} \tab Gets the shadow for the specified thing\cr
\link[=iotdataplane_list_named_shadows_for_thing]{list_named_shadows_for_thing} \tab Lists the shadows for the specified thing\cr
\link[=iotdataplane_publish]{publish} \tab Publishes state information\cr
\link[=iotdataplane_update_thing_shadow]{update_thing_shadow} \tab Updates the shadow for the specified thing
}
}

\examples{
\dontrun{
svc <- iotdataplane()
svc$delete_thing_shadow(
  Foo = 123
)
}

}
