% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotdataplane_operations.R
\name{iotdataplane_get_thing_shadow}
\alias{iotdataplane_get_thing_shadow}
\title{Gets the shadow for the specified thing}
\usage{
iotdataplane_get_thing_shadow(thingName, shadowName)
}
\arguments{
\item{thingName}{[required] The name of the thing.}

\item{shadowName}{The name of the shadow.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  payload = raw
)
}\if{html}{\out{</div>}}
}
\description{
Gets the shadow for the specified thing.

For more information, see
\code{\link[=iotdataplane_get_thing_shadow]{get_thing_shadow}} in the AWS IoT
Developer Guide.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_thing_shadow(
  thingName = "string",
  shadowName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
