% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotdataplane_operations.R
\name{iotdataplane_list_named_shadows_for_thing}
\alias{iotdataplane_list_named_shadows_for_thing}
\title{Lists the shadows for the specified thing}
\usage{
iotdataplane_list_named_shadows_for_thing(thingName, nextToken,
  pageSize)
}
\arguments{
\item{thingName}{[required] The name of the thing.}

\item{nextToken}{The token to retrieve the next set of results.}

\item{pageSize}{The result page size.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  results = list(
    "string"
  ),
  nextToken = "string",
  timestamp = 123
)
}\if{html}{\out{</div>}}
}
\description{
Lists the shadows for the specified thing.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_named_shadows_for_thing(
  thingName = "string",
  nextToken = "string",
  pageSize = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
