% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotjobsdataplane_operations.R
\name{iotjobsdataplane_describe_job_execution}
\alias{iotjobsdataplane_describe_job_execution}
\title{Gets details of a job execution}
\usage{
iotjobsdataplane_describe_job_execution(jobId, thingName,
  includeJobDocument, executionNumber)
}
\arguments{
\item{jobId}{[required] The unique identifier assigned to this job when it was created.}

\item{thingName}{[required] The thing name associated with the device the job execution is running
on.}

\item{includeJobDocument}{Optional. When set to true, the response contains the job document. The
default is false.}

\item{executionNumber}{Optional. A number that identifies a particular job execution on a
particular device. If not specified, the latest job execution is
returned.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  execution = list(
    jobId = "string",
    thingName = "string",
    status = "QUEUED"|"IN_PROGRESS"|"SUCCEEDED"|"FAILED"|"TIMED_OUT"|"REJECTED"|"REMOVED"|"CANCELED",
    statusDetails = list(
      "string"
    ),
    queuedAt = 123,
    startedAt = 123,
    lastUpdatedAt = 123,
    approximateSecondsBeforeTimedOut = 123,
    versionNumber = 123,
    executionNumber = 123,
    jobDocument = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets details of a job execution.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_job_execution(
  jobId = "string",
  thingName = "string",
  includeJobDocument = TRUE|FALSE,
  executionNumber = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
