% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signer_service.R
\name{signer}
\alias{signer}
\title{AWS Signer}
\usage{
signer(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\description{
With code signing for IoT, you can sign code that you create for any IoT
device that is supported by Amazon Web Services (AWS). Code signing is
available through \href{http://docs.aws.amazon.com/freertos/latest/userguide/}{Amazon FreeRTOS} and
\href{http://docs.aws.amazon.com/iot/latest/developerguide/}{AWS IoT Device Management}, and
integrated with \href{http://docs.aws.amazon.com/acm/latest/userguide/}{AWS Certificate Manager (ACM)}. In order to
sign code, you import a third-party code signing certificate with ACM
that is used to sign updates in Amazon FreeRTOS and AWS IoT Device
Management. For general information about using code signing, see the
\href{http://docs.aws.amazon.com/signer/latest/developerguide/Welcome.html}{Code Signing for IoT Developer Guide}.
}
\section{Service syntax}{
\preformatted{svc <- signer(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=signer_cancel_signing_profile]{cancel_signing_profile} \tab Changes the state of an ACTIVE signing profile to CANCELED \cr
\link[=signer_describe_signing_job]{describe_signing_job} \tab Returns information about a specific code signing job \cr
\link[=signer_get_signing_platform]{get_signing_platform} \tab Returns information on a specific signing platform \cr
\link[=signer_get_signing_profile]{get_signing_profile} \tab Returns information on a specific signing profile \cr
\link[=signer_list_signing_jobs]{list_signing_jobs} \tab Lists all your signing jobs \cr
\link[=signer_list_signing_platforms]{list_signing_platforms} \tab Lists all signing platforms available in code signing that match the request parameters\cr
\link[=signer_list_signing_profiles]{list_signing_profiles} \tab Lists all available signing profiles in your AWS account \cr
\link[=signer_list_tags_for_resource]{list_tags_for_resource} \tab Returns a list of the tags associated with a signing profile resource \cr
\link[=signer_put_signing_profile]{put_signing_profile} \tab Creates a signing profile \cr
\link[=signer_start_signing_job]{start_signing_job} \tab Initiates a signing job to be performed on the code provided \cr
\link[=signer_tag_resource]{tag_resource} \tab Adds one or more tags to a signing profile \cr
\link[=signer_untag_resource]{untag_resource} \tab Removes one or more tags from a signing profile
}
}

\examples{
\dontrun{
svc <- signer()
svc$cancel_signing_profile(
  Foo = 123
)
}

}
