% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signer_operations.R
\name{signer_add_profile_permission}
\alias{signer_add_profile_permission}
\title{Adds cross-account permissions to a signing profile}
\usage{
signer_add_profile_permission(profileName, profileVersion, action,
  principal, revisionId, statementId)
}
\arguments{
\item{profileName}{[required] The human-readable name of the signing profile.}

\item{profileVersion}{The version of the signing profile.}

\item{action}{[required] The AWS Signer action permitted as part of cross-account permissions.}

\item{principal}{[required] The AWS principal receiving cross-account permissions. This may be an
IAM role or another AWS account ID.}

\item{revisionId}{A unique identifier for the current profile revision.}

\item{statementId}{[required] A unique identifier for the cross-account permission statement.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  revisionId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Adds cross-account permissions to a signing profile.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$add_profile_permission(
  profileName = "string",
  profileVersion = "string",
  action = "string",
  principal = "string",
  revisionId = "string",
  statementId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
