% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signer_operations.R
\name{signer_get_signing_platform}
\alias{signer_get_signing_platform}
\title{Returns information on a specific signing platform}
\usage{
signer_get_signing_platform(platformId)
}
\arguments{
\item{platformId}{[required] The ID of the target signing platform.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  platformId = "string",
  displayName = "string",
  partner = "string",
  target = "string",
  category = "AWSIoT",
  signingConfiguration = list(
    encryptionAlgorithmOptions = list(
      allowedValues = list(
        "RSA"|"ECDSA"
      ),
      defaultValue = "RSA"|"ECDSA"
    ),
    hashAlgorithmOptions = list(
      allowedValues = list(
        "SHA1"|"SHA256"
      ),
      defaultValue = "SHA1"|"SHA256"
    )
  ),
  signingImageFormat = list(
    supportedFormats = list(
      "JSON"|"JSONEmbedded"|"JSONDetached"
    ),
    defaultFormat = "JSON"|"JSONEmbedded"|"JSONDetached"
  ),
  maxSizeInMB = 123,
  revocationSupported = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}
\description{
Returns information on a specific signing platform.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_signing_platform(
  platformId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
