% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signer_operations.R
\name{signer_list_signing_jobs}
\alias{signer_list_signing_jobs}
\title{Lists all your signing jobs}
\usage{
signer_list_signing_jobs(status, platformId, requestedBy, maxResults,
  nextToken, isRevoked, signatureExpiresBefore, signatureExpiresAfter,
  jobInvoker)
}
\arguments{
\item{status}{A status value with which to filter your results.}

\item{platformId}{The ID of microcontroller platform that you specified for the
distribution of your code image.}

\item{requestedBy}{The IAM principal that requested the signing job.}

\item{maxResults}{Specifies the maximum number of items to return in the response. Use
this parameter when paginating results. If additional items exist beyond
the number you specify, the \code{nextToken} element is set in the response.
Use the \code{nextToken} value in a subsequent request to retrieve additional
items.}

\item{nextToken}{String for specifying the next set of paginated results to return. After
you receive a response with truncated results, use this parameter in a
subsequent request. Set it to the value of \code{nextToken} from the response
that you just received.}

\item{isRevoked}{Filters results to return only signing jobs with revoked signatures.}

\item{signatureExpiresBefore}{Filters results to return only signing jobs with signatures expiring
before a specified timestamp.}

\item{signatureExpiresAfter}{Filters results to return only signing jobs with signatures expiring
after a specified timestamp.}

\item{jobInvoker}{Filters results to return only signing jobs initiated by a specified IAM
entity.}
}
\description{
Lists all your signing jobs. You can use the \code{maxResults} parameter to
limit the number of signing jobs that are returned in the response. If
additional jobs remain to be listed, code signing returns a \code{nextToken}
value. Use this value in subsequent calls to
\code{\link[=signer_list_signing_jobs]{list_signing_jobs}} to fetch the remaining
values. You can continue calling
\code{\link[=signer_list_signing_jobs]{list_signing_jobs}} with your \code{maxResults}
parameter and with new values that code signing returns in the
\code{nextToken} parameter until all of your signing jobs have been returned.
}
\section{Request syntax}{
\preformatted{svc$list_signing_jobs(
  status = "InProgress"|"Failed"|"Succeeded",
  platformId = "string",
  requestedBy = "string",
  maxResults = 123,
  nextToken = "string",
  isRevoked = TRUE|FALSE,
  signatureExpiresBefore = as.POSIXct(
    "2015-01-01"
  ),
  signatureExpiresAfter = as.POSIXct(
    "2015-01-01"
  ),
  jobInvoker = "string"
)
}
}

\keyword{internal}
