% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signer_operations.R
\name{signer_list_signing_profiles}
\alias{signer_list_signing_profiles}
\title{Lists all available signing profiles in your AWS account}
\usage{
signer_list_signing_profiles(includeCanceled, maxResults, nextToken,
  platformId, statuses)
}
\arguments{
\item{includeCanceled}{Designates whether to include profiles with the status of \code{CANCELED}.}

\item{maxResults}{The maximum number of profiles to be returned.}

\item{nextToken}{Value for specifying the next set of paginated results to return. After
you receive a response with truncated results, use this parameter in a
subsequent request. Set it to the value of \code{nextToken} from the response
that you just received.}

\item{platformId}{Filters results to return only signing jobs initiated for a specified
signing platform.}

\item{statuses}{Filters results to return only signing jobs with statuses in the
specified list.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  profiles = list(
    list(
      profileName = "string",
      profileVersion = "string",
      profileVersionArn = "string",
      signingMaterial = list(
        certificateArn = "string"
      ),
      signatureValidityPeriod = list(
        value = 123,
        type = "DAYS"|"MONTHS"|"YEARS"
      ),
      platformId = "string",
      platformDisplayName = "string",
      signingParameters = list(
        "string"
      ),
      status = "Active"|"Canceled"|"Revoked",
      arn = "string",
      tags = list(
        "string"
      )
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists all available signing profiles in your AWS account. Returns only
profiles with an \code{ACTIVE} status unless the \code{includeCanceled} request
field is set to \code{true}. If additional jobs remain to be listed, code
signing returns a \code{nextToken} value. Use this value in subsequent calls
to \code{\link[=signer_list_signing_jobs]{list_signing_jobs}} to fetch the
remaining values. You can continue calling
\code{\link[=signer_list_signing_jobs]{list_signing_jobs}} with your \code{maxResults}
parameter and with new values that code signing returns in the
\code{nextToken} parameter until all of your signing jobs have been returned.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_signing_profiles(
  includeCanceled = TRUE|FALSE,
  maxResults = 123,
  nextToken = "string",
  platformId = "string",
  statuses = list(
    "Active"|"Canceled"|"Revoked"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
