% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signer_operations.R
\name{signer_list_tags_for_resource}
\alias{signer_list_tags_for_resource}
\title{Returns a list of the tags associated with a signing profile resource}
\usage{
signer_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) for the signing profile.}
}
\description{
Returns a list of the tags associated with a signing profile resource.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  resourceArn = "string"
)
}
}

\keyword{internal}
