% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signer_operations.R
\name{signer_remove_profile_permission}
\alias{signer_remove_profile_permission}
\title{Removes cross-account permissions from a signing profile}
\usage{
signer_remove_profile_permission(profileName, revisionId, statementId)
}
\arguments{
\item{profileName}{[required] A human-readable name for the signing profile with permissions to be
removed.}

\item{revisionId}{[required] An identifier for the current revision of the signing profile
permissions.}

\item{statementId}{[required] A unique identifier for the cross-account permissions statement.}
}
\value{
A list with the following syntax:\preformatted{list(
  revisionId = "string"
)
}
}
\description{
Removes cross-account permissions from a signing profile.
}
\section{Request syntax}{
\preformatted{svc$remove_profile_permission(
  profileName = "string",
  revisionId = "string",
  statementId = "string"
)
}
}

\keyword{internal}
