% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signer_operations.R
\name{signer_revoke_signature}
\alias{signer_revoke_signature}
\title{Changes the state of a signing job to REVOKED}
\usage{
signer_revoke_signature(jobId, jobOwner, reason)
}
\arguments{
\item{jobId}{[required] ID of the signing job to be revoked.}

\item{jobOwner}{AWS account ID of the job owner.}

\item{reason}{[required] The reason for revoking the signing job.}
}
\value{
An empty list.
}
\description{
Changes the state of a signing job to REVOKED. This indicates that the
signature is no longer valid.
}
\section{Request syntax}{
\preformatted{svc$revoke_signature(
  jobId = "string",
  jobOwner = "string",
  reason = "string"
)
}
}

\keyword{internal}
