% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signer_operations.R
\name{signer_start_signing_job}
\alias{signer_start_signing_job}
\title{Initiates a signing job to be performed on the code provided}
\usage{
signer_start_signing_job(source, destination, profileName,
  clientRequestToken, profileOwner)
}
\arguments{
\item{source}{[required] The S3 bucket that contains the object to sign or a BLOB that contains
your raw code.}

\item{destination}{[required] The S3 bucket in which to save your signed object. The destination
contains the name of your bucket and an optional prefix.}

\item{profileName}{[required] The name of the signing profile.}

\item{clientRequestToken}{[required] String that identifies the signing request. All calls after the first
that use this token return the same response as the first call.}

\item{profileOwner}{The AWS account ID of the signing profile owner.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  jobId = "string",
  jobOwner = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Initiates a signing job to be performed on the code provided. Signing
jobs are viewable by the \code{\link[=signer_list_signing_jobs]{list_signing_jobs}}
operation for two years after they are performed. Note the following
requirements:
\itemize{
\item You must create an Amazon S3 source bucket. For more information,
see \href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/creating-bucket.html}{Create a Bucket}
in the \emph{Amazon S3 Getting Started Guide}.
\item Your S3 source bucket must be version enabled.
\item You must create an S3 destination bucket. Code signing uses your S3
destination bucket to write your signed code.
\item You specify the name of the source and destination buckets when
calling the \code{\link[=signer_start_signing_job]{start_signing_job}}
operation.
\item You must also specify a request token that identifies your request
to code signing.
}

You can call the \code{\link[=signer_describe_signing_job]{describe_signing_job}}
and the \code{\link[=signer_list_signing_jobs]{list_signing_jobs}} actions after
you call \code{\link[=signer_start_signing_job]{start_signing_job}}.

For a Java example that shows how to use this action, see
\href{https://docs.aws.amazon.com/acm/latest/userguide/}{http://docs.aws.amazon.com/acm/latest/userguide/}
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$start_signing_job(
  source = list(
    s3 = list(
      bucketName = "string",
      key = "string",
      version = "string"
    )
  ),
  destination = list(
    s3 = list(
      bucketName = "string",
      prefix = "string"
    )
  ),
  profileName = "string",
  clientRequestToken = "string",
  profileOwner = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
