% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_create_group_certificate_authority}
\alias{greengrass_create_group_certificate_authority}
\title{Creates a CA for the group}
\usage{
greengrass_create_group_certificate_authority(AmznClientToken, GroupId)
}
\arguments{
\item{AmznClientToken}{A client token used to correlate requests and responses.}

\item{GroupId}{[required] The ID of the Greengrass group.}
}
\value{
A list with the following syntax:\preformatted{list(
  GroupCertificateAuthorityArn = "string"
)
}
}
\description{
Creates a CA for the group. If a CA already exists, it will rotate the
existing CA.
}
\section{Request syntax}{
\preformatted{svc$create_group_certificate_authority(
  AmznClientToken = "string",
  GroupId = "string"
)
}
}

\keyword{internal}
