% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_get_device_definition}
\alias{greengrass_get_device_definition}
\title{Retrieves information about a device definition}
\usage{
greengrass_get_device_definition(DeviceDefinitionId)
}
\arguments{
\item{DeviceDefinitionId}{[required] The ID of the device definition.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Arn = "string",
  CreationTimestamp = "string",
  Id = "string",
  LastUpdatedTimestamp = "string",
  LatestVersion = "string",
  LatestVersionArn = "string",
  Name = "string",
  tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves information about a device definition.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_device_definition(
  DeviceDefinitionId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
