% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_get_device_definition_version}
\alias{greengrass_get_device_definition_version}
\title{Retrieves information about a device definition version}
\usage{
greengrass_get_device_definition_version(DeviceDefinitionId,
  DeviceDefinitionVersionId, NextToken)
}
\arguments{
\item{DeviceDefinitionId}{[required] The ID of the device definition.}

\item{DeviceDefinitionVersionId}{[required] The ID of the device definition version.}

\item{NextToken}{The token for the next set of results, or ''null'' if there are no additional results.}
}
\description{
Retrieves information about a device definition version.
}
\section{Request syntax}{
\preformatted{svc$get_device_definition_version(
  DeviceDefinitionId = "string",
  DeviceDefinitionVersionId = "string",
  NextToken = "string"
)
}
}

\keyword{internal}
