% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_get_group}
\alias{greengrass_get_group}
\title{Retrieves information about a group}
\usage{
greengrass_get_group(GroupId)
}
\arguments{
\item{GroupId}{[required] The ID of the Greengrass group.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Arn = "string",
  CreationTimestamp = "string",
  Id = "string",
  LastUpdatedTimestamp = "string",
  LatestVersion = "string",
  LatestVersionArn = "string",
  Name = "string",
  tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves information about a group.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_group(
  GroupId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
