% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_get_group_certificate_configuration}
\alias{greengrass_get_group_certificate_configuration}
\title{Retrieves the current configuration for the CA used by the group}
\usage{
greengrass_get_group_certificate_configuration(GroupId)
}
\arguments{
\item{GroupId}{[required] The ID of the Greengrass group.}
}
\value{
A list with the following syntax:\preformatted{list(
  CertificateAuthorityExpiryInMilliseconds = "string",
  CertificateExpiryInMilliseconds = "string",
  GroupId = "string"
)
}
}
\description{
Retrieves the current configuration for the CA used by the group.
}
\section{Request syntax}{
\preformatted{svc$get_group_certificate_configuration(
  GroupId = "string"
)
}
}

\keyword{internal}
