% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_get_group_version}
\alias{greengrass_get_group_version}
\title{Retrieves information about a group version}
\usage{
greengrass_get_group_version(GroupId, GroupVersionId)
}
\arguments{
\item{GroupId}{[required] The ID of the Greengrass group.}

\item{GroupVersionId}{[required] The ID of the group version.}
}
\description{
Retrieves information about a group version.
}
\section{Request syntax}{
\preformatted{svc$get_group_version(
  GroupId = "string",
  GroupVersionId = "string"
)
}
}

\keyword{internal}
