% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_get_logger_definition}
\alias{greengrass_get_logger_definition}
\title{Retrieves information about a logger definition}
\usage{
greengrass_get_logger_definition(LoggerDefinitionId)
}
\arguments{
\item{LoggerDefinitionId}{[required] The ID of the logger definition.}
}
\value{
A list with the following syntax:\preformatted{list(
  Arn = "string",
  CreationTimestamp = "string",
  Id = "string",
  LastUpdatedTimestamp = "string",
  LatestVersion = "string",
  LatestVersionArn = "string",
  Name = "string",
  tags = list(
    "string"
  )
)
}
}
\description{
Retrieves information about a logger definition.
}
\section{Request syntax}{
\preformatted{svc$get_logger_definition(
  LoggerDefinitionId = "string"
)
}
}

\keyword{internal}
