% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_get_logger_definition_version}
\alias{greengrass_get_logger_definition_version}
\title{Retrieves information about a logger definition version}
\usage{
greengrass_get_logger_definition_version(LoggerDefinitionId,
  LoggerDefinitionVersionId, NextToken)
}
\arguments{
\item{LoggerDefinitionId}{[required] The ID of the logger definition.}

\item{LoggerDefinitionVersionId}{[required] The ID of the logger definition version. This value maps to the
''Version'' property of the corresponding ''VersionInformation'' object,
which is returned by ''ListLoggerDefinitionVersions'' requests. If the
version is the last one that was associated with a logger definition,
the value also maps to the ''LatestVersion'' property of the
corresponding ''DefinitionInformation'' object.}

\item{NextToken}{The token for the next set of results, or ''null'' if there are no
additional results.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Arn = "string",
  CreationTimestamp = "string",
  Definition = list(
    Loggers = list(
      list(
        Component = "GreengrassSystem"|"Lambda",
        Id = "string",
        Level = "DEBUG"|"INFO"|"WARN"|"ERROR"|"FATAL",
        Space = 123,
        Type = "FileSystem"|"AWSCloudWatch"
      )
    )
  ),
  Id = "string",
  Version = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves information about a logger definition version.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_logger_definition_version(
  LoggerDefinitionId = "string",
  LoggerDefinitionVersionId = "string",
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
