% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_get_service_role_for_account}
\alias{greengrass_get_service_role_for_account}
\title{Retrieves the service role that is attached to your account}
\usage{
greengrass_get_service_role_for_account()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  AssociatedAt = "string",
  RoleArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves the service role that is attached to your account.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_service_role_for_account()
}\if{html}{\out{</div>}}
}

\keyword{internal}
