% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_list_groups}
\alias{greengrass_list_groups}
\title{Retrieves a list of groups}
\usage{
greengrass_list_groups(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{The maximum number of results to be returned per request.}

\item{NextToken}{The token for the next set of results, or ''null'' if there are no
additional results.}
}
\value{
A list with the following syntax:\preformatted{list(
  Groups = list(
    list(
      Arn = "string",
      CreationTimestamp = "string",
      Id = "string",
      LastUpdatedTimestamp = "string",
      LatestVersion = "string",
      LatestVersionArn = "string",
      Name = "string"
    )
  ),
  NextToken = "string"
)
}
}
\description{
Retrieves a list of groups.
}
\section{Request syntax}{
\preformatted{svc$list_groups(
  MaxResults = "string",
  NextToken = "string"
)
}
}

\keyword{internal}
