% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_tag_resource}
\alias{greengrass_tag_resource}
\title{Add tags to a resource}
\usage{
greengrass_tag_resource(ResourceArn, tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{tags}{[required] A map of the key-value pairs for the resource tag.}
}
\description{
Add tags to a resource.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  ResourceArn = "string",
  tags = list(
    "string"
  )
)
}
}

\keyword{internal}
