% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_update_group_certificate_configuration}
\alias{greengrass_update_group_certificate_configuration}
\title{Updates the Certificate expiry time for a group}
\usage{
greengrass_update_group_certificate_configuration(
  CertificateExpiryInMilliseconds, GroupId)
}
\arguments{
\item{CertificateExpiryInMilliseconds}{The amount of time remaining before the certificate expires, in milliseconds.}

\item{GroupId}{[required] The ID of the Greengrass group.}
}
\description{
Updates the Certificate expiry time for a group.
}
\section{Request syntax}{
\preformatted{svc$update_group_certificate_configuration(
  CertificateExpiryInMilliseconds = "string",
  GroupId = "string"
)
}
}

\keyword{internal}
