% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_update_thing_runtime_configuration}
\alias{greengrass_update_thing_runtime_configuration}
\title{Updates the runtime configuration of a thing}
\usage{
greengrass_update_thing_runtime_configuration(TelemetryConfiguration,
  ThingName)
}
\arguments{
\item{TelemetryConfiguration}{Configuration for telemetry service.}

\item{ThingName}{[required] The thing name.}
}
\value{
An empty list.
}
\description{
Updates the runtime configuration of a thing.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_thing_runtime_configuration(
  TelemetryConfiguration = list(
    Telemetry = "On"|"Off"
  ),
  ThingName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
