% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot1clickprojects_service.R
\name{iot1clickprojects}
\alias{iot1clickprojects}
\title{AWS IoT 1-Click Projects Service}
\usage{
iot1clickprojects(config = list())
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.}
}
\value{
A client for the service. You can call the service's operations using
syntax like \code{svc$operation(...)}, where \code{svc} is the name you've assigned
to the client. The available operations are listed in the
Operations section.
}
\description{
The AWS IoT 1-Click Projects API Reference
}
\section{Service syntax}{
\preformatted{svc <- iot1clickprojects(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string"
    ),
    endpoint = "string",
    region = "string"
  )
)
}
}

\section{Operations}{

\tabular{ll}{
\link[=iot1clickprojects_associate_device_with_placement]{associate_device_with_placement} \tab Associates a physical device with a placement\cr
\link[=iot1clickprojects_create_placement]{create_placement} \tab Creates an empty placement\cr
\link[=iot1clickprojects_create_project]{create_project} \tab Creates an empty project with a placement template\cr
\link[=iot1clickprojects_delete_placement]{delete_placement} \tab Deletes a placement\cr
\link[=iot1clickprojects_delete_project]{delete_project} \tab Deletes a project\cr
\link[=iot1clickprojects_describe_placement]{describe_placement} \tab Describes a placement in a project\cr
\link[=iot1clickprojects_describe_project]{describe_project} \tab Returns an object describing a project\cr
\link[=iot1clickprojects_disassociate_device_from_placement]{disassociate_device_from_placement} \tab Removes a physical device from a placement\cr
\link[=iot1clickprojects_get_devices_in_placement]{get_devices_in_placement} \tab Returns an object enumerating the devices in a placement\cr
\link[=iot1clickprojects_list_placements]{list_placements} \tab Lists the placement(s) of a project\cr
\link[=iot1clickprojects_list_projects]{list_projects} \tab Lists the AWS IoT 1-Click project(s) associated with your AWS account and region\cr
\link[=iot1clickprojects_list_tags_for_resource]{list_tags_for_resource} \tab Lists the tags (metadata key/value pairs) which you have assigned to the resource\cr
\link[=iot1clickprojects_tag_resource]{tag_resource} \tab Creates or modifies tags for a resource\cr
\link[=iot1clickprojects_untag_resource]{untag_resource} \tab Removes one or more tags (metadata key/value pairs) from a resource\cr
\link[=iot1clickprojects_update_placement]{update_placement} \tab Updates a placement with the given attributes\cr
\link[=iot1clickprojects_update_project]{update_project} \tab Updates a project associated with your AWS account and region
}
}

\examples{
\dontrun{
svc <- iot1clickprojects()
svc$associate_device_with_placement(
  Foo = 123
)
}

}
