% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_associate_targets_with_job}
\alias{iot_associate_targets_with_job}
\title{Associates a group with a continuous job}
\usage{
iot_associate_targets_with_job(targets, jobId, comment, namespaceId)
}
\arguments{
\item{targets}{[required] A list of thing group ARNs that define the targets of the job.}

\item{jobId}{[required] The unique identifier you assigned to this job when it was created.}

\item{comment}{An optional comment string describing why the job was associated with
the targets.}

\item{namespaceId}{The namespace used to indicate that a job is a customer-managed job.

When you specify a value for this parameter, AWS IoT Core sends jobs
notifications to MQTT topics that contain the value in the following
format.

\verb{$aws/things/THING_NAME/jobs/JOB_ID/notify-namespace-NAMESPACE_ID/}

The \code{namespaceId} feature is in public preview.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  jobArn = "string",
  jobId = "string",
  description = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Associates a group with a continuous job. The following criteria must be
met:
\itemize{
\item The job must have been created with the \code{targetSelection} field set
to "CONTINUOUS".
\item The job status must currently be "IN_PROGRESS".
\item The total number of targets associated with a job must not exceed
100.
}
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$associate_targets_with_job(
  targets = list(
    "string"
  ),
  jobId = "string",
  comment = "string",
  namespaceId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
