% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_clear_default_authorizer}
\alias{iot_clear_default_authorizer}
\title{Clears the default authorizer}
\usage{
iot_clear_default_authorizer()
}
\value{
An empty list.
}
\description{
Clears the default authorizer.
}
\section{Request syntax}{
\preformatted{svc$clear_default_authorizer()
}
}

\keyword{internal}
