% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_create_authorizer}
\alias{iot_create_authorizer}
\title{Creates an authorizer}
\usage{
iot_create_authorizer(authorizerName, authorizerFunctionArn,
  tokenKeyName, tokenSigningPublicKeys, status)
}
\arguments{
\item{authorizerName}{[required] The authorizer name.}

\item{authorizerFunctionArn}{[required] The ARN of the authorizer's Lambda function.}

\item{tokenKeyName}{[required] The name of the token key used to extract the token from the HTTP
headers.}

\item{tokenSigningPublicKeys}{[required] The public keys used to verify the digital signature returned by your
custom authentication service.}

\item{status}{The status of the create authorizer request.}
}
\description{
Creates an authorizer.
}
\section{Request syntax}{
\preformatted{svc$create_authorizer(
  authorizerName = "string",
  authorizerFunctionArn = "string",
  tokenKeyName = "string",
  tokenSigningPublicKeys = list(
    "string"
  ),
  status = "ACTIVE"|"INACTIVE"
)
}
}

\keyword{internal}
