% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_create_billing_group}
\alias{iot_create_billing_group}
\title{Creates a billing group}
\usage{
iot_create_billing_group(billingGroupName, billingGroupProperties, tags)
}
\arguments{
\item{billingGroupName}{[required] The name you wish to give to the billing group.}

\item{billingGroupProperties}{The properties of the billing group.}

\item{tags}{Metadata which can be used to manage the billing group.}
}
\description{
Creates a billing group.
}
\section{Request syntax}{
\preformatted{svc$create_billing_group(
  billingGroupName = "string",
  billingGroupProperties = list(
    billingGroupDescription = "string"
  ),
  tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
