% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_create_stream}
\alias{iot_create_stream}
\title{Creates a stream for delivering one or more large files in chunks over
MQTT}
\usage{
iot_create_stream(streamId, description, files, roleArn, tags)
}
\arguments{
\item{streamId}{[required] The stream ID.}

\item{description}{A description of the stream.}

\item{files}{[required] The files to stream.}

\item{roleArn}{[required] An IAM role that allows the IoT service principal assumes to access your
S3 files.}

\item{tags}{Metadata which can be used to manage streams.}
}
\description{
Creates a stream for delivering one or more large files in chunks over
MQTT. A stream transports data bytes in chunks or blocks packaged as
MQTT messages from a source like S3. You can have one or more files
associated with a stream.
}
\section{Request syntax}{
\preformatted{svc$create_stream(
  streamId = "string",
  description = "string",
  files = list(
    list(
      fileId = 123,
      s3Location = list(
        bucket = "string",
        key = "string",
        version = "string"
      )
    )
  ),
  roleArn = "string",
  tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
