% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_create_topic_rule_destination}
\alias{iot_create_topic_rule_destination}
\title{Creates a topic rule destination}
\usage{
iot_create_topic_rule_destination(destinationConfiguration)
}
\arguments{
\item{destinationConfiguration}{[required] The topic rule destination configuration.}
}
\description{
Creates a topic rule destination. The destination must be confirmed
prior to use.
}
\section{Request syntax}{
\preformatted{svc$create_topic_rule_destination(
  destinationConfiguration = list(
    httpUrlConfiguration = list(
      confirmationUrl = "string"
    ),
    vpcConfiguration = list(
      subnetIds = list(
        "string"
      ),
      securityGroups = list(
        "string"
      ),
      vpcId = "string",
      roleArn = "string"
    )
  )
)
}
}

\keyword{internal}
