% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_delete_dimension}
\alias{iot_delete_dimension}
\title{Removes the specified dimension from your AWS account}
\usage{
iot_delete_dimension(name)
}
\arguments{
\item{name}{[required] The unique identifier for the dimension that you want to delete.}
}
\value{
An empty list.
}
\description{
Removes the specified dimension from your AWS account.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_dimension(
  name = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
