% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_delete_domain_configuration}
\alias{iot_delete_domain_configuration}
\title{Deletes the specified domain configuration}
\usage{
iot_delete_domain_configuration(domainConfigurationName)
}
\arguments{
\item{domainConfigurationName}{[required] The name of the domain configuration to be deleted.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Deletes the specified domain configuration.

The domain configuration feature is in public preview and is subject to
change.
}
\section{Request syntax}{
\preformatted{svc$delete_domain_configuration(
  domainConfigurationName = "string"
)
}
}

\keyword{internal}
