% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_delete_ota_update}
\alias{iot_delete_ota_update}
\title{Delete an OTA update}
\usage{
iot_delete_ota_update(otaUpdateId, deleteStream, forceDeleteAWSJob)
}
\arguments{
\item{otaUpdateId}{[required] The ID of the OTA update to delete.}

\item{deleteStream}{Specifies if the stream associated with an OTA update should be deleted
when the OTA update is deleted.}

\item{forceDeleteAWSJob}{Specifies if the AWS Job associated with the OTA update should be
deleted when the OTA update is deleted.}
}
\value{
An empty list.
}
\description{
Delete an OTA update.
}
\section{Request syntax}{
\preformatted{svc$delete_ota_update(
  otaUpdateId = "string",
  deleteStream = TRUE|FALSE,
  forceDeleteAWSJob = TRUE|FALSE
)
}
}

\keyword{internal}
