% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_audit_suppression}
\alias{iot_describe_audit_suppression}
\title{Gets information about a Device Defender audit suppression}
\usage{
iot_describe_audit_suppression(checkName, resourceIdentifier)
}
\arguments{
\item{checkName}{[required]}

\item{resourceIdentifier}{[required]}
}
\value{
A list with the following syntax:\preformatted{list(
  checkName = "string",
  resourceIdentifier = list(
    deviceCertificateId = "string",
    caCertificateId = "string",
    cognitoIdentityPoolId = "string",
    clientId = "string",
    policyVersionIdentifier = list(
      policyName = "string",
      policyVersionId = "string"
    ),
    account = "string",
    iamRoleArn = "string",
    roleAliasArn = "string"
  ),
  expirationDate = as.POSIXct(
    "2015-01-01"
  ),
  suppressIndefinitely = TRUE|FALSE,
  description = "string"
)
}
}
\description{
Gets information about a Device Defender audit suppression.
}
\section{Request syntax}{
\preformatted{svc$describe_audit_suppression(
  checkName = "string",
  resourceIdentifier = list(
    deviceCertificateId = "string",
    caCertificateId = "string",
    cognitoIdentityPoolId = "string",
    clientId = "string",
    policyVersionIdentifier = list(
      policyName = "string",
      policyVersionId = "string"
    ),
    account = "string",
    iamRoleArn = "string",
    roleAliasArn = "string"
  )
)
}
}

\keyword{internal}
