% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_certificate}
\alias{iot_describe_certificate}
\title{Gets information about the specified certificate}
\usage{
iot_describe_certificate(certificateId)
}
\arguments{
\item{certificateId}{[required] The ID of the certificate. (The last part of the certificate ARN
contains the certificate ID.)}
}
\description{
Gets information about the specified certificate.
}
\section{Request syntax}{
\preformatted{svc$describe_certificate(
  certificateId = "string"
)
}
}

\keyword{internal}
