% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_custom_metric}
\alias{iot_describe_custom_metric}
\title{Gets information about a Device Defender detect custom metric}
\usage{
iot_describe_custom_metric(metricName)
}
\arguments{
\item{metricName}{[required] The name of the custom metric.}
}
\description{
Gets information about a Device Defender detect custom metric.
}
\section{Request syntax}{
\preformatted{svc$describe_custom_metric(
  metricName = "string"
)
}
}

\keyword{internal}
