% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_default_authorizer}
\alias{iot_describe_default_authorizer}
\title{Describes the default authorizer}
\usage{
iot_describe_default_authorizer()
}
\description{
Describes the default authorizer.
}
\section{Request syntax}{
\preformatted{svc$describe_default_authorizer()
}
}

\keyword{internal}
