% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_default_authorizer}
\alias{iot_describe_default_authorizer}
\title{Describes the default authorizer}
\usage{
iot_describe_default_authorizer()
}
\value{
A list with the following syntax:\preformatted{list(
  authorizerDescription = list(
    authorizerName = "string",
    authorizerArn = "string",
    authorizerFunctionArn = "string",
    tokenKeyName = "string",
    tokenSigningPublicKeys = list(
      "string"
    ),
    status = "ACTIVE"|"INACTIVE",
    creationDate = as.POSIXct(
      "2015-01-01"
    ),
    lastModifiedDate = as.POSIXct(
      "2015-01-01"
    ),
    signingDisabled = TRUE|FALSE
  )
)
}
}
\description{
Describes the default authorizer.
}
\section{Request syntax}{
\preformatted{svc$describe_default_authorizer()
}
}

\keyword{internal}
