% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_dimension}
\alias{iot_describe_dimension}
\title{Provides details about a dimension that is defined in your AWS account}
\usage{
iot_describe_dimension(name)
}
\arguments{
\item{name}{[required] The unique identifier for the dimension.}
}
\value{
A list with the following syntax:\preformatted{list(
  name = "string",
  arn = "string",
  type = "TOPIC_FILTER",
  stringValues = list(
    "string"
  ),
  creationDate = as.POSIXct(
    "2015-01-01"
  ),
  lastModifiedDate = as.POSIXct(
    "2015-01-01"
  )
)
}
}
\description{
Provides details about a dimension that is defined in your AWS account.
}
\section{Request syntax}{
\preformatted{svc$describe_dimension(
  name = "string"
)
}
}

\keyword{internal}
