% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_domain_configuration}
\alias{iot_describe_domain_configuration}
\title{Gets summary information about a domain configuration}
\usage{
iot_describe_domain_configuration(domainConfigurationName)
}
\arguments{
\item{domainConfigurationName}{[required] The name of the domain configuration.}
}
\description{
Gets summary information about a domain configuration.
}
\details{
The domain configuration feature is in public preview and is subject to
change.
}
\section{Request syntax}{
\preformatted{svc$describe_domain_configuration(
  domainConfigurationName = "string"
)
}
}

\keyword{internal}
