% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_event_configurations}
\alias{iot_describe_event_configurations}
\title{Describes event configurations}
\usage{
iot_describe_event_configurations()
}
\description{
Describes event configurations.
}
\section{Request syntax}{
\preformatted{svc$describe_event_configurations()
}
}

\keyword{internal}
