% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_index}
\alias{iot_describe_index}
\title{Describes a search index}
\usage{
iot_describe_index(indexName)
}
\arguments{
\item{indexName}{[required] The index name.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  indexName = "string",
  indexStatus = "ACTIVE"|"BUILDING"|"REBUILDING",
  schema = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Describes a search index.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_index(
  indexName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
