% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_thing_type}
\alias{iot_describe_thing_type}
\title{Gets information about the specified thing type}
\usage{
iot_describe_thing_type(thingTypeName)
}
\arguments{
\item{thingTypeName}{[required] The name of the thing type.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  thingTypeName = "string",
  thingTypeId = "string",
  thingTypeArn = "string",
  thingTypeProperties = list(
    thingTypeDescription = "string",
    searchableAttributes = list(
      "string"
    )
  ),
  thingTypeMetadata = list(
    deprecated = TRUE|FALSE,
    deprecationDate = as.POSIXct(
      "2015-01-01"
    ),
    creationDate = as.POSIXct(
      "2015-01-01"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets information about the specified thing type.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_thing_type(
  thingTypeName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
