% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_detach_principal_policy}
\alias{iot_detach_principal_policy}
\title{Removes the specified policy from the specified certificate}
\usage{
iot_detach_principal_policy(policyName, principal)
}
\arguments{
\item{policyName}{[required] The name of the policy to detach.}

\item{principal}{[required] The principal.

Valid principals are CertificateArn
(arn:aws:iot:\emph{region}:\emph{accountId}:cert/\emph{certificateId}), thingGroupArn
(arn:aws:iot:\emph{region}:\emph{accountId}:thinggroup/\emph{groupName}) and CognitoId
(\emph{region}:\emph{id}).}
}
\value{
An empty list.
}
\description{
Removes the specified policy from the specified certificate.

\strong{Note:} This API is deprecated. Please use
\code{\link[=iot_detach_policy]{detach_policy}} instead.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$detach_principal_policy(
  policyName = "string",
  principal = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
