% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_get_behavior_model_training_summaries}
\alias{iot_get_behavior_model_training_summaries}
\title{Returns a Device Defender's ML Detect Security Profile training model's
status}
\usage{
iot_get_behavior_model_training_summaries(securityProfileName,
  maxResults, nextToken)
}
\arguments{
\item{securityProfileName}{The name of the security profile.}

\item{maxResults}{The maximum number of results to return at one time. The default is 25.}

\item{nextToken}{The token for the next set of results.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  summaries = list(
    list(
      securityProfileName = "string",
      behaviorName = "string",
      trainingDataCollectionStartDate = as.POSIXct(
        "2015-01-01"
      ),
      modelStatus = "PENDING_BUILD"|"ACTIVE"|"EXPIRED",
      datapointsCollectionPercentage = 123.0,
      lastModelRefreshDate = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns a Device Defender's ML Detect Security Profile training model's
status.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_behavior_model_training_summaries(
  securityProfileName = "string",
  maxResults = 123,
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
