% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_get_effective_policies}
\alias{iot_get_effective_policies}
\title{Gets a list of the policies that have an effect on the authorization
behavior of the specified device when it connects to the AWS IoT device
gateway}
\usage{
iot_get_effective_policies(principal, cognitoIdentityPoolId, thingName)
}
\arguments{
\item{principal}{The principal. Valid principals are CertificateArn
(arn:aws:iot:\emph{region}:\emph{accountId}:cert/\emph{certificateId}), thingGroupArn
(arn:aws:iot:\emph{region}:\emph{accountId}:thinggroup/\emph{groupName}) and CognitoId
(\emph{region}:\emph{id}).}

\item{cognitoIdentityPoolId}{The Cognito identity pool ID.}

\item{thingName}{The thing name.}
}
\value{
A list with the following syntax:\preformatted{list(
  effectivePolicies = list(
    list(
      policyName = "string",
      policyArn = "string",
      policyDocument = "string"
    )
  )
)
}
}
\description{
Gets a list of the policies that have an effect on the authorization
behavior of the specified device when it connects to the AWS IoT device
gateway.
}
\section{Request syntax}{
\preformatted{svc$get_effective_policies(
  principal = "string",
  cognitoIdentityPoolId = "string",
  thingName = "string"
)
}
}

\keyword{internal}
