% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_get_indexing_configuration}
\alias{iot_get_indexing_configuration}
\title{Gets the indexing configuration}
\usage{
iot_get_indexing_configuration()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  thingIndexingConfiguration = list(
    thingIndexingMode = "OFF"|"REGISTRY"|"REGISTRY_AND_SHADOW",
    thingConnectivityIndexingMode = "OFF"|"STATUS",
    managedFields = list(
      list(
        name = "string",
        type = "Number"|"String"|"Boolean"
      )
    ),
    customFields = list(
      list(
        name = "string",
        type = "Number"|"String"|"Boolean"
      )
    )
  ),
  thingGroupIndexingConfiguration = list(
    thingGroupIndexingMode = "OFF"|"ON",
    managedFields = list(
      list(
        name = "string",
        type = "Number"|"String"|"Boolean"
      )
    ),
    customFields = list(
      list(
        name = "string",
        type = "Number"|"String"|"Boolean"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets the indexing configuration.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_indexing_configuration()
}\if{html}{\out{</div>}}
}

\keyword{internal}
