% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_get_indexing_configuration}
\alias{iot_get_indexing_configuration}
\title{Gets the indexing configuration}
\usage{
iot_get_indexing_configuration()
}
\description{
Gets the indexing configuration.
}
\section{Request syntax}{
\preformatted{svc$get_indexing_configuration()
}
}

\keyword{internal}
