% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_get_job_document}
\alias{iot_get_job_document}
\title{Gets a job document}
\usage{
iot_get_job_document(jobId)
}
\arguments{
\item{jobId}{[required] The unique identifier you assigned to this job when it was created.}
}
\description{
Gets a job document.
}
\section{Request syntax}{
\preformatted{svc$get_job_document(
  jobId = "string"
)
}
}

\keyword{internal}
