% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_get_logging_options}
\alias{iot_get_logging_options}
\title{Gets the logging options}
\usage{
iot_get_logging_options()
}
\description{
Gets the logging options.
}
\details{
NOTE: use of this command is not recommended. Use \code{GetV2LoggingOptions}
instead.
}
\section{Request syntax}{
\preformatted{svc$get_logging_options()
}
}

\keyword{internal}
