% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_get_v2_logging_options}
\alias{iot_get_v2_logging_options}
\title{Gets the fine grained logging options}
\usage{
iot_get_v2_logging_options()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  roleArn = "string",
  defaultLogLevel = "DEBUG"|"INFO"|"ERROR"|"WARN"|"DISABLED",
  disableAllLogs = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}
\description{
Gets the fine grained logging options.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_v2_logging_options()
}\if{html}{\out{</div>}}
}

\keyword{internal}
