% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_detect_mitigation_actions_tasks}
\alias{iot_list_detect_mitigation_actions_tasks}
\title{List of Device Defender ML Detect mitigation actions tasks}
\usage{
iot_list_detect_mitigation_actions_tasks(maxResults, nextToken,
  startTime, endTime)
}
\arguments{
\item{maxResults}{The maximum number of results to return at one time. The default is 25.}

\item{nextToken}{The token for the next set of results.}

\item{startTime}{[required] A filter to limit results to those found after the specified time. You
must specify either the startTime and endTime or the taskId, but not
both.}

\item{endTime}{[required] The end of the time period for which ML Detect mitigation actions tasks
are returned.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  tasks = list(
    list(
      taskId = "string",
      taskStatus = "IN_PROGRESS"|"SUCCESSFUL"|"FAILED"|"CANCELED",
      taskStartTime = as.POSIXct(
        "2015-01-01"
      ),
      taskEndTime = as.POSIXct(
        "2015-01-01"
      ),
      target = list(
        violationIds = list(
          "string"
        ),
        securityProfileName = "string",
        behaviorName = "string"
      ),
      violationEventOccurrenceRange = list(
        startTime = as.POSIXct(
          "2015-01-01"
        ),
        endTime = as.POSIXct(
          "2015-01-01"
        )
      ),
      onlyActiveViolationsIncluded = TRUE|FALSE,
      suppressedAlertsIncluded = TRUE|FALSE,
      actionsDefinition = list(
        list(
          name = "string",
          id = "string",
          roleArn = "string",
          actionParams = list(
            updateDeviceCertificateParams = list(
              action = "DEACTIVATE"
            ),
            updateCACertificateParams = list(
              action = "DEACTIVATE"
            ),
            addThingsToThingGroupParams = list(
              thingGroupNames = list(
                "string"
              ),
              overrideDynamicGroups = TRUE|FALSE
            ),
            replaceDefaultPolicyVersionParams = list(
              templateName = "BLANK_POLICY"
            ),
            enableIoTLoggingParams = list(
              roleArnForLogging = "string",
              logLevel = "DEBUG"|"INFO"|"ERROR"|"WARN"|"DISABLED"
            ),
            publishFindingToSnsParams = list(
              topicArn = "string"
            )
          )
        )
      ),
      taskStatistics = list(
        actionsExecuted = 123,
        actionsSkipped = 123,
        actionsFailed = 123
      )
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
List of Device Defender ML Detect mitigation actions tasks.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_detect_mitigation_actions_tasks(
  maxResults = 123,
  nextToken = "string",
  startTime = as.POSIXct(
    "2015-01-01"
  ),
  endTime = as.POSIXct(
    "2015-01-01"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
