% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_dimensions}
\alias{iot_list_dimensions}
\title{List the set of dimensions that are defined for your AWS account}
\usage{
iot_list_dimensions(nextToken, maxResults)
}
\arguments{
\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to retrieve at one time.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  dimensionNames = list(
    "string"
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
List the set of dimensions that are defined for your AWS account.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_dimensions(
  nextToken = "string",
  maxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
