% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_policy_principals}
\alias{iot_list_policy_principals}
\title{Lists the principals associated with the specified policy}
\usage{
iot_list_policy_principals(policyName, marker, pageSize, ascendingOrder)
}
\arguments{
\item{policyName}{[required] The policy name.}

\item{marker}{The marker for the next set of results.}

\item{pageSize}{The result page size.}

\item{ascendingOrder}{Specifies the order for results. If true, the results are returned in
ascending creation order.}
}
\description{
Lists the principals associated with the specified policy.

\strong{Note:} This API is deprecated. Please use ListTargetsForPolicy
instead.
}
\section{Request syntax}{
\preformatted{svc$list_policy_principals(
  policyName = "string",
  marker = "string",
  pageSize = 123,
  ascendingOrder = TRUE|FALSE
)
}
}

\keyword{internal}
