% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_provisioning_templates}
\alias{iot_list_provisioning_templates}
\title{Lists the fleet provisioning templates in your AWS account}
\usage{
iot_list_provisioning_templates(maxResults, nextToken)
}
\arguments{
\item{maxResults}{The maximum number of results to return at one time.}

\item{nextToken}{A token to retrieve the next set of results.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  templates = list(
    list(
      templateArn = "string",
      templateName = "string",
      description = "string",
      creationDate = as.POSIXct(
        "2015-01-01"
      ),
      lastModifiedDate = as.POSIXct(
        "2015-01-01"
      ),
      enabled = TRUE|FALSE
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists the fleet provisioning templates in your AWS account.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_provisioning_templates(
  maxResults = 123,
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
