% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_tags_for_resource}
\alias{iot_list_tags_for_resource}
\title{Lists the tags (metadata) you have assigned to the resource}
\usage{
iot_list_tags_for_resource(resourceArn, nextToken)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource.}

\item{nextToken}{To retrieve the next set of results, the \code{nextToken} value from a
previous response; otherwise \strong{null} to receive the first set of
results.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists the tags (metadata) you have assigned to the resource.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_tags_for_resource(
  resourceArn = "string",
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
