% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_targets_for_policy}
\alias{iot_list_targets_for_policy}
\title{List targets for the specified policy}
\usage{
iot_list_targets_for_policy(policyName, marker, pageSize)
}
\arguments{
\item{policyName}{[required] The policy name.}

\item{marker}{A marker used to get the next set of results.}

\item{pageSize}{The maximum number of results to return at one time.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  targets = list(
    "string"
  ),
  nextMarker = "string"
)
}\if{html}{\out{</div>}}
}
\description{
List targets for the specified policy.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_targets_for_policy(
  policyName = "string",
  marker = "string",
  pageSize = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
