% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_things}
\alias{iot_list_things}
\title{Lists your things}
\usage{
iot_list_things(nextToken, maxResults, attributeName, attributeValue,
  thingTypeName)
}
\arguments{
\item{nextToken}{The token to retrieve the next set of results.}

\item{maxResults}{The maximum number of results to return in this operation.}

\item{attributeName}{The attribute name used to search for things.}

\item{attributeValue}{The attribute value used to search for things.}

\item{thingTypeName}{The name of the thing type used to search for things.}
}
\description{
Lists your things. Use the \strong{attributeName} and \strong{attributeValue}
parameters to filter your things. For example, calling \code{ListThings} with
attributeName=Color and attributeValue=Red retrieves all things in the
registry that contain an attribute \strong{Color} with the value \strong{Red}.
}
\section{Request syntax}{
\preformatted{svc$list_things(
  nextToken = "string",
  maxResults = 123,
  attributeName = "string",
  attributeValue = "string",
  thingTypeName = "string"
)
}
}

\keyword{internal}
