% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_things_in_thing_group}
\alias{iot_list_things_in_thing_group}
\title{Lists the things in the specified group}
\usage{
iot_list_things_in_thing_group(thingGroupName, recursive, nextToken,
  maxResults)
}
\arguments{
\item{thingGroupName}{[required] The thing group name.}

\item{recursive}{When true, list things in this thing group and in all child groups as
well.}

\item{nextToken}{The token to retrieve the next set of results.}

\item{maxResults}{The maximum number of results to return at one time.}
}
\description{
Lists the things in the specified group.
}
\section{Request syntax}{
\preformatted{svc$list_things_in_thing_group(
  thingGroupName = "string",
  recursive = TRUE|FALSE,
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
