% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_topic_rule_destinations}
\alias{iot_list_topic_rule_destinations}
\title{Lists all the topic rule destinations in your AWS account}
\usage{
iot_list_topic_rule_destinations(maxResults, nextToken)
}
\arguments{
\item{maxResults}{The maximum number of results to return at one time.}

\item{nextToken}{To retrieve the next set of results, the \code{nextToken} value from a
previous response; otherwise \strong{null} to receive the first set of
results.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  destinationSummaries = list(
    list(
      arn = "string",
      status = "ENABLED"|"IN_PROGRESS"|"DISABLED"|"ERROR"|"DELETING",
      createdAt = as.POSIXct(
        "2015-01-01"
      ),
      lastUpdatedAt = as.POSIXct(
        "2015-01-01"
      ),
      statusReason = "string",
      httpUrlSummary = list(
        confirmationUrl = "string"
      ),
      vpcDestinationSummary = list(
        subnetIds = list(
          "string"
        ),
        securityGroups = list(
          "string"
        ),
        vpcId = "string",
        roleArn = "string"
      )
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists all the topic rule destinations in your AWS account.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_topic_rule_destinations(
  maxResults = 123,
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
