% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_topic_rule_destinations}
\alias{iot_list_topic_rule_destinations}
\title{Lists all the topic rule destinations in your AWS account}
\usage{
iot_list_topic_rule_destinations(maxResults, nextToken)
}
\arguments{
\item{maxResults}{The maximum number of results to return at one time.}

\item{nextToken}{To retrieve the next set of results, the \code{nextToken} value from a
previous response; otherwise \strong{null} to receive the first set of
results.}
}
\description{
Lists all the topic rule destinations in your AWS account.
}
\section{Request syntax}{
\preformatted{svc$list_topic_rule_destinations(
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
