% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_topic_rules}
\alias{iot_list_topic_rules}
\title{Lists the rules for the specific topic}
\usage{
iot_list_topic_rules(topic, maxResults, nextToken, ruleDisabled)
}
\arguments{
\item{topic}{The topic.}

\item{maxResults}{The maximum number of results to return.}

\item{nextToken}{To retrieve the next set of results, the \code{nextToken} value from a
previous response; otherwise \strong{null} to receive the first set of
results.}

\item{ruleDisabled}{Specifies whether the rule is disabled.}
}
\value{
A list with the following syntax:\preformatted{list(
  rules = list(
    list(
      ruleArn = "string",
      ruleName = "string",
      topicPattern = "string",
      createdAt = as.POSIXct(
        "2015-01-01"
      ),
      ruleDisabled = TRUE|FALSE
    )
  ),
  nextToken = "string"
)
}
}
\description{
Lists the rules for the specific topic.
}
\section{Request syntax}{
\preformatted{svc$list_topic_rules(
  topic = "string",
  maxResults = 123,
  nextToken = "string",
  ruleDisabled = TRUE|FALSE
)
}
}

\keyword{internal}
