% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_violation_events}
\alias{iot_list_violation_events}
\title{Lists the Device Defender security profile violations discovered during
the given time period}
\usage{
iot_list_violation_events(startTime, endTime, thingName,
  securityProfileName, nextToken, maxResults)
}
\arguments{
\item{startTime}{[required] The start time for the alerts to be listed.}

\item{endTime}{[required] The end time for the alerts to be listed.}

\item{thingName}{A filter to limit results to those alerts caused by the specified thing.}

\item{securityProfileName}{A filter to limit results to those alerts generated by the specified
security profile.}

\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to return at one time.}
}
\description{
Lists the Device Defender security profile violations discovered during
the given time period. You can use filters to limit the results to those
alerts issued for a particular security profile, behavior or thing
(device).
}
\section{Request syntax}{
\preformatted{svc$list_violation_events(
  startTime = as.POSIXct(
    "2015-01-01"
  ),
  endTime = as.POSIXct(
    "2015-01-01"
  ),
  thingName = "string",
  securityProfileName = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
