% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_set_default_authorizer}
\alias{iot_set_default_authorizer}
\title{Sets the default authorizer}
\usage{
iot_set_default_authorizer(authorizerName)
}
\arguments{
\item{authorizerName}{[required] The authorizer name.}
}
\value{
A list with the following syntax:\preformatted{list(
  authorizerName = "string",
  authorizerArn = "string"
)
}
}
\description{
Sets the default authorizer. This will be used if a websocket connection
is made without specifying an authorizer.
}
\section{Request syntax}{
\preformatted{svc$set_default_authorizer(
  authorizerName = "string"
)
}
}

\keyword{internal}
