% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_stop_thing_registration_task}
\alias{iot_stop_thing_registration_task}
\title{Cancels a bulk thing provisioning task}
\usage{
iot_stop_thing_registration_task(taskId)
}
\arguments{
\item{taskId}{[required] The bulk thing provisioning task ID.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Cancels a bulk thing provisioning task.
}
\section{Request syntax}{
\preformatted{svc$stop_thing_registration_task(
  taskId = "string"
)
}
}

\keyword{internal}
