% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_test_invoke_authorizer}
\alias{iot_test_invoke_authorizer}
\title{Tests a custom authorization behavior by invoking a specified custom
authorizer}
\usage{
iot_test_invoke_authorizer(authorizerName, token, tokenSignature,
  httpContext, mqttContext, tlsContext)
}
\arguments{
\item{authorizerName}{[required] The custom authorizer name.}

\item{token}{The token returned by your custom authentication service.}

\item{tokenSignature}{The signature made with the token and your custom authentication
service\'s private key.}

\item{httpContext}{Specifies a test HTTP authorization request.}

\item{mqttContext}{Specifies a test MQTT authorization request.\\>}

\item{tlsContext}{Specifies a test TLS authorization request.}
}
\description{
Tests a custom authorization behavior by invoking a specified custom
authorizer. Use this to test and debug the custom authorization behavior
of devices that connect to the AWS IoT device gateway.
}
\section{Request syntax}{
\preformatted{svc$test_invoke_authorizer(
  authorizerName = "string",
  token = "string",
  tokenSignature = "string",
  httpContext = list(
    headers = list(
      "string"
    ),
    queryString = "string"
  ),
  mqttContext = list(
    username = "string",
    password = raw,
    clientId = "string"
  ),
  tlsContext = list(
    serverName = "string"
  )
)
}
}

\keyword{internal}
