% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_update_billing_group}
\alias{iot_update_billing_group}
\title{Updates information about the billing group}
\usage{
iot_update_billing_group(billingGroupName, billingGroupProperties,
  expectedVersion)
}
\arguments{
\item{billingGroupName}{[required] The name of the billing group.}

\item{billingGroupProperties}{[required] The properties of the billing group.}

\item{expectedVersion}{The expected version of the billing group. If the version of the billing
group does not match the expected version specified in the request, the
\code{\link[=iot_update_billing_group]{update_billing_group}} request is rejected
with a \code{VersionConflictException}.}
}
\value{
A list with the following syntax:\preformatted{list(
  version = 123
)
}
}
\description{
Updates information about the billing group.
}
\section{Request syntax}{
\preformatted{svc$update_billing_group(
  billingGroupName = "string",
  billingGroupProperties = list(
    billingGroupDescription = "string"
  ),
  expectedVersion = 123
)
}
}

\keyword{internal}
