% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_update_topic_rule_destination}
\alias{iot_update_topic_rule_destination}
\title{Updates a topic rule destination}
\usage{
iot_update_topic_rule_destination(arn, status)
}
\arguments{
\item{arn}{[required] The ARN of the topic rule destination.}

\item{status}{[required] The status of the topic rule destination. Valid values are:
\subsection{IN\\_PROGRESS}{

A topic rule destination was created but has not been confirmed. You can
set \code{status} to \code{IN_PROGRESS} by calling \code{UpdateTopicRuleDestination}.
Calling \code{UpdateTopicRuleDestination} causes a new confirmation challenge
to be sent to your confirmation endpoint.
}

\subsection{ENABLED}{

Confirmation was completed, and traffic to this destination is allowed.
You can set \code{status} to \code{DISABLED} by calling
\code{UpdateTopicRuleDestination}.
}

\subsection{DISABLED}{

Confirmation was completed, and traffic to this destination is not
allowed. You can set \code{status} to \code{ENABLED} by calling
\code{UpdateTopicRuleDestination}.
}

\subsection{ERROR}{

Confirmation could not be completed, for example if the confirmation
timed out. You can call \code{GetTopicRuleDestination} for details about the
error. You can set \code{status} to \code{IN_PROGRESS} by calling
\code{UpdateTopicRuleDestination}. Calling \code{UpdateTopicRuleDestination}
causes a new confirmation challenge to be sent to your confirmation
endpoint.
}}
}
\description{
Updates a topic rule destination. You use this to change the status,
endpoint URL, or confirmation URL of the destination.
}
\section{Request syntax}{
\preformatted{svc$update_topic_rule_destination(
  arn = "string",
  status = "ENABLED"|"IN_PROGRESS"|"DISABLED"|"ERROR"
)
}
}

\keyword{internal}
