% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_create_dataset_content}
\alias{iotanalytics_create_dataset_content}
\title{Creates the content of a data set by applying a "queryAction" (a SQL
query) or a "containerAction" (executing a containerized application)}
\usage{
iotanalytics_create_dataset_content(datasetName)
}
\arguments{
\item{datasetName}{[required] The name of the data set.}
}
\description{
Creates the content of a data set by applying a \"queryAction\" (a SQL
query) or a \"containerAction\" (executing a containerized application).
}
\section{Request syntax}{
\preformatted{svc$create_dataset_content(
  datasetName = "string"
)
}
}

\keyword{internal}
