% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_create_datastore}
\alias{iotanalytics_create_datastore}
\title{Creates a data store, which is a repository for messages}
\usage{
iotanalytics_create_datastore(datastoreName, datastoreStorage,
  retentionPeriod, tags, fileFormatConfiguration)
}
\arguments{
\item{datastoreName}{[required] The name of the data store.}

\item{datastoreStorage}{Where data store data is stored. You can choose one of
\code{serviceManagedS3} or \code{customerManagedS3} storage. If not specified, the
default is \code{serviceManagedS3}. You cannot change this storage option
after the data store is created.}

\item{retentionPeriod}{How long, in days, message data is kept for the data store. When
\code{customerManagedS3} storage is selected, this parameter is ignored.}

\item{tags}{Metadata which can be used to manage the data store.}

\item{fileFormatConfiguration}{Contains the configuration information of file formats. AWS IoT
Analytics data stores support JSON and Parquet.

The default file format is JSON. You can specify only one format.

You can't change the file format after you create the data store.}
}
\value{
A list with the following syntax:\preformatted{list(
  datastoreName = "string",
  datastoreArn = "string",
  retentionPeriod = list(
    unlimited = TRUE|FALSE,
    numberOfDays = 123
  )
)
}
}
\description{
Creates a data store, which is a repository for messages.
}
\section{Request syntax}{
\preformatted{svc$create_datastore(
  datastoreName = "string",
  datastoreStorage = list(
    serviceManagedS3 = list(),
    customerManagedS3 = list(
      bucket = "string",
      keyPrefix = "string",
      roleArn = "string"
    )
  ),
  retentionPeriod = list(
    unlimited = TRUE|FALSE,
    numberOfDays = 123
  ),
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  ),
  fileFormatConfiguration = list(
    jsonConfiguration = list(),
    parquetConfiguration = list(
      schemaDefinition = list(
        columns = list(
          list(
            name = "string",
            type = "string"
          )
        )
      )
    )
  )
)
}
}

\keyword{internal}
