% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_delete_dataset_content}
\alias{iotanalytics_delete_dataset_content}
\title{Deletes the content of the specified dataset}
\usage{
iotanalytics_delete_dataset_content(datasetName, versionId)
}
\arguments{
\item{datasetName}{[required] The name of the dataset whose content is deleted.}

\item{versionId}{The version of the dataset whose content is deleted. You can also use
the strings "$LATEST" or "$LATEST_SUCCEEDED" to delete the latest or
latest successfully completed data set. If not specified,
"$LATEST_SUCCEEDED" is the default.}
}
\value{
An empty list.
}
\description{
Deletes the content of the specified dataset.
}
\section{Request syntax}{
\preformatted{svc$delete_dataset_content(
  datasetName = "string",
  versionId = "string"
)
}
}

\keyword{internal}
