% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_describe_datastore}
\alias{iotanalytics_describe_datastore}
\title{Retrieves information about a data store}
\usage{
iotanalytics_describe_datastore(datastoreName, includeStatistics)
}
\arguments{
\item{datastoreName}{[required] The name of the data store}

\item{includeStatistics}{If true, additional statistical information about the data store is
included in the response. This feature cannot be used with a data store
whose S3 storage is customer-managed.}
}
\value{
A list with the following syntax:\preformatted{list(
  datastore = list(
    name = "string",
    storage = list(
      serviceManagedS3 = list(),
      customerManagedS3 = list(
        bucket = "string",
        keyPrefix = "string",
        roleArn = "string"
      )
    ),
    arn = "string",
    status = "CREATING"|"ACTIVE"|"DELETING",
    retentionPeriod = list(
      unlimited = TRUE|FALSE,
      numberOfDays = 123
    ),
    creationTime = as.POSIXct(
      "2015-01-01"
    ),
    lastUpdateTime = as.POSIXct(
      "2015-01-01"
    ),
    lastMessageArrivalTime = as.POSIXct(
      "2015-01-01"
    ),
    fileFormatConfiguration = list(
      jsonConfiguration = list(),
      parquetConfiguration = list(
        schemaDefinition = list(
          columns = list(
            list(
              name = "string",
              type = "string"
            )
          )
        )
      )
    )
  ),
  statistics = list(
    size = list(
      estimatedSizeInBytes = 123.0,
      estimatedOn = as.POSIXct(
        "2015-01-01"
      )
    )
  )
)
}
}
\description{
Retrieves information about a data store.
}
\section{Request syntax}{
\preformatted{svc$describe_datastore(
  datastoreName = "string",
  includeStatistics = TRUE|FALSE
)
}
}

\keyword{internal}
