% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_describe_logging_options}
\alias{iotanalytics_describe_logging_options}
\title{Retrieves the current settings of the AWS IoT Analytics logging options}
\usage{
iotanalytics_describe_logging_options()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  loggingOptions = list(
    roleArn = "string",
    level = "ERROR",
    enabled = TRUE|FALSE
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves the current settings of the AWS IoT Analytics logging options.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_logging_options()
}\if{html}{\out{</div>}}
}

\keyword{internal}
