% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_list_tags_for_resource}
\alias{iotanalytics_list_tags_for_resource}
\title{Lists the tags (metadata) that you have assigned to the resource}
\usage{
iotanalytics_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource whose tags you want to list.}
}
\value{
A list with the following syntax:\preformatted{list(
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  )
)
}
}
\description{
Lists the tags (metadata) that you have assigned to the resource.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  resourceArn = "string"
)
}
}

\keyword{internal}
