% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_update_datastore}
\alias{iotanalytics_update_datastore}
\title{Updates the settings of a data store}
\usage{
iotanalytics_update_datastore(datastoreName, retentionPeriod)
}
\arguments{
\item{datastoreName}{[required] The name of the data store to be updated.}

\item{retentionPeriod}{How long, in days, message data is kept for the data store.}
}
\description{
Updates the settings of a data store.
}
\section{Request syntax}{
\preformatted{svc$update_datastore(
  datastoreName = "string",
  retentionPeriod = list(
    unlimited = TRUE|FALSE,
    numberOfDays = 123
  )
)
}
}

\keyword{internal}
