% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotjobsdataplane_operations.R
\name{iotjobsdataplane_get_pending_job_executions}
\alias{iotjobsdataplane_get_pending_job_executions}
\title{Gets the list of all jobs for a thing that are not in a terminal status}
\usage{
iotjobsdataplane_get_pending_job_executions(thingName)
}
\arguments{
\item{thingName}{[required] The name of the thing that is executing the job.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  inProgressJobs = list(
    list(
      jobId = "string",
      queuedAt = 123,
      startedAt = 123,
      lastUpdatedAt = 123,
      versionNumber = 123,
      executionNumber = 123
    )
  ),
  queuedJobs = list(
    list(
      jobId = "string",
      queuedAt = 123,
      startedAt = 123,
      lastUpdatedAt = 123,
      versionNumber = 123,
      executionNumber = 123
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets the list of all jobs for a thing that are not in a terminal status.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_pending_job_executions(
  thingName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
