% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signer_operations.R
\name{signer_get_signing_profile}
\alias{signer_get_signing_profile}
\title{Returns information on a specific signing profile}
\usage{
signer_get_signing_profile(profileName, profileOwner)
}
\arguments{
\item{profileName}{[required] The name of the target signing profile.}

\item{profileOwner}{The AWS account ID of the profile owner.}
}
\value{
A list with the following syntax:\preformatted{list(
  profileName = "string",
  profileVersion = "string",
  profileVersionArn = "string",
  revocationRecord = list(
    revocationEffectiveFrom = as.POSIXct(
      "2015-01-01"
    ),
    revokedAt = as.POSIXct(
      "2015-01-01"
    ),
    revokedBy = "string"
  ),
  signingMaterial = list(
    certificateArn = "string"
  ),
  platformId = "string",
  platformDisplayName = "string",
  signatureValidityPeriod = list(
    value = 123,
    type = "DAYS"|"MONTHS"|"YEARS"
  ),
  overrides = list(
    signingConfiguration = list(
      encryptionAlgorithm = "RSA"|"ECDSA",
      hashAlgorithm = "SHA1"|"SHA256"
    ),
    signingImageFormat = "JSON"|"JSONEmbedded"|"JSONDetached"
  ),
  signingParameters = list(
    "string"
  ),
  status = "Active"|"Canceled"|"Revoked",
  statusReason = "string",
  arn = "string",
  tags = list(
    "string"
  )
)
}
}
\description{
Returns information on a specific signing profile.
}
\section{Request syntax}{
\preformatted{svc$get_signing_profile(
  profileName = "string",
  profileOwner = "string"
)
}
}

\keyword{internal}
