% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signer_operations.R
\name{signer_put_signing_profile}
\alias{signer_put_signing_profile}
\title{Creates a signing profile}
\usage{
signer_put_signing_profile(profileName, signingMaterial, platformId,
  overrides, signingParameters)
}
\arguments{
\item{profileName}{[required] The name of the signing profile to be created.}

\item{signingMaterial}{[required] The AWS Certificate Manager certificate that will be used to sign code
with the new signing profile.}

\item{platformId}{[required] The ID of the signing profile to be created.}

\item{overrides}{A subfield of \code{platform}. This specifies any different configuration
options that you want to apply to the chosen platform (such as a
different \code{hash-algorithm} or \code{signing-algorithm}).}

\item{signingParameters}{Map of key-value pairs for signing. These can include any information
that you want to use during signing.}
}
\description{
Creates a signing profile. A signing profile is an AWS Signer template
that can be used to carry out a pre-defined signing job. For more
information, see
\url{http://docs.aws.amazon.com/signer/latest/developerguide/gs-profile.html}
}
\section{Request syntax}{
\preformatted{svc$put_signing_profile(
  profileName = "string",
  signingMaterial = list(
    certificateArn = "string"
  ),
  platformId = "string",
  overrides = list(
    signingConfiguration = list(
      encryptionAlgorithm = "RSA"|"ECDSA",
      hashAlgorithm = "SHA1"|"SHA256"
    )
  ),
  signingParameters = list(
    "string"
  )
)
}
}

\keyword{internal}
