% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signer_operations.R
\name{signer_untag_resource}
\alias{signer_untag_resource}
\title{Removes one or more tags from a signing profile}
\usage{
signer_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) for the signing profile.}

\item{tagKeys}{[required] A list of tag keys to be removed from the signing profile.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Removes one or more tags from a signing profile. To remove the tags,
specify a list of tag keys.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  resourceArn = "string",
  tagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
